Evaluate an R expression while collecting journals from completed futures

Evaluate an R expression while collecting journals from completed futures

Usage

capture_journals(expr, substitute = TRUE, envir = parent.frame())

Arguments

expr

The R expression to evaluate

substitute

If TRUE, then expr is substituted, otherwise not.

envir

The environment where expr should be evaluated

Value

A list of FutureJournal:s.

Details

This function evaluates an R expression and capture the journals signaled by futures as they are completed. A future journal comprise a log of events appearing during the life-span of a future, e.g. the timestamps when the future was created, launched, queried, resolved, and its results are collected.

Examples

slow_fcn <- function(x) {
  Sys.sleep(x / 10)
  sqrt(x)
}

plan(multisession, workers = 2)
js <- capture_journals({
  fs <- lapply(3:1, FUN = function(x) future(slow_fcn(x)))
  value(fs)
})

## Summarize all journals
js_all <- Reduce(rbind, js)
print(summary(js_all), digits = 2L)
#> Number of futures: 3
#>   summary  evaluate evaluate_ratio   overhead overhead_ratio  duration
#> 1     min 0.11 secs           0.18 0.018 secs          0.030 0.61 secs
#> 2    mean 0.31 secs           0.43 0.223 secs          0.317 0.70 secs
#> 3  median 0.36 secs           0.50 0.026 secs          0.036 0.72 secs
#> 4     max 0.45 secs           0.58 0.624 secs          0.800 0.78 secs
#> 5   total 0.92 secs           0.43 0.668 secs          0.317 2.11 secs
#>    walltime
#> 1 0.61 secs
#> 2 0.70 secs
#> 3 0.72 secs
#> 4 0.78 secs
#> 5 2.11 secs

## Shut down parallel workers
plan(sequential)