Evaluate an R expression while collecting journals from completed futures

Evaluate an R expression while collecting journals from completed futures

gather_journals(expr, substitute = TRUE, envir = parent.frame())

Arguments

expr

The R expression to evaluate

substitute

If TRUE, then expr is subtituted, otherwise not.

envir

The environment where expr should be evaluated

Value

A list of FutureJournal:s.

Examples

library(future.apply)
#> Loading required package: future

slow_fcn <- function(x) {
  Sys.sleep(x / 10)
  sqrt(x)
}

plan(multisession, workers = 2)
js <- gather_journals({
  vs <- future_lapply(3:1, FUN = slow_fcn)
})
print(js)
#> [[1]]
#>              step                   start               at          duration
#> 1          create 2022-03-09 17:59:47.183 0.000000000 secs 1.344681e-03 secs
#> 6          launch 2022-03-09 17:59:47.187 0.003739834 secs 5.626154e-02 secs
#> 2       getWorker 2022-03-09 17:59:47.188 0.004810333 secs 2.200603e-04 secs
#> 3     eraseWorker 2022-03-09 17:59:47.189 0.006147146 secs 7.452965e-04 secs
#> 4  attachPackages 2022-03-09 17:59:47.191 0.007812738 secs 6.318092e-05 secs
#> 5   exportGlobals 2022-03-09 17:59:47.192 0.008764029 secs 4.723763e-02 secs
#> 10       evaluate 2022-03-09 17:59:47.287 0.103488207 secs 3.041193e-01 secs
#> 7        resolved 2022-03-09 17:59:47.380 0.197264433 secs 1.228070e-02 secs
#> 8        resolved 2022-03-09 17:59:47.507 0.323828220 secs 1.212215e-02 secs
#> 9          gather 2022-03-09 17:59:47.637 0.453879356 secs 1.695156e-04 secs
#>       future_label                          future_uuid
#> 1  future_lapply-1 a3c7fd72-c4c8-e4b8-c154-03a77e8a4a35
#> 6  future_lapply-1 a3c7fd72-c4c8-e4b8-c154-03a77e8a4a35
#> 2  future_lapply-1 a3c7fd72-c4c8-e4b8-c154-03a77e8a4a35
#> 3  future_lapply-1 a3c7fd72-c4c8-e4b8-c154-03a77e8a4a35
#> 4  future_lapply-1 a3c7fd72-c4c8-e4b8-c154-03a77e8a4a35
#> 5  future_lapply-1 a3c7fd72-c4c8-e4b8-c154-03a77e8a4a35
#> 10 future_lapply-1 a3c7fd72-c4c8-e4b8-c154-03a77e8a4a35
#> 7  future_lapply-1 a3c7fd72-c4c8-e4b8-c154-03a77e8a4a35
#> 8  future_lapply-1 a3c7fd72-c4c8-e4b8-c154-03a77e8a4a35
#> 9  future_lapply-1 a3c7fd72-c4c8-e4b8-c154-03a77e8a4a35
#>                            session_uuid
#> 1  44997193-9d7d-050a-a13c-9614b8203dc7
#> 6  44997193-9d7d-050a-a13c-9614b8203dc7
#> 2  44997193-9d7d-050a-a13c-9614b8203dc7
#> 3  44997193-9d7d-050a-a13c-9614b8203dc7
#> 4  44997193-9d7d-050a-a13c-9614b8203dc7
#> 5  44997193-9d7d-050a-a13c-9614b8203dc7
#> 10 63fd3903-d2e1-d1e8-4d9c-9b89f027c696
#> 7  44997193-9d7d-050a-a13c-9614b8203dc7
#> 8  44997193-9d7d-050a-a13c-9614b8203dc7
#> 9  44997193-9d7d-050a-a13c-9614b8203dc7
#> 
#> [[2]]
#>              step                   start               at          duration
#> 1          create 2022-03-09 17:59:47.247 0.000000000 secs 2.345085e-03 secs
#> 6          launch 2022-03-09 17:59:47.253 0.005711555 secs 1.265402e-01 secs
#> 2       getWorker 2022-03-09 17:59:47.255 0.008201599 secs 5.145073e-04 secs
#> 3     eraseWorker 2022-03-09 17:59:47.258 0.010605812 secs 1.151621e-01 secs
#> 4  attachPackages 2022-03-09 17:59:47.374 0.126840830 secs 6.604195e-05 secs
#> 5   exportGlobals 2022-03-09 17:59:47.375 0.127815723 secs 2.491474e-03 secs
#> 7        resolved 2022-03-09 17:59:47.394 0.147044420 secs 1.170588e-02 secs
#> 12       evaluate 2022-03-09 17:59:47.422 0.175366402 secs 3.978889e-01 secs
#> 8        resolved 2022-03-09 17:59:47.520 0.273427725 secs 1.171041e-02 secs
#> 9        resolved 2022-03-09 17:59:47.655 0.407881498 secs 1.279593e-02 secs
#> 10       resolved 2022-03-09 17:59:47.770 0.523483038 secs 1.333857e-02 secs
#> 11         gather 2022-03-09 17:59:47.890 0.643003702 secs 1.354218e-04 secs
#>       future_label                          future_uuid
#> 1  future_lapply-2 38c646c1-dc60-9be9-a6b4-37615df737e6
#> 6  future_lapply-2 38c646c1-dc60-9be9-a6b4-37615df737e6
#> 2  future_lapply-2 38c646c1-dc60-9be9-a6b4-37615df737e6
#> 3  future_lapply-2 38c646c1-dc60-9be9-a6b4-37615df737e6
#> 4  future_lapply-2 38c646c1-dc60-9be9-a6b4-37615df737e6
#> 5  future_lapply-2 38c646c1-dc60-9be9-a6b4-37615df737e6
#> 7  future_lapply-2 38c646c1-dc60-9be9-a6b4-37615df737e6
#> 12 future_lapply-2 38c646c1-dc60-9be9-a6b4-37615df737e6
#> 8  future_lapply-2 38c646c1-dc60-9be9-a6b4-37615df737e6
#> 9  future_lapply-2 38c646c1-dc60-9be9-a6b4-37615df737e6
#> 10 future_lapply-2 38c646c1-dc60-9be9-a6b4-37615df737e6
#> 11 future_lapply-2 38c646c1-dc60-9be9-a6b4-37615df737e6
#>                            session_uuid
#> 1  44997193-9d7d-050a-a13c-9614b8203dc7
#> 6  44997193-9d7d-050a-a13c-9614b8203dc7
#> 2  44997193-9d7d-050a-a13c-9614b8203dc7
#> 3  44997193-9d7d-050a-a13c-9614b8203dc7
#> 4  44997193-9d7d-050a-a13c-9614b8203dc7
#> 5  44997193-9d7d-050a-a13c-9614b8203dc7
#> 7  44997193-9d7d-050a-a13c-9614b8203dc7
#> 12 de9a73d8-0517-bf4e-c2c9-836741a5e6d2
#> 8  44997193-9d7d-050a-a13c-9614b8203dc7
#> 9  44997193-9d7d-050a-a13c-9614b8203dc7
#> 10 44997193-9d7d-050a-a13c-9614b8203dc7
#> 11 44997193-9d7d-050a-a13c-9614b8203dc7
#> 

## Shut down parallel workers
plan(sequential)