Evaluate an R expression while collecting journals from completed futures

Evaluate an R expression while collecting journals from completed futures

gather_journals(expr, substitute = TRUE, envir = parent.frame())

Arguments

expr

The R expression to evaluate

substitute

If TRUE, then expr is subtituted, otherwise not.

envir

The environment where expr should be evaluated

Value

A list of FutureJournal:s.

Examples

library(future.apply)
#> Loading required package: future

slow_fcn <- function(x) {
  Sys.sleep(x / 10)
  sqrt(x)
}

plan(multisession, workers = 2)
js <- gather_journals({
  vs <- future_lapply(3:1, FUN = slow_fcn)
})
print(js)
#> [[1]]
#>              step                   start               at          duration
#> 1          create 2022-03-09 18:17:19.995 0.000000000 secs 1.581907e-03 secs
#> 6          launch 2022-03-09 18:17:19.999 0.004158735 secs 5.574155e-02 secs
#> 2       getWorker 2022-03-09 18:17:20.000 0.005308151 secs 2.510548e-04 secs
#> 3     eraseWorker 2022-03-09 18:17:20.002 0.006817818 secs 8.172989e-04 secs
#> 4  attachPackages 2022-03-09 18:17:20.004 0.008734226 secs 6.818771e-05 secs
#> 5   exportGlobals 2022-03-09 18:17:20.005 0.009809732 secs 4.605436e-02 secs
#> 10       evaluate 2022-03-09 18:17:20.099 0.103916645 secs 3.039248e-01 secs
#> 7        resolved 2022-03-09 18:17:20.215 0.219864130 secs 1.214623e-02 secs
#> 8        resolved 2022-03-09 18:17:20.341 0.346513510 secs 1.251435e-02 secs
#> 9          gather 2022-03-09 18:17:20.474 0.478917599 secs 1.235008e-04 secs
#>       future_label                          future_uuid
#> 1  future_lapply-1 4982183b-df76-4997-ed46-d6ccab257c7e
#> 6  future_lapply-1 4982183b-df76-4997-ed46-d6ccab257c7e
#> 2  future_lapply-1 4982183b-df76-4997-ed46-d6ccab257c7e
#> 3  future_lapply-1 4982183b-df76-4997-ed46-d6ccab257c7e
#> 4  future_lapply-1 4982183b-df76-4997-ed46-d6ccab257c7e
#> 5  future_lapply-1 4982183b-df76-4997-ed46-d6ccab257c7e
#> 10 future_lapply-1 4982183b-df76-4997-ed46-d6ccab257c7e
#> 7  future_lapply-1 4982183b-df76-4997-ed46-d6ccab257c7e
#> 8  future_lapply-1 4982183b-df76-4997-ed46-d6ccab257c7e
#> 9  future_lapply-1 4982183b-df76-4997-ed46-d6ccab257c7e
#>                            session_uuid
#> 1  5c22e3c2-7a07-5f6c-11ef-ec8e7db7559f
#> 6  5c22e3c2-7a07-5f6c-11ef-ec8e7db7559f
#> 2  5c22e3c2-7a07-5f6c-11ef-ec8e7db7559f
#> 3  5c22e3c2-7a07-5f6c-11ef-ec8e7db7559f
#> 4  5c22e3c2-7a07-5f6c-11ef-ec8e7db7559f
#> 5  5c22e3c2-7a07-5f6c-11ef-ec8e7db7559f
#> 10 3ddb1684-2fe8-bfdf-6930-8384763d8a00
#> 7  5c22e3c2-7a07-5f6c-11ef-ec8e7db7559f
#> 8  5c22e3c2-7a07-5f6c-11ef-ec8e7db7559f
#> 9  5c22e3c2-7a07-5f6c-11ef-ec8e7db7559f
#> 
#> [[2]]
#>              step                   start               at          duration
#> 1          create 2022-03-09 18:17:20.058 0.000000000 secs 2.317429e-03 secs
#> 6          launch 2022-03-09 18:17:20.064 0.005751133 secs 1.490655e-01 secs
#> 2       getWorker 2022-03-09 18:17:20.067 0.008122444 secs 5.061626e-04 secs
#> 3     eraseWorker 2022-03-09 18:17:20.069 0.010541201 secs 1.369939e-01 secs
#> 4  attachPackages 2022-03-09 18:17:20.207 0.148644447 secs 7.653236e-05 secs
#> 5   exportGlobals 2022-03-09 18:17:20.208 0.149682283 secs 3.183126e-03 secs
#> 7        resolved 2022-03-09 18:17:20.228 0.169597387 secs 1.176906e-02 secs
#> 12       evaluate 2022-03-09 18:17:20.254 0.196013689 secs 4.586196e-01 secs
#> 8        resolved 2022-03-09 18:17:20.356 0.297394991 secs 1.239657e-02 secs
#> 9        resolved 2022-03-09 18:17:20.490 0.431609869 secs 1.240897e-02 secs
#> 10       resolved 2022-03-09 18:17:20.605 0.546939611 secs 1.321316e-02 secs
#> 11         gather 2022-03-09 18:17:20.725 0.666794062 secs 1.499653e-04 secs
#>       future_label                          future_uuid
#> 1  future_lapply-2 7bf1716c-f185-44e5-9436-1bf075db26ce
#> 6  future_lapply-2 7bf1716c-f185-44e5-9436-1bf075db26ce
#> 2  future_lapply-2 7bf1716c-f185-44e5-9436-1bf075db26ce
#> 3  future_lapply-2 7bf1716c-f185-44e5-9436-1bf075db26ce
#> 4  future_lapply-2 7bf1716c-f185-44e5-9436-1bf075db26ce
#> 5  future_lapply-2 7bf1716c-f185-44e5-9436-1bf075db26ce
#> 7  future_lapply-2 7bf1716c-f185-44e5-9436-1bf075db26ce
#> 12 future_lapply-2 7bf1716c-f185-44e5-9436-1bf075db26ce
#> 8  future_lapply-2 7bf1716c-f185-44e5-9436-1bf075db26ce
#> 9  future_lapply-2 7bf1716c-f185-44e5-9436-1bf075db26ce
#> 10 future_lapply-2 7bf1716c-f185-44e5-9436-1bf075db26ce
#> 11 future_lapply-2 7bf1716c-f185-44e5-9436-1bf075db26ce
#>                            session_uuid
#> 1  5c22e3c2-7a07-5f6c-11ef-ec8e7db7559f
#> 6  5c22e3c2-7a07-5f6c-11ef-ec8e7db7559f
#> 2  5c22e3c2-7a07-5f6c-11ef-ec8e7db7559f
#> 3  5c22e3c2-7a07-5f6c-11ef-ec8e7db7559f
#> 4  5c22e3c2-7a07-5f6c-11ef-ec8e7db7559f
#> 5  5c22e3c2-7a07-5f6c-11ef-ec8e7db7559f
#> 7  5c22e3c2-7a07-5f6c-11ef-ec8e7db7559f
#> 12 e33e77b0-3a05-0828-ce09-15d7f5990532
#> 8  5c22e3c2-7a07-5f6c-11ef-ec8e7db7559f
#> 9  5c22e3c2-7a07-5f6c-11ef-ec8e7db7559f
#> 10 5c22e3c2-7a07-5f6c-11ef-ec8e7db7559f
#> 11 5c22e3c2-7a07-5f6c-11ef-ec8e7db7559f
#> 

## Shut down parallel workers
plan(sequential)