Gets the logged journal of events for a future

WARNING: This function is under development. It can change at any time. For now, please, do not depend on this function in a published R package.

Usage

journal(x, ...)

Arguments

x

A future::Future object.

...

Not used.

Value

A data frame of class FutureJournal with columns:

event (character string) - type of event that took place

category (character string) - the category of the event

parent (character string) - (to be describe)

start (POSIXct) - the timestamp when the event started

at (difftime) - the time when the event started relative to first event

duration (difftime) - the duration of the event

future_label (character string) - the label of the future

future_uuid (character string) - the UUID of the future

session_uuid (character string) - the UUID of the R session where the event took place

The common events are:

create - the future was created (an overhead)

launch - the future was launched (an overhead)

evaluate - the future was evaluated (an evaluation)

resolved - the future was queried (may be occur multiple times) (an overhead)

gather - the results was retrieved (an overhead)

but others may be added by other Future classes.

Common event categories are:

evaluation - processing time is spent on evaluation

overhead - processing time is spent on orchestrating the future

waiting - processing time is spent on waiting to set up or querying the future

but others may be added by other Future classes.

The data frame is sorted by the at time. Note that the timestamps for the evaluate event are based on the local time on the worker. The system clocks on the worker and the calling R system may not be in perfect sync.

Enabling and disabling event logging

To enable logging of events, set option future.journal is TRUE. To disable, set it to FALSE (default).

See also

Examples

## Enable journaling of futures
oopts <- options(future.journal = TRUE)

plan(multisession, workers = 2L)

t_start <- Sys.time()
fs <- lapply(1:3, FUN = function(x) future({ Sys.sleep(x); sqrt(x) }))
vs <- value(fs)
js <- lapply(fs, FUN = journal, baseline = t_start)
print(js)
#> [[1]]
#>            event   category parent               start               at
#> 1         create   overhead   <NA> 2026-04-07 08:57:35 0.000000000 secs
#> 4         launch   overhead   <NA> 2026-04-07 08:57:35 0.006395340 secs
#> 2      getWorker   overhead launch 2026-04-07 08:57:35 0.006527901 secs
#> 3    eraseWorker   overhead launch 2026-04-07 08:57:35 0.007649660 secs
#> 5       resolved       <NA>   <NA> 2026-04-07 08:57:35 0.055186987 secs
#> 6       resolved       <NA>   <NA> 2026-04-07 08:57:35 0.089033127 secs
#> 7       resolved       <NA>   <NA> 2026-04-07 08:57:35 0.131754875 secs
#> 8       resolved       <NA>   <NA> 2026-04-07 08:57:35 0.165295839 secs
#> 9       resolved       <NA>   <NA> 2026-04-07 08:57:35 0.198984861 secs
#> 41      evaluate evaluation   <NA> 2026-04-07 08:57:35 0.207314253 secs
#> 10      resolved       <NA>   <NA> 2026-04-07 08:57:36 0.232787371 secs
#> 11      resolved       <NA>   <NA> 2026-04-07 08:57:36 0.266866207 secs
#> 12      resolved       <NA>   <NA> 2026-04-07 08:57:36 0.301012039 secs
#> 13      resolved       <NA>   <NA> 2026-04-07 08:57:36 0.335442543 secs
#> 14      resolved       <NA>   <NA> 2026-04-07 08:57:36 0.370602131 secs
#> 15      resolved       <NA>   <NA> 2026-04-07 08:57:36 0.408231020 secs
#> 16      resolved       <NA>   <NA> 2026-04-07 08:57:36 0.445653439 secs
#> 17      resolved       <NA>   <NA> 2026-04-07 08:57:36 0.481858969 secs
#> 18      resolved       <NA>   <NA> 2026-04-07 08:57:36 0.519630432 secs
#> 19      resolved       <NA>   <NA> 2026-04-07 08:57:36 0.556129217 secs
#> 20      resolved       <NA>   <NA> 2026-04-07 08:57:36 0.593373299 secs
#> 21      resolved       <NA>   <NA> 2026-04-07 08:57:36 0.630234003 secs
#> 22      resolved       <NA>   <NA> 2026-04-07 08:57:36 0.668174744 secs
#> 23      resolved       <NA>   <NA> 2026-04-07 08:57:36 0.706638575 secs
#> 24      resolved       <NA>   <NA> 2026-04-07 08:57:36 0.744083881 secs
#> 25      resolved       <NA>   <NA> 2026-04-07 08:57:36 0.782227516 secs
#> 26      resolved       <NA>   <NA> 2026-04-07 08:57:36 0.819393158 secs
#> 27      resolved       <NA>   <NA> 2026-04-07 08:57:36 0.857527733 secs
#> 28      resolved       <NA>   <NA> 2026-04-07 08:57:36 0.896339655 secs
#> 29      resolved       <NA>   <NA> 2026-04-07 08:57:36 0.933403254 secs
#> 30      resolved       <NA>   <NA> 2026-04-07 08:57:36 0.972538948 secs
#> 31      resolved       <NA>   <NA> 2026-04-07 08:57:36 1.011084557 secs
#> 32      resolved       <NA>   <NA> 2026-04-07 08:57:36 1.050303936 secs
#> 33      resolved       <NA>   <NA> 2026-04-07 08:57:36 1.089630604 secs
#> 34      resolved       <NA>   <NA> 2026-04-07 08:57:36 1.128599644 secs
#> 35      resolved       <NA>   <NA> 2026-04-07 08:57:36 1.168787479 secs
#> 36      resolved       <NA>   <NA> 2026-04-07 08:57:36 1.209238291 secs
#> 37      resolved       <NA>   <NA> 2026-04-07 08:57:37 1.248574257 secs
#> 38      resolved       <NA>   <NA> 2026-04-07 08:57:37 1.288259506 secs
#> 42      resolved       <NA>   <NA> 2026-04-07 08:57:37 1.328148842 secs
#> 39 receiveResult   overhead gather 2026-04-07 08:57:37 1.334411621 secs
#> 40        gather   overhead   <NA> 2026-04-07 08:57:37 1.336301088 secs
#> 43      resolved       <NA>   <NA> 2026-04-07 08:57:37 1.352408886 secs
#> 44       resolve   overhead   <NA> 2026-04-07 08:57:37 1.353143930 secs
#>             duration future_label                         future_uuid
#> 1  6.092310e-03 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-17
#> 4  6.464243e-03 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-17
#> 2  5.323887e-04 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-17
#> 3  3.428459e-04 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-17
#> 5  1.092625e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-17
#> 6  1.884675e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-17
#> 7  1.077366e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-17
#> 8  1.080298e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-17
#> 9  1.074147e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-17
#> 41 1.125111e+00 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-17
#> 10 1.078987e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-17
#> 11 1.079631e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-17
#> 12 1.083302e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-17
#> 13 1.085019e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-17
#> 14 1.151180e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-17
#> 15 1.150417e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-17
#> 16 1.120830e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-17
#> 17 1.144934e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-17
#> 18 1.126099e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-17
#> 19 1.117897e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-17
#> 20 1.131916e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-17
#> 21 1.117253e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-17
#> 22 1.151133e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-17
#> 23 1.129842e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-17
#> 24 1.145220e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-17
#> 25 1.073599e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-17
#> 26 1.135874e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-17
#> 27 1.142454e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-17
#> 28 1.134157e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-17
#> 29 1.135468e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-17
#> 30 1.124048e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-17
#> 31 1.133823e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-17
#> 32 1.139164e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-17
#> 33 1.133370e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-17
#> 34 1.197195e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-17
#> 35 1.128149e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-17
#> 36 1.130986e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-17
#> 37 1.122832e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-17
#> 38 1.125240e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-17
#> 42 1.364422e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-17
#> 39 3.445148e-04 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-17
#> 40 1.282692e-04 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-17
#> 43 3.480911e-05 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-17
#> 44 9.536743e-06 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-17
#>                        session_uuid
#> 1  fec0ec46ebd2534db0dd26773e1d198d
#> 4  fec0ec46ebd2534db0dd26773e1d198d
#> 2  fec0ec46ebd2534db0dd26773e1d198d
#> 3  fec0ec46ebd2534db0dd26773e1d198d
#> 5  fec0ec46ebd2534db0dd26773e1d198d
#> 6  fec0ec46ebd2534db0dd26773e1d198d
#> 7  fec0ec46ebd2534db0dd26773e1d198d
#> 8  fec0ec46ebd2534db0dd26773e1d198d
#> 9  fec0ec46ebd2534db0dd26773e1d198d
#> 41 fec0ec46ebd2534db0dd26773e1d198d
#> 10 fec0ec46ebd2534db0dd26773e1d198d
#> 11 fec0ec46ebd2534db0dd26773e1d198d
#> 12 fec0ec46ebd2534db0dd26773e1d198d
#> 13 fec0ec46ebd2534db0dd26773e1d198d
#> 14 fec0ec46ebd2534db0dd26773e1d198d
#> 15 fec0ec46ebd2534db0dd26773e1d198d
#> 16 fec0ec46ebd2534db0dd26773e1d198d
#> 17 fec0ec46ebd2534db0dd26773e1d198d
#> 18 fec0ec46ebd2534db0dd26773e1d198d
#> 19 fec0ec46ebd2534db0dd26773e1d198d
#> 20 fec0ec46ebd2534db0dd26773e1d198d
#> 21 fec0ec46ebd2534db0dd26773e1d198d
#> 22 fec0ec46ebd2534db0dd26773e1d198d
#> 23 fec0ec46ebd2534db0dd26773e1d198d
#> 24 fec0ec46ebd2534db0dd26773e1d198d
#> 25 fec0ec46ebd2534db0dd26773e1d198d
#> 26 fec0ec46ebd2534db0dd26773e1d198d
#> 27 fec0ec46ebd2534db0dd26773e1d198d
#> 28 fec0ec46ebd2534db0dd26773e1d198d
#> 29 fec0ec46ebd2534db0dd26773e1d198d
#> 30 fec0ec46ebd2534db0dd26773e1d198d
#> 31 fec0ec46ebd2534db0dd26773e1d198d
#> 32 fec0ec46ebd2534db0dd26773e1d198d
#> 33 fec0ec46ebd2534db0dd26773e1d198d
#> 34 fec0ec46ebd2534db0dd26773e1d198d
#> 35 fec0ec46ebd2534db0dd26773e1d198d
#> 36 fec0ec46ebd2534db0dd26773e1d198d
#> 37 fec0ec46ebd2534db0dd26773e1d198d
#> 38 fec0ec46ebd2534db0dd26773e1d198d
#> 42 fec0ec46ebd2534db0dd26773e1d198d
#> 39 fec0ec46ebd2534db0dd26773e1d198d
#> 40 fec0ec46ebd2534db0dd26773e1d198d
#> 43 fec0ec46ebd2534db0dd26773e1d198d
#> 44 fec0ec46ebd2534db0dd26773e1d198d
#> 
#> [[2]]
#>            event   category parent               start              at
#> 1         create   overhead   <NA> 2026-04-07 08:57:35 0.00000000 secs
#> 4         launch   overhead   <NA> 2026-04-07 08:57:35 0.01188278 secs
#> 2      getWorker   overhead launch 2026-04-07 08:57:35 0.01207590 secs
#> 3    eraseWorker   overhead launch 2026-04-07 08:57:35 0.01378870 secs
#> 5       resolved       <NA>   <NA> 2026-04-07 08:57:35 0.05306315 secs
#> 6       resolved       <NA>   <NA> 2026-04-07 08:57:35 0.09536695 secs
#> 7       resolved       <NA>   <NA> 2026-04-07 08:57:35 0.12951827 secs
#> 8       resolved       <NA>   <NA> 2026-04-07 08:57:35 0.16308188 secs
#> 9       resolved       <NA>   <NA> 2026-04-07 08:57:35 0.19675732 secs
#> 69      evaluate evaluation   <NA> 2026-04-07 08:57:35 0.20071793 secs
#> 10      resolved       <NA>   <NA> 2026-04-07 08:57:36 0.23058939 secs
#> 11      resolved       <NA>   <NA> 2026-04-07 08:57:36 0.26469493 secs
#> 12      resolved       <NA>   <NA> 2026-04-07 08:57:36 0.29890871 secs
#> 13      resolved       <NA>   <NA> 2026-04-07 08:57:36 0.33341885 secs
#> 14      resolved       <NA>   <NA> 2026-04-07 08:57:36 0.36996245 secs
#> 15      resolved       <NA>   <NA> 2026-04-07 08:57:36 0.40755391 secs
#> 16      resolved       <NA>   <NA> 2026-04-07 08:57:36 0.44431138 secs
#> 17      resolved       <NA>   <NA> 2026-04-07 08:57:36 0.48126674 secs
#> 18      resolved       <NA>   <NA> 2026-04-07 08:57:36 0.51836705 secs
#> 19      resolved       <NA>   <NA> 2026-04-07 08:57:36 0.55494690 secs
#> 20      resolved       <NA>   <NA> 2026-04-07 08:57:36 0.59225202 secs
#> 21      resolved       <NA>   <NA> 2026-04-07 08:57:36 0.62915421 secs
#> 22      resolved       <NA>   <NA> 2026-04-07 08:57:36 0.66749620 secs
#> 23      resolved       <NA>   <NA> 2026-04-07 08:57:36 0.70553160 secs
#> 24      resolved       <NA>   <NA> 2026-04-07 08:57:36 0.74409008 secs
#> 25      resolved       <NA>   <NA> 2026-04-07 08:57:36 0.78059888 secs
#> 26      resolved       <NA>   <NA> 2026-04-07 08:57:36 0.81824589 secs
#> 27      resolved       <NA>   <NA> 2026-04-07 08:57:36 0.85680676 secs
#> 28      resolved       <NA>   <NA> 2026-04-07 08:57:36 0.89468431 secs
#> 29      resolved       <NA>   <NA> 2026-04-07 08:57:36 0.93266368 secs
#> 30      resolved       <NA>   <NA> 2026-04-07 08:57:36 0.97139621 secs
#> 31      resolved       <NA>   <NA> 2026-04-07 08:57:36 1.01014781 secs
#> 32      resolved       <NA>   <NA> 2026-04-07 08:57:36 1.04938531 secs
#> 33      resolved       <NA>   <NA> 2026-04-07 08:57:36 1.08851957 secs
#> 34      resolved       <NA>   <NA> 2026-04-07 08:57:36 1.12827826 secs
#> 35      resolved       <NA>   <NA> 2026-04-07 08:57:36 1.16810584 secs
#> 36      resolved       <NA>   <NA> 2026-04-07 08:57:36 1.20816040 secs
#> 37      resolved       <NA>   <NA> 2026-04-07 08:57:37 1.24757600 secs
#> 38      resolved       <NA>   <NA> 2026-04-07 08:57:37 1.28720021 secs
#> 39      resolved       <NA>   <NA> 2026-04-07 08:57:37 1.33979297 secs
#> 40      resolved       <NA>   <NA> 2026-04-07 08:57:37 1.37335563 secs
#> 41      resolved       <NA>   <NA> 2026-04-07 08:57:37 1.40870214 secs
#> 42      resolved       <NA>   <NA> 2026-04-07 08:57:37 1.44441891 secs
#> 43      resolved       <NA>   <NA> 2026-04-07 08:57:37 1.48068833 secs
#> 44      resolved       <NA>   <NA> 2026-04-07 08:57:37 1.51662111 secs
#> 45      resolved       <NA>   <NA> 2026-04-07 08:57:37 1.55212498 secs
#> 46      resolved       <NA>   <NA> 2026-04-07 08:57:37 1.58898997 secs
#> 47      resolved       <NA>   <NA> 2026-04-07 08:57:37 1.62469411 secs
#> 48      resolved       <NA>   <NA> 2026-04-07 08:57:37 1.65999389 secs
#> 49      resolved       <NA>   <NA> 2026-04-07 08:57:37 1.69627380 secs
#> 50      resolved       <NA>   <NA> 2026-04-07 08:57:37 1.73207998 secs
#> 51      resolved       <NA>   <NA> 2026-04-07 08:57:37 1.76835632 secs
#> 52      resolved       <NA>   <NA> 2026-04-07 08:57:37 1.80430293 secs
#> 53      resolved       <NA>   <NA> 2026-04-07 08:57:37 1.83926606 secs
#> 54      resolved       <NA>   <NA> 2026-04-07 08:57:37 1.87552547 secs
#> 55      resolved       <NA>   <NA> 2026-04-07 08:57:37 1.91127753 secs
#> 56      resolved       <NA>   <NA> 2026-04-07 08:57:37 1.94668388 secs
#> 57      resolved       <NA>   <NA> 2026-04-07 08:57:37 1.98276973 secs
#> 58      resolved       <NA>   <NA> 2026-04-07 08:57:37 2.01783156 secs
#> 59      resolved       <NA>   <NA> 2026-04-07 08:57:37 2.05384207 secs
#> 60      resolved       <NA>   <NA> 2026-04-07 08:57:37 2.09014797 secs
#> 61      resolved       <NA>   <NA> 2026-04-07 08:57:37 2.12652326 secs
#> 62      resolved       <NA>   <NA> 2026-04-07 08:57:37 2.16282845 secs
#> 63      resolved       <NA>   <NA> 2026-04-07 08:57:37 2.20468259 secs
#> 64      resolved       <NA>   <NA> 2026-04-07 08:57:38 2.24083018 secs
#> 65      resolved       <NA>   <NA> 2026-04-07 08:57:38 2.27671385 secs
#> 66      resolved       <NA>   <NA> 2026-04-07 08:57:38 2.31179166 secs
#> 70      resolved       <NA>   <NA> 2026-04-07 08:57:38 2.34742451 secs
#> 67 receiveResult   overhead gather 2026-04-07 08:57:38 2.34867954 secs
#> 68        gather   overhead   <NA> 2026-04-07 08:57:38 2.35060048 secs
#> 71       resolve   overhead   <NA> 2026-04-07 08:57:38 2.35745525 secs
#>             duration future_label                         future_uuid
#> 1  1.136661e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 4  7.779837e-03 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 2  8.304119e-04 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 3  3.883839e-04 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 5  1.084542e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 6  1.144195e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 7  1.068950e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 8  1.070976e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 9  1.077080e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 69 2.126884e+00 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 10 1.082468e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 11 1.073790e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 12 1.078153e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 13 1.086116e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 14 1.139927e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 15 1.132965e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 16 1.100516e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 17 1.126719e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 18 1.101422e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 19 1.108050e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 20 1.100779e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 21 1.121974e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 22 1.140261e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 23 1.100588e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 24 1.114130e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 25 1.108408e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 26 1.110697e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 27 1.129436e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 28 1.073146e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 29 1.133466e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 30 1.122046e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 31 1.136112e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 32 1.127362e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 33 1.103759e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 34 1.128554e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 35 1.186538e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 36 1.105332e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 37 1.124215e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 38 1.116467e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 39 1.049352e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 40 1.135540e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 41 1.179671e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 42 1.215434e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 43 1.142812e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 44 1.122570e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 45 1.183963e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 46 1.129484e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 47 1.126289e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 48 1.140881e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 49 1.155043e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 50 1.154780e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 51 1.142383e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 52 1.145291e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 53 1.121020e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 54 1.151109e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 55 1.123428e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 56 1.135230e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 57 1.115012e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 58 1.137710e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 59 1.155806e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 60 1.147175e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 61 1.154828e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 62 1.129103e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 63 1.153946e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 64 1.154351e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 65 1.113343e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 66 1.129723e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 70 7.914543e-03 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 67 2.889633e-04 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 68 1.244545e-04 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#> 71 2.813339e-05 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-18
#>                        session_uuid
#> 1  fec0ec46ebd2534db0dd26773e1d198d
#> 4  fec0ec46ebd2534db0dd26773e1d198d
#> 2  fec0ec46ebd2534db0dd26773e1d198d
#> 3  fec0ec46ebd2534db0dd26773e1d198d
#> 5  fec0ec46ebd2534db0dd26773e1d198d
#> 6  fec0ec46ebd2534db0dd26773e1d198d
#> 7  fec0ec46ebd2534db0dd26773e1d198d
#> 8  fec0ec46ebd2534db0dd26773e1d198d
#> 9  fec0ec46ebd2534db0dd26773e1d198d
#> 69 fec0ec46ebd2534db0dd26773e1d198d
#> 10 fec0ec46ebd2534db0dd26773e1d198d
#> 11 fec0ec46ebd2534db0dd26773e1d198d
#> 12 fec0ec46ebd2534db0dd26773e1d198d
#> 13 fec0ec46ebd2534db0dd26773e1d198d
#> 14 fec0ec46ebd2534db0dd26773e1d198d
#> 15 fec0ec46ebd2534db0dd26773e1d198d
#> 16 fec0ec46ebd2534db0dd26773e1d198d
#> 17 fec0ec46ebd2534db0dd26773e1d198d
#> 18 fec0ec46ebd2534db0dd26773e1d198d
#> 19 fec0ec46ebd2534db0dd26773e1d198d
#> 20 fec0ec46ebd2534db0dd26773e1d198d
#> 21 fec0ec46ebd2534db0dd26773e1d198d
#> 22 fec0ec46ebd2534db0dd26773e1d198d
#> 23 fec0ec46ebd2534db0dd26773e1d198d
#> 24 fec0ec46ebd2534db0dd26773e1d198d
#> 25 fec0ec46ebd2534db0dd26773e1d198d
#> 26 fec0ec46ebd2534db0dd26773e1d198d
#> 27 fec0ec46ebd2534db0dd26773e1d198d
#> 28 fec0ec46ebd2534db0dd26773e1d198d
#> 29 fec0ec46ebd2534db0dd26773e1d198d
#> 30 fec0ec46ebd2534db0dd26773e1d198d
#> 31 fec0ec46ebd2534db0dd26773e1d198d
#> 32 fec0ec46ebd2534db0dd26773e1d198d
#> 33 fec0ec46ebd2534db0dd26773e1d198d
#> 34 fec0ec46ebd2534db0dd26773e1d198d
#> 35 fec0ec46ebd2534db0dd26773e1d198d
#> 36 fec0ec46ebd2534db0dd26773e1d198d
#> 37 fec0ec46ebd2534db0dd26773e1d198d
#> 38 fec0ec46ebd2534db0dd26773e1d198d
#> 39 fec0ec46ebd2534db0dd26773e1d198d
#> 40 fec0ec46ebd2534db0dd26773e1d198d
#> 41 fec0ec46ebd2534db0dd26773e1d198d
#> 42 fec0ec46ebd2534db0dd26773e1d198d
#> 43 fec0ec46ebd2534db0dd26773e1d198d
#> 44 fec0ec46ebd2534db0dd26773e1d198d
#> 45 fec0ec46ebd2534db0dd26773e1d198d
#> 46 fec0ec46ebd2534db0dd26773e1d198d
#> 47 fec0ec46ebd2534db0dd26773e1d198d
#> 48 fec0ec46ebd2534db0dd26773e1d198d
#> 49 fec0ec46ebd2534db0dd26773e1d198d
#> 50 fec0ec46ebd2534db0dd26773e1d198d
#> 51 fec0ec46ebd2534db0dd26773e1d198d
#> 52 fec0ec46ebd2534db0dd26773e1d198d
#> 53 fec0ec46ebd2534db0dd26773e1d198d
#> 54 fec0ec46ebd2534db0dd26773e1d198d
#> 55 fec0ec46ebd2534db0dd26773e1d198d
#> 56 fec0ec46ebd2534db0dd26773e1d198d
#> 57 fec0ec46ebd2534db0dd26773e1d198d
#> 58 fec0ec46ebd2534db0dd26773e1d198d
#> 59 fec0ec46ebd2534db0dd26773e1d198d
#> 60 fec0ec46ebd2534db0dd26773e1d198d
#> 61 fec0ec46ebd2534db0dd26773e1d198d
#> 62 fec0ec46ebd2534db0dd26773e1d198d
#> 63 fec0ec46ebd2534db0dd26773e1d198d
#> 64 fec0ec46ebd2534db0dd26773e1d198d
#> 65 fec0ec46ebd2534db0dd26773e1d198d
#> 66 fec0ec46ebd2534db0dd26773e1d198d
#> 70 fec0ec46ebd2534db0dd26773e1d198d
#> 67 fec0ec46ebd2534db0dd26773e1d198d
#> 68 fec0ec46ebd2534db0dd26773e1d198d
#> 71 fec0ec46ebd2534db0dd26773e1d198d
#> 
#> [[3]]
#>             event   category parent               start              at
#> 1          create   overhead   <NA> 2026-04-07 08:57:35 0.00000000 secs
#> 4          launch   overhead   <NA> 2026-04-07 08:57:35 0.01042485 secs
#> 2       getWorker   overhead launch 2026-04-07 08:57:35 0.01055121 secs
#> 3     eraseWorker   overhead launch 2026-04-07 08:57:37 1.31277156 secs
#> 121      evaluate evaluation   <NA> 2026-04-07 08:57:37 1.31879902 secs
#> 5        resolved       <NA>   <NA> 2026-04-07 08:57:37 1.33031583 secs
#> 6        resolved       <NA>   <NA> 2026-04-07 08:57:37 1.36540008 secs
#> 7        resolved       <NA>   <NA> 2026-04-07 08:57:37 1.40131450 secs
#> 8        resolved       <NA>   <NA> 2026-04-07 08:57:37 1.43737555 secs
#> 9        resolved       <NA>   <NA> 2026-04-07 08:57:37 1.47322989 secs
#> 10       resolved       <NA>   <NA> 2026-04-07 08:57:37 1.50874329 secs
#> 11       resolved       <NA>   <NA> 2026-04-07 08:57:37 1.54546285 secs
#> 12       resolved       <NA>   <NA> 2026-04-07 08:57:37 1.58124137 secs
#> 13       resolved       <NA>   <NA> 2026-04-07 08:57:37 1.61676717 secs
#> 14       resolved       <NA>   <NA> 2026-04-07 08:57:37 1.65243173 secs
#> 15       resolved       <NA>   <NA> 2026-04-07 08:57:37 1.68879414 secs
#> 16       resolved       <NA>   <NA> 2026-04-07 08:57:37 1.72463775 secs
#> 17       resolved       <NA>   <NA> 2026-04-07 08:57:37 1.76074409 secs
#> 18       resolved       <NA>   <NA> 2026-04-07 08:57:37 1.79631066 secs
#> 19       resolved       <NA>   <NA> 2026-04-07 08:57:37 1.83137584 secs
#> 20       resolved       <NA>   <NA> 2026-04-07 08:57:37 1.86802506 secs
#> 21       resolved       <NA>   <NA> 2026-04-07 08:57:37 1.90333247 secs
#> 22       resolved       <NA>   <NA> 2026-04-07 08:57:37 1.93905020 secs
#> 23       resolved       <NA>   <NA> 2026-04-07 08:57:37 1.97466302 secs
#> 24       resolved       <NA>   <NA> 2026-04-07 08:57:37 2.01019764 secs
#> 25       resolved       <NA>   <NA> 2026-04-07 08:57:37 2.04646969 secs
#> 26       resolved       <NA>   <NA> 2026-04-07 08:57:37 2.08265209 secs
#> 27       resolved       <NA>   <NA> 2026-04-07 08:57:37 2.11906409 secs
#> 28       resolved       <NA>   <NA> 2026-04-07 08:57:37 2.15485930 secs
#> 29       resolved       <NA>   <NA> 2026-04-07 08:57:38 2.19720840 secs
#> 30       resolved       <NA>   <NA> 2026-04-07 08:57:38 2.23333168 secs
#> 31       resolved       <NA>   <NA> 2026-04-07 08:57:38 2.26845574 secs
#> 32       resolved       <NA>   <NA> 2026-04-07 08:57:38 2.30401111 secs
#> 33       resolved       <NA>   <NA> 2026-04-07 08:57:38 2.33799219 secs
#> 34       resolved       <NA>   <NA> 2026-04-07 08:57:38 2.36075377 secs
#> 35       resolved       <NA>   <NA> 2026-04-07 08:57:38 2.38340068 secs
#> 36       resolved       <NA>   <NA> 2026-04-07 08:57:38 2.40625167 secs
#> 37       resolved       <NA>   <NA> 2026-04-07 08:57:38 2.42940354 secs
#> 38       resolved       <NA>   <NA> 2026-04-07 08:57:38 2.45232201 secs
#> 39       resolved       <NA>   <NA> 2026-04-07 08:57:38 2.47475505 secs
#> 40       resolved       <NA>   <NA> 2026-04-07 08:57:38 2.49796176 secs
#> 41       resolved       <NA>   <NA> 2026-04-07 08:57:38 2.52116847 secs
#> 42       resolved       <NA>   <NA> 2026-04-07 08:57:38 2.54436135 secs
#> 43       resolved       <NA>   <NA> 2026-04-07 08:57:38 2.56744266 secs
#> 44       resolved       <NA>   <NA> 2026-04-07 08:57:38 2.59064293 secs
#> 45       resolved       <NA>   <NA> 2026-04-07 08:57:38 2.61384773 secs
#> 46       resolved       <NA>   <NA> 2026-04-07 08:57:38 2.63681960 secs
#> 47       resolved       <NA>   <NA> 2026-04-07 08:57:38 2.65946770 secs
#> 48       resolved       <NA>   <NA> 2026-04-07 08:57:38 2.68234944 secs
#> 49       resolved       <NA>   <NA> 2026-04-07 08:57:38 2.70571136 secs
#> 50       resolved       <NA>   <NA> 2026-04-07 08:57:38 2.72869611 secs
#> 51       resolved       <NA>   <NA> 2026-04-07 08:57:38 2.75187755 secs
#> 52       resolved       <NA>   <NA> 2026-04-07 08:57:38 2.77451134 secs
#> 53       resolved       <NA>   <NA> 2026-04-07 08:57:38 2.79738593 secs
#> 54       resolved       <NA>   <NA> 2026-04-07 08:57:38 2.82049417 secs
#> 55       resolved       <NA>   <NA> 2026-04-07 08:57:38 2.84431529 secs
#> 56       resolved       <NA>   <NA> 2026-04-07 08:57:38 2.86729312 secs
#> 57       resolved       <NA>   <NA> 2026-04-07 08:57:38 2.89063859 secs
#> 58       resolved       <NA>   <NA> 2026-04-07 08:57:38 2.91396403 secs
#> 59       resolved       <NA>   <NA> 2026-04-07 08:57:38 2.93730307 secs
#> 60       resolved       <NA>   <NA> 2026-04-07 08:57:38 2.96034384 secs
#> 61       resolved       <NA>   <NA> 2026-04-07 08:57:38 2.98370123 secs
#> 62       resolved       <NA>   <NA> 2026-04-07 08:57:38 3.00663352 secs
#> 63       resolved       <NA>   <NA> 2026-04-07 08:57:38 3.02930236 secs
#> 64       resolved       <NA>   <NA> 2026-04-07 08:57:38 3.05259728 secs
#> 65       resolved       <NA>   <NA> 2026-04-07 08:57:38 3.07556081 secs
#> 66       resolved       <NA>   <NA> 2026-04-07 08:57:38 3.09851217 secs
#> 67       resolved       <NA>   <NA> 2026-04-07 08:57:38 3.12170219 secs
#> 68       resolved       <NA>   <NA> 2026-04-07 08:57:38 3.14473510 secs
#> 69       resolved       <NA>   <NA> 2026-04-07 08:57:38 3.16795301 secs
#> 70       resolved       <NA>   <NA> 2026-04-07 08:57:39 3.19071507 secs
#> 71       resolved       <NA>   <NA> 2026-04-07 08:57:39 3.21431184 secs
#> 72       resolved       <NA>   <NA> 2026-04-07 08:57:39 3.23766494 secs
#> 73       resolved       <NA>   <NA> 2026-04-07 08:57:39 3.26082969 secs
#> 74       resolved       <NA>   <NA> 2026-04-07 08:57:39 3.28390145 secs
#> 75       resolved       <NA>   <NA> 2026-04-07 08:57:39 3.30684495 secs
#> 76       resolved       <NA>   <NA> 2026-04-07 08:57:39 3.33003139 secs
#> 77       resolved       <NA>   <NA> 2026-04-07 08:57:39 3.35317302 secs
#> 78       resolved       <NA>   <NA> 2026-04-07 08:57:39 3.37635541 secs
#> 79       resolved       <NA>   <NA> 2026-04-07 08:57:39 3.39973760 secs
#> 80       resolved       <NA>   <NA> 2026-04-07 08:57:39 3.42289233 secs
#> 81       resolved       <NA>   <NA> 2026-04-07 08:57:39 3.44598293 secs
#> 82       resolved       <NA>   <NA> 2026-04-07 08:57:39 3.46938634 secs
#> 83       resolved       <NA>   <NA> 2026-04-07 08:57:39 3.49286914 secs
#> 84       resolved       <NA>   <NA> 2026-04-07 08:57:39 3.51617432 secs
#> 85       resolved       <NA>   <NA> 2026-04-07 08:57:39 3.53928876 secs
#> 86       resolved       <NA>   <NA> 2026-04-07 08:57:39 3.56229854 secs
#> 87       resolved       <NA>   <NA> 2026-04-07 08:57:39 3.58549023 secs
#> 88       resolved       <NA>   <NA> 2026-04-07 08:57:39 3.60838509 secs
#> 89       resolved       <NA>   <NA> 2026-04-07 08:57:39 3.63116956 secs
#> 90       resolved       <NA>   <NA> 2026-04-07 08:57:39 3.65431929 secs
#> 91       resolved       <NA>   <NA> 2026-04-07 08:57:39 3.67762327 secs
#> 92       resolved       <NA>   <NA> 2026-04-07 08:57:39 3.70020151 secs
#> 93       resolved       <NA>   <NA> 2026-04-07 08:57:39 3.72308588 secs
#> 94       resolved       <NA>   <NA> 2026-04-07 08:57:39 3.74625587 secs
#> 95       resolved       <NA>   <NA> 2026-04-07 08:57:39 3.76949358 secs
#> 96       resolved       <NA>   <NA> 2026-04-07 08:57:39 3.79248166 secs
#> 97       resolved       <NA>   <NA> 2026-04-07 08:57:39 3.81564069 secs
#> 98       resolved       <NA>   <NA> 2026-04-07 08:57:39 3.83891296 secs
#> 99       resolved       <NA>   <NA> 2026-04-07 08:57:39 3.86175179 secs
#> 100      resolved       <NA>   <NA> 2026-04-07 08:57:39 3.88495541 secs
#> 101      resolved       <NA>   <NA> 2026-04-07 08:57:39 3.90763402 secs
#> 102      resolved       <NA>   <NA> 2026-04-07 08:57:39 3.93275142 secs
#> 103      resolved       <NA>   <NA> 2026-04-07 08:57:39 3.95389843 secs
#> 104      resolved       <NA>   <NA> 2026-04-07 08:57:39 3.97625160 secs
#> 105      resolved       <NA>   <NA> 2026-04-07 08:57:39 3.99946928 secs
#> 106      resolved       <NA>   <NA> 2026-04-07 08:57:39 4.02262497 secs
#> 107      resolved       <NA>   <NA> 2026-04-07 08:57:39 4.04608202 secs
#> 108      resolved       <NA>   <NA> 2026-04-07 08:57:39 4.06951046 secs
#> 109      resolved       <NA>   <NA> 2026-04-07 08:57:39 4.09269214 secs
#> 110      resolved       <NA>   <NA> 2026-04-07 08:57:39 4.11539102 secs
#> 111      resolved       <NA>   <NA> 2026-04-07 08:57:39 4.13829541 secs
#> 112      resolved       <NA>   <NA> 2026-04-07 08:57:39 4.16145062 secs
#> 113      resolved       <NA>   <NA> 2026-04-07 08:57:39 4.18432713 secs
#> 114      resolved       <NA>   <NA> 2026-04-07 08:57:40 4.20763993 secs
#> 115      resolved       <NA>   <NA> 2026-04-07 08:57:40 4.23069072 secs
#> 116      resolved       <NA>   <NA> 2026-04-07 08:57:40 4.25388432 secs
#> 117      resolved       <NA>   <NA> 2026-04-07 08:57:40 4.27690315 secs
#> 118      resolved       <NA>   <NA> 2026-04-07 08:57:40 4.29991531 secs
#> 122      resolved       <NA>   <NA> 2026-04-07 08:57:40 4.32302189 secs
#> 119 receiveResult   overhead gather 2026-04-07 08:57:40 4.32686400 secs
#> 120        gather   overhead   <NA> 2026-04-07 08:57:40 4.32892585 secs
#> 123       resolve   overhead   <NA> 2026-04-07 08:57:40 4.33563375 secs
#>              duration future_label                         future_uuid
#> 1   1.011109e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 4   1.306635e+00 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 2   1.301031e+00 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 3   4.732609e-04 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 121 3.006358e+00 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 5   1.069522e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 6   1.100850e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 7   1.111531e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 8   1.118398e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 9   1.118898e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 10  1.120257e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 11  1.123333e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 12  1.120210e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 13  1.112294e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 14  1.138353e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 15  1.121521e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 16  1.128721e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 17  1.125669e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 18  1.075435e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 19  1.175117e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 20  1.123142e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 21  1.107955e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 22  1.143551e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 23  1.105928e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 24  1.127267e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 25  1.134109e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 26  1.130533e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 27  1.145601e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 28  1.109958e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 29  1.124930e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 30  1.123834e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 31  1.108193e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 32  1.117873e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 33  1.112103e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 34  1.127005e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 35  1.122570e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 36  1.134229e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 37  1.143909e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 38  1.104784e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 39  1.144385e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 40  1.148677e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 41  1.127887e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 42  1.145530e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 43  1.148796e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 44  1.145458e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 45  1.135159e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 46  1.115751e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 47  1.118660e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 48  1.156783e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 49  1.147318e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 50  1.167607e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 51  1.118112e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 52  1.121330e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 53  1.141715e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 54  1.132536e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 55  1.135039e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 56  1.152778e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 57  1.153111e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 58  1.143456e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 59  1.142454e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 60  1.153469e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 61  1.137686e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 62  1.109076e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 63  1.149774e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 64  1.143289e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 65  1.120496e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 66  1.146984e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 67  1.146865e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 68  1.138926e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 69  1.113963e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 70  1.136994e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 71  1.160955e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 72  1.129413e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 73  1.143789e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 74  1.120925e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 75  1.145530e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 76  1.136994e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 77  1.136899e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 78  1.158690e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 79  1.136518e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 80  1.134634e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 81  1.158214e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 82  1.160240e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 83  1.149154e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 84  1.142406e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 85  1.142859e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 86  1.140475e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 87  1.129770e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 88  1.121402e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 89  1.145482e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 90  1.151013e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 91  1.125288e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 92  1.115251e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 93  1.140690e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 94  1.149702e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 95  1.134062e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 96  1.144314e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 97  1.153135e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 98  1.147294e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 99  1.145554e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 100 1.129746e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 101 1.315117e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 102 1.056337e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 103 1.065636e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 104 1.144028e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 105 1.145864e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 106 1.159573e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 107 1.156282e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 108 1.147914e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 109 1.117039e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 110 1.131344e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 111 1.128650e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 112 1.145363e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 113 1.142049e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 114 1.137233e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 115 1.155853e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 116 1.136518e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 117 1.138091e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 118 1.130319e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 122 1.048708e-02 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 119 3.526211e-04 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 120 1.425743e-04 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#> 123 3.385544e-05 secs         <NA> fec0ec46ebd2534db0dd26773e1d198d-19
#>                         session_uuid
#> 1   fec0ec46ebd2534db0dd26773e1d198d
#> 4   fec0ec46ebd2534db0dd26773e1d198d
#> 2   fec0ec46ebd2534db0dd26773e1d198d
#> 3   fec0ec46ebd2534db0dd26773e1d198d
#> 121 fec0ec46ebd2534db0dd26773e1d198d
#> 5   fec0ec46ebd2534db0dd26773e1d198d
#> 6   fec0ec46ebd2534db0dd26773e1d198d
#> 7   fec0ec46ebd2534db0dd26773e1d198d
#> 8   fec0ec46ebd2534db0dd26773e1d198d
#> 9   fec0ec46ebd2534db0dd26773e1d198d
#> 10  fec0ec46ebd2534db0dd26773e1d198d
#> 11  fec0ec46ebd2534db0dd26773e1d198d
#> 12  fec0ec46ebd2534db0dd26773e1d198d
#> 13  fec0ec46ebd2534db0dd26773e1d198d
#> 14  fec0ec46ebd2534db0dd26773e1d198d
#> 15  fec0ec46ebd2534db0dd26773e1d198d
#> 16  fec0ec46ebd2534db0dd26773e1d198d
#> 17  fec0ec46ebd2534db0dd26773e1d198d
#> 18  fec0ec46ebd2534db0dd26773e1d198d
#> 19  fec0ec46ebd2534db0dd26773e1d198d
#> 20  fec0ec46ebd2534db0dd26773e1d198d
#> 21  fec0ec46ebd2534db0dd26773e1d198d
#> 22  fec0ec46ebd2534db0dd26773e1d198d
#> 23  fec0ec46ebd2534db0dd26773e1d198d
#> 24  fec0ec46ebd2534db0dd26773e1d198d
#> 25  fec0ec46ebd2534db0dd26773e1d198d
#> 26  fec0ec46ebd2534db0dd26773e1d198d
#> 27  fec0ec46ebd2534db0dd26773e1d198d
#> 28  fec0ec46ebd2534db0dd26773e1d198d
#> 29  fec0ec46ebd2534db0dd26773e1d198d
#> 30  fec0ec46ebd2534db0dd26773e1d198d
#> 31  fec0ec46ebd2534db0dd26773e1d198d
#> 32  fec0ec46ebd2534db0dd26773e1d198d
#> 33  fec0ec46ebd2534db0dd26773e1d198d
#> 34  fec0ec46ebd2534db0dd26773e1d198d
#> 35  fec0ec46ebd2534db0dd26773e1d198d
#> 36  fec0ec46ebd2534db0dd26773e1d198d
#> 37  fec0ec46ebd2534db0dd26773e1d198d
#> 38  fec0ec46ebd2534db0dd26773e1d198d
#> 39  fec0ec46ebd2534db0dd26773e1d198d
#> 40  fec0ec46ebd2534db0dd26773e1d198d
#> 41  fec0ec46ebd2534db0dd26773e1d198d
#> 42  fec0ec46ebd2534db0dd26773e1d198d
#> 43  fec0ec46ebd2534db0dd26773e1d198d
#> 44  fec0ec46ebd2534db0dd26773e1d198d
#> 45  fec0ec46ebd2534db0dd26773e1d198d
#> 46  fec0ec46ebd2534db0dd26773e1d198d
#> 47  fec0ec46ebd2534db0dd26773e1d198d
#> 48  fec0ec46ebd2534db0dd26773e1d198d
#> 49  fec0ec46ebd2534db0dd26773e1d198d
#> 50  fec0ec46ebd2534db0dd26773e1d198d
#> 51  fec0ec46ebd2534db0dd26773e1d198d
#> 52  fec0ec46ebd2534db0dd26773e1d198d
#> 53  fec0ec46ebd2534db0dd26773e1d198d
#> 54  fec0ec46ebd2534db0dd26773e1d198d
#> 55  fec0ec46ebd2534db0dd26773e1d198d
#> 56  fec0ec46ebd2534db0dd26773e1d198d
#> 57  fec0ec46ebd2534db0dd26773e1d198d
#> 58  fec0ec46ebd2534db0dd26773e1d198d
#> 59  fec0ec46ebd2534db0dd26773e1d198d
#> 60  fec0ec46ebd2534db0dd26773e1d198d
#> 61  fec0ec46ebd2534db0dd26773e1d198d
#> 62  fec0ec46ebd2534db0dd26773e1d198d
#> 63  fec0ec46ebd2534db0dd26773e1d198d
#> 64  fec0ec46ebd2534db0dd26773e1d198d
#> 65  fec0ec46ebd2534db0dd26773e1d198d
#> 66  fec0ec46ebd2534db0dd26773e1d198d
#> 67  fec0ec46ebd2534db0dd26773e1d198d
#> 68  fec0ec46ebd2534db0dd26773e1d198d
#> 69  fec0ec46ebd2534db0dd26773e1d198d
#> 70  fec0ec46ebd2534db0dd26773e1d198d
#> 71  fec0ec46ebd2534db0dd26773e1d198d
#> 72  fec0ec46ebd2534db0dd26773e1d198d
#> 73  fec0ec46ebd2534db0dd26773e1d198d
#> 74  fec0ec46ebd2534db0dd26773e1d198d
#> 75  fec0ec46ebd2534db0dd26773e1d198d
#> 76  fec0ec46ebd2534db0dd26773e1d198d
#> 77  fec0ec46ebd2534db0dd26773e1d198d
#> 78  fec0ec46ebd2534db0dd26773e1d198d
#> 79  fec0ec46ebd2534db0dd26773e1d198d
#> 80  fec0ec46ebd2534db0dd26773e1d198d
#> 81  fec0ec46ebd2534db0dd26773e1d198d
#> 82  fec0ec46ebd2534db0dd26773e1d198d
#> 83  fec0ec46ebd2534db0dd26773e1d198d
#> 84  fec0ec46ebd2534db0dd26773e1d198d
#> 85  fec0ec46ebd2534db0dd26773e1d198d
#> 86  fec0ec46ebd2534db0dd26773e1d198d
#> 87  fec0ec46ebd2534db0dd26773e1d198d
#> 88  fec0ec46ebd2534db0dd26773e1d198d
#> 89  fec0ec46ebd2534db0dd26773e1d198d
#> 90  fec0ec46ebd2534db0dd26773e1d198d
#> 91  fec0ec46ebd2534db0dd26773e1d198d
#> 92  fec0ec46ebd2534db0dd26773e1d198d
#> 93  fec0ec46ebd2534db0dd26773e1d198d
#> 94  fec0ec46ebd2534db0dd26773e1d198d
#> 95  fec0ec46ebd2534db0dd26773e1d198d
#> 96  fec0ec46ebd2534db0dd26773e1d198d
#> 97  fec0ec46ebd2534db0dd26773e1d198d
#> 98  fec0ec46ebd2534db0dd26773e1d198d
#> 99  fec0ec46ebd2534db0dd26773e1d198d
#> 100 fec0ec46ebd2534db0dd26773e1d198d
#> 101 fec0ec46ebd2534db0dd26773e1d198d
#> 102 fec0ec46ebd2534db0dd26773e1d198d
#> 103 fec0ec46ebd2534db0dd26773e1d198d
#> 104 fec0ec46ebd2534db0dd26773e1d198d
#> 105 fec0ec46ebd2534db0dd26773e1d198d
#> 106 fec0ec46ebd2534db0dd26773e1d198d
#> 107 fec0ec46ebd2534db0dd26773e1d198d
#> 108 fec0ec46ebd2534db0dd26773e1d198d
#> 109 fec0ec46ebd2534db0dd26773e1d198d
#> 110 fec0ec46ebd2534db0dd26773e1d198d
#> 111 fec0ec46ebd2534db0dd26773e1d198d
#> 112 fec0ec46ebd2534db0dd26773e1d198d
#> 113 fec0ec46ebd2534db0dd26773e1d198d
#> 114 fec0ec46ebd2534db0dd26773e1d198d
#> 115 fec0ec46ebd2534db0dd26773e1d198d
#> 116 fec0ec46ebd2534db0dd26773e1d198d
#> 117 fec0ec46ebd2534db0dd26773e1d198d
#> 118 fec0ec46ebd2534db0dd26773e1d198d
#> 122 fec0ec46ebd2534db0dd26773e1d198d
#> 119 fec0ec46ebd2534db0dd26773e1d198d
#> 120 fec0ec46ebd2534db0dd26773e1d198d
#> 123 fec0ec46ebd2534db0dd26773e1d198d
#> 

## Stop parallel workers and disable journal logging and signaling
plan(sequential)
options(oopts)