Gets the logged journal of events for a future

WARNING: This function is under development. It can change at any time. For now, please, do not depend on this function in a published R package.

Usage

journal(x, ...)

Arguments

x

A future::Future object.

...

Not used.

Value

A data frame of class FutureJournal with columns:

event (character string) - type of event that took place

category (character string) - the category of the event

parent (character string) - (to be describe)

start (POSIXct) - the timestamp when the event started

at (difftime) - the time when the event started relative to first event

duration (difftime) - the duration of the event

future_label (character string) - the label of the future

future_uuid (character string) - the UUID of the future

session_uuid (character string) - the UUID of the R session where the event took place

The common events are:

create - the future was created (an overhead)

launch - the future was launched (an overhead)

evaluate - the future was evaluated (an evaluation)

resolved - the future was queried (may be occur multiple times) (an overhead)

gather - the results was retrieved (an overhead)

but others may be added by other Future classes.

Common event categories are:

evaluation - processing time is spent on evaluation

overhead - processing time is spent on orchestrating the future

waiting - processing time is spent on waiting to set up or querying the future

but others may be added by other Future classes.

The data frame is sorted by the at time. Note that the timestamps for the evaluate event are based on the local time on the worker. The system clocks on the worker and the calling R system may not be in perfect sync.

Enabling and disabling event logging

To enable logging of events, set option future.journal is TRUE. To disable, set it to FALSE (default).

See also

Examples

## Enable journaling of futures
oopts <- options(future.journal = TRUE)

plan(multisession, workers = 2L)

t_start <- Sys.time()
fs <- lapply(1:3, FUN = function(x) future({ Sys.sleep(x); sqrt(x) }))
vs <- value(fs)
js <- lapply(fs, FUN = journal, baseline = t_start)
print(js)
#> [[1]]
#>            event   category parent               start               at
#> 1         create   overhead   <NA> 2026-01-25 16:33:35 0.000000000 secs
#> 4         launch   overhead   <NA> 2026-01-25 16:33:35 0.007212162 secs
#> 2      getWorker   overhead launch 2026-01-25 16:33:35 0.007328987 secs
#> 3    eraseWorker   overhead launch 2026-01-25 16:33:35 0.008412361 secs
#> 5       resolved       <NA>   <NA> 2026-01-25 16:33:35 0.042851925 secs
#> 6       resolved       <NA>   <NA> 2026-01-25 16:33:35 0.076385021 secs
#> 7       resolved       <NA>   <NA> 2026-01-25 16:33:35 0.109673023 secs
#> 8       resolved       <NA>   <NA> 2026-01-25 16:33:35 0.143195391 secs
#> 9       resolved       <NA>   <NA> 2026-01-25 16:33:35 0.176886559 secs
#> 41      evaluate evaluation   <NA> 2026-01-25 16:33:35 0.187297106 secs
#> 10      resolved       <NA>   <NA> 2026-01-25 16:33:35 0.211236715 secs
#> 11      resolved       <NA>   <NA> 2026-01-25 16:33:35 0.246399403 secs
#> 12      resolved       <NA>   <NA> 2026-01-25 16:33:35 0.280499458 secs
#> 13      resolved       <NA>   <NA> 2026-01-25 16:33:35 0.319164753 secs
#> 14      resolved       <NA>   <NA> 2026-01-25 16:33:35 0.353872776 secs
#> 15      resolved       <NA>   <NA> 2026-01-25 16:33:35 0.391916513 secs
#> 16      resolved       <NA>   <NA> 2026-01-25 16:33:36 0.429578781 secs
#> 17      resolved       <NA>   <NA> 2026-01-25 16:33:36 0.467386961 secs
#> 18      resolved       <NA>   <NA> 2026-01-25 16:33:36 0.505439758 secs
#> 19      resolved       <NA>   <NA> 2026-01-25 16:33:36 0.543430328 secs
#> 20      resolved       <NA>   <NA> 2026-01-25 16:33:36 0.581537724 secs
#> 21      resolved       <NA>   <NA> 2026-01-25 16:33:36 0.620006800 secs
#> 22      resolved       <NA>   <NA> 2026-01-25 16:33:36 0.659044027 secs
#> 23      resolved       <NA>   <NA> 2026-01-25 16:33:36 0.697744846 secs
#> 24      resolved       <NA>   <NA> 2026-01-25 16:33:36 0.736318827 secs
#> 25      resolved       <NA>   <NA> 2026-01-25 16:33:36 0.775137901 secs
#> 26      resolved       <NA>   <NA> 2026-01-25 16:33:36 0.813734055 secs
#> 27      resolved       <NA>   <NA> 2026-01-25 16:33:36 0.852498770 secs
#> 28      resolved       <NA>   <NA> 2026-01-25 16:33:36 0.891539574 secs
#> 29      resolved       <NA>   <NA> 2026-01-25 16:33:36 0.930731058 secs
#> 30      resolved       <NA>   <NA> 2026-01-25 16:33:36 0.970166683 secs
#> 31      resolved       <NA>   <NA> 2026-01-25 16:33:36 1.009792328 secs
#> 32      resolved       <NA>   <NA> 2026-01-25 16:33:36 1.049490929 secs
#> 33      resolved       <NA>   <NA> 2026-01-25 16:33:36 1.089111328 secs
#> 34      resolved       <NA>   <NA> 2026-01-25 16:33:36 1.129076242 secs
#> 35      resolved       <NA>   <NA> 2026-01-25 16:33:36 1.169508219 secs
#> 36      resolved       <NA>   <NA> 2026-01-25 16:33:36 1.209581852 secs
#> 37      resolved       <NA>   <NA> 2026-01-25 16:33:36 1.249819994 secs
#> 38      resolved       <NA>   <NA> 2026-01-25 16:33:36 1.290004969 secs
#> 42      resolved       <NA>   <NA> 2026-01-25 16:33:36 1.330100060 secs
#> 39 receiveResult   overhead gather 2026-01-25 16:33:36 1.331326962 secs
#> 40        gather   overhead   <NA> 2026-01-25 16:33:36 1.333472252 secs
#> 43      resolved       <NA>   <NA> 2026-01-25 16:33:36 1.350581408 secs
#> 44       resolve   overhead   <NA> 2026-01-25 16:33:36 1.351481915 secs
#>             duration future_label                         future_uuid
#> 1  6.910801e-03 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-17
#> 4  4.907846e-03 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-17
#> 2  5.192757e-04 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-17
#> 3  3.986359e-04 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-17
#> 5  1.091003e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-17
#> 6  1.069665e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-17
#> 7  1.078415e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-17
#> 8  1.077700e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-17
#> 9  1.081920e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-17
#> 41 1.124770e+00 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-17
#> 10 1.076102e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-17
#> 11 1.080489e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-17
#> 12 1.079369e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-17
#> 13 1.078320e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-17
#> 14 1.148629e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-17
#> 15 1.154780e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-17
#> 16 1.150751e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-17
#> 17 1.146936e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-17
#> 18 1.149821e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-17
#> 19 1.154423e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-17
#> 20 1.150370e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-17
#> 21 1.187968e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-17
#> 22 1.150012e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-17
#> 23 1.155710e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-17
#> 24 1.148820e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-17
#> 25 1.141286e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-17
#> 26 1.149130e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-17
#> 27 1.143289e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-17
#> 28 1.138616e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-17
#> 29 1.156569e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-17
#> 30 1.163697e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-17
#> 31 1.148200e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-17
#> 32 1.149893e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-17
#> 33 1.145339e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-17
#> 34 1.165009e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-17
#> 35 1.150131e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-17
#> 36 1.148129e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-17
#> 37 1.143694e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-17
#> 38 1.152611e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-17
#> 42 8.465767e-03 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-17
#> 39 2.605915e-04 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-17
#> 40 1.401901e-04 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-17
#> 43 4.315376e-05 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-17
#> 44 1.192093e-05 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-17
#>                        session_uuid
#> 1  6128f31e79476bcb789b83bdb3b16a22
#> 4  6128f31e79476bcb789b83bdb3b16a22
#> 2  6128f31e79476bcb789b83bdb3b16a22
#> 3  6128f31e79476bcb789b83bdb3b16a22
#> 5  6128f31e79476bcb789b83bdb3b16a22
#> 6  6128f31e79476bcb789b83bdb3b16a22
#> 7  6128f31e79476bcb789b83bdb3b16a22
#> 8  6128f31e79476bcb789b83bdb3b16a22
#> 9  6128f31e79476bcb789b83bdb3b16a22
#> 41 6128f31e79476bcb789b83bdb3b16a22
#> 10 6128f31e79476bcb789b83bdb3b16a22
#> 11 6128f31e79476bcb789b83bdb3b16a22
#> 12 6128f31e79476bcb789b83bdb3b16a22
#> 13 6128f31e79476bcb789b83bdb3b16a22
#> 14 6128f31e79476bcb789b83bdb3b16a22
#> 15 6128f31e79476bcb789b83bdb3b16a22
#> 16 6128f31e79476bcb789b83bdb3b16a22
#> 17 6128f31e79476bcb789b83bdb3b16a22
#> 18 6128f31e79476bcb789b83bdb3b16a22
#> 19 6128f31e79476bcb789b83bdb3b16a22
#> 20 6128f31e79476bcb789b83bdb3b16a22
#> 21 6128f31e79476bcb789b83bdb3b16a22
#> 22 6128f31e79476bcb789b83bdb3b16a22
#> 23 6128f31e79476bcb789b83bdb3b16a22
#> 24 6128f31e79476bcb789b83bdb3b16a22
#> 25 6128f31e79476bcb789b83bdb3b16a22
#> 26 6128f31e79476bcb789b83bdb3b16a22
#> 27 6128f31e79476bcb789b83bdb3b16a22
#> 28 6128f31e79476bcb789b83bdb3b16a22
#> 29 6128f31e79476bcb789b83bdb3b16a22
#> 30 6128f31e79476bcb789b83bdb3b16a22
#> 31 6128f31e79476bcb789b83bdb3b16a22
#> 32 6128f31e79476bcb789b83bdb3b16a22
#> 33 6128f31e79476bcb789b83bdb3b16a22
#> 34 6128f31e79476bcb789b83bdb3b16a22
#> 35 6128f31e79476bcb789b83bdb3b16a22
#> 36 6128f31e79476bcb789b83bdb3b16a22
#> 37 6128f31e79476bcb789b83bdb3b16a22
#> 38 6128f31e79476bcb789b83bdb3b16a22
#> 42 6128f31e79476bcb789b83bdb3b16a22
#> 39 6128f31e79476bcb789b83bdb3b16a22
#> 40 6128f31e79476bcb789b83bdb3b16a22
#> 43 6128f31e79476bcb789b83bdb3b16a22
#> 44 6128f31e79476bcb789b83bdb3b16a22
#> 
#> [[2]]
#>            event   category parent               start               at
#> 1         create   overhead   <NA> 2026-01-25 16:33:35 0.000000000 secs
#> 4         launch   overhead   <NA> 2026-01-25 16:33:35 0.007308006 secs
#> 2      getWorker   overhead launch 2026-01-25 16:33:35 0.007428646 secs
#> 3    eraseWorker   overhead launch 2026-01-25 16:33:35 0.008457899 secs
#> 5       resolved       <NA>   <NA> 2026-01-25 16:33:35 0.041819096 secs
#> 6       resolved       <NA>   <NA> 2026-01-25 16:33:35 0.075146437 secs
#> 7       resolved       <NA>   <NA> 2026-01-25 16:33:35 0.108551741 secs
#> 8       resolved       <NA>   <NA> 2026-01-25 16:33:35 0.142183065 secs
#> 9       resolved       <NA>   <NA> 2026-01-25 16:33:35 0.175957441 secs
#> 68      evaluate evaluation   <NA> 2026-01-25 16:33:35 0.188623905 secs
#> 10      resolved       <NA>   <NA> 2026-01-25 16:33:35 0.210693836 secs
#> 11      resolved       <NA>   <NA> 2026-01-25 16:33:35 0.245344877 secs
#> 12      resolved       <NA>   <NA> 2026-01-25 16:33:35 0.279444933 secs
#> 13      resolved       <NA>   <NA> 2026-01-25 16:33:35 0.318154573 secs
#> 14      resolved       <NA>   <NA> 2026-01-25 16:33:35 0.354622602 secs
#> 15      resolved       <NA>   <NA> 2026-01-25 16:33:35 0.392350674 secs
#> 16      resolved       <NA>   <NA> 2026-01-25 16:33:36 0.429974079 secs
#> 17      resolved       <NA>   <NA> 2026-01-25 16:33:36 0.467769146 secs
#> 18      resolved       <NA>   <NA> 2026-01-25 16:33:36 0.505842209 secs
#> 19      resolved       <NA>   <NA> 2026-01-25 16:33:36 0.543853998 secs
#> 20      resolved       <NA>   <NA> 2026-01-25 16:33:36 0.581930161 secs
#> 21      resolved       <NA>   <NA> 2026-01-25 16:33:36 0.620913506 secs
#> 22      resolved       <NA>   <NA> 2026-01-25 16:33:36 0.659479856 secs
#> 23      resolved       <NA>   <NA> 2026-01-25 16:33:36 0.698144197 secs
#> 24      resolved       <NA>   <NA> 2026-01-25 16:33:36 0.736701250 secs
#> 25      resolved       <NA>   <NA> 2026-01-25 16:33:36 0.775371552 secs
#> 26      resolved       <NA>   <NA> 2026-01-25 16:33:36 0.814052582 secs
#> 27      resolved       <NA>   <NA> 2026-01-25 16:33:36 0.852897167 secs
#> 28      resolved       <NA>   <NA> 2026-01-25 16:33:36 0.891817570 secs
#> 29      resolved       <NA>   <NA> 2026-01-25 16:33:36 0.931170464 secs
#> 30      resolved       <NA>   <NA> 2026-01-25 16:33:36 0.970746517 secs
#> 31      resolved       <NA>   <NA> 2026-01-25 16:33:36 1.010342598 secs
#> 32      resolved       <NA>   <NA> 2026-01-25 16:33:36 1.049866915 secs
#> 33      resolved       <NA>   <NA> 2026-01-25 16:33:36 1.089494705 secs
#> 34      resolved       <NA>   <NA> 2026-01-25 16:33:36 1.129735708 secs
#> 35      resolved       <NA>   <NA> 2026-01-25 16:33:36 1.169920921 secs
#> 36      resolved       <NA>   <NA> 2026-01-25 16:33:36 1.210059643 secs
#> 37      resolved       <NA>   <NA> 2026-01-25 16:33:36 1.250178099 secs
#> 38      resolved       <NA>   <NA> 2026-01-25 16:33:36 1.290344715 secs
#> 39      resolved       <NA>   <NA> 2026-01-25 16:33:36 1.339327097 secs
#> 40      resolved       <NA>   <NA> 2026-01-25 16:33:36 1.374291420 secs
#> 41      resolved       <NA>   <NA> 2026-01-25 16:33:37 1.410438538 secs
#> 42      resolved       <NA>   <NA> 2026-01-25 16:33:37 1.446633339 secs
#> 43      resolved       <NA>   <NA> 2026-01-25 16:33:37 1.482783794 secs
#> 44      resolved       <NA>   <NA> 2026-01-25 16:33:37 1.518812656 secs
#> 45      resolved       <NA>   <NA> 2026-01-25 16:33:37 1.555108547 secs
#> 46      resolved       <NA>   <NA> 2026-01-25 16:33:37 1.591328144 secs
#> 47      resolved       <NA>   <NA> 2026-01-25 16:33:37 1.628352404 secs
#> 48      resolved       <NA>   <NA> 2026-01-25 16:33:37 1.664637089 secs
#> 49      resolved       <NA>   <NA> 2026-01-25 16:33:37 1.700621367 secs
#> 50      resolved       <NA>   <NA> 2026-01-25 16:33:37 1.736844540 secs
#> 51      resolved       <NA>   <NA> 2026-01-25 16:33:37 1.773596764 secs
#> 52      resolved       <NA>   <NA> 2026-01-25 16:33:37 1.810637236 secs
#> 53      resolved       <NA>   <NA> 2026-01-25 16:33:37 1.847327948 secs
#> 54      resolved       <NA>   <NA> 2026-01-25 16:33:37 1.884053469 secs
#> 55      resolved       <NA>   <NA> 2026-01-25 16:33:37 1.920361996 secs
#> 56      resolved       <NA>   <NA> 2026-01-25 16:33:37 1.956422567 secs
#> 57      resolved       <NA>   <NA> 2026-01-25 16:33:37 1.992339849 secs
#> 58      resolved       <NA>   <NA> 2026-01-25 16:33:37 2.029011726 secs
#> 59      resolved       <NA>   <NA> 2026-01-25 16:33:37 2.065688372 secs
#> 60      resolved       <NA>   <NA> 2026-01-25 16:33:37 2.101906061 secs
#> 61      resolved       <NA>   <NA> 2026-01-25 16:33:37 2.138700247 secs
#> 62      resolved       <NA>   <NA> 2026-01-25 16:33:37 2.174826145 secs
#> 63      resolved       <NA>   <NA> 2026-01-25 16:33:37 2.211444139 secs
#> 64      resolved       <NA>   <NA> 2026-01-25 16:33:37 2.247658968 secs
#> 65      resolved       <NA>   <NA> 2026-01-25 16:33:37 2.281295300 secs
#> 69      resolved       <NA>   <NA> 2026-01-25 16:33:37 2.317137957 secs
#> 66 receiveResult   overhead gather 2026-01-25 16:33:37 2.318475008 secs
#> 67        gather   overhead   <NA> 2026-01-25 16:33:37 2.320580959 secs
#> 70       resolve   overhead   <NA> 2026-01-25 16:33:37 2.327844858 secs
#>             duration future_label                         future_uuid
#> 1  7.016420e-03 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 4  4.707336e-03 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 2  4.904270e-04 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 3  2.722740e-04 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 5  1.080847e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 6  1.063275e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 7  1.070046e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 8  1.063228e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 9  1.100254e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 68 2.124472e+00 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 10 1.117110e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 11 1.071477e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 12 1.513743e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 13 1.078391e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 14 1.143479e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 15 1.134849e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 16 1.142287e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 17 1.148653e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 18 1.129818e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 19 1.133657e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 20 1.139617e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 21 1.137805e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 22 1.128554e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 23 1.136589e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 24 1.135445e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 25 1.134253e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 26 1.135850e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 27 1.136947e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 28 1.134944e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 29 1.142216e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 30 1.135516e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 31 1.141000e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 32 1.136351e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 33 1.144695e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 34 1.135445e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 35 1.137471e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 36 1.130939e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 37 1.132345e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 38 1.127219e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 39 1.066017e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 40 1.156187e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 41 1.145554e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 42 1.151156e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 43 1.140547e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 44 1.161408e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 45 1.156187e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 46 1.150799e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 47 1.147366e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 48 1.154470e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 49 1.145625e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 50 1.182032e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 51 1.181269e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 52 1.144719e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 53 1.157618e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 54 1.157045e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 55 1.148915e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 56 1.153255e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 57 1.152015e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 58 1.160097e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 59 1.152658e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 60 1.182508e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 61 1.155162e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 62 1.154613e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 63 1.145935e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 64 1.122022e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 65 1.148987e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 69 8.338213e-03 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 66 3.001690e-04 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 67 1.294613e-04 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#> 70 3.170967e-05 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-18
#>                        session_uuid
#> 1  6128f31e79476bcb789b83bdb3b16a22
#> 4  6128f31e79476bcb789b83bdb3b16a22
#> 2  6128f31e79476bcb789b83bdb3b16a22
#> 3  6128f31e79476bcb789b83bdb3b16a22
#> 5  6128f31e79476bcb789b83bdb3b16a22
#> 6  6128f31e79476bcb789b83bdb3b16a22
#> 7  6128f31e79476bcb789b83bdb3b16a22
#> 8  6128f31e79476bcb789b83bdb3b16a22
#> 9  6128f31e79476bcb789b83bdb3b16a22
#> 68 6128f31e79476bcb789b83bdb3b16a22
#> 10 6128f31e79476bcb789b83bdb3b16a22
#> 11 6128f31e79476bcb789b83bdb3b16a22
#> 12 6128f31e79476bcb789b83bdb3b16a22
#> 13 6128f31e79476bcb789b83bdb3b16a22
#> 14 6128f31e79476bcb789b83bdb3b16a22
#> 15 6128f31e79476bcb789b83bdb3b16a22
#> 16 6128f31e79476bcb789b83bdb3b16a22
#> 17 6128f31e79476bcb789b83bdb3b16a22
#> 18 6128f31e79476bcb789b83bdb3b16a22
#> 19 6128f31e79476bcb789b83bdb3b16a22
#> 20 6128f31e79476bcb789b83bdb3b16a22
#> 21 6128f31e79476bcb789b83bdb3b16a22
#> 22 6128f31e79476bcb789b83bdb3b16a22
#> 23 6128f31e79476bcb789b83bdb3b16a22
#> 24 6128f31e79476bcb789b83bdb3b16a22
#> 25 6128f31e79476bcb789b83bdb3b16a22
#> 26 6128f31e79476bcb789b83bdb3b16a22
#> 27 6128f31e79476bcb789b83bdb3b16a22
#> 28 6128f31e79476bcb789b83bdb3b16a22
#> 29 6128f31e79476bcb789b83bdb3b16a22
#> 30 6128f31e79476bcb789b83bdb3b16a22
#> 31 6128f31e79476bcb789b83bdb3b16a22
#> 32 6128f31e79476bcb789b83bdb3b16a22
#> 33 6128f31e79476bcb789b83bdb3b16a22
#> 34 6128f31e79476bcb789b83bdb3b16a22
#> 35 6128f31e79476bcb789b83bdb3b16a22
#> 36 6128f31e79476bcb789b83bdb3b16a22
#> 37 6128f31e79476bcb789b83bdb3b16a22
#> 38 6128f31e79476bcb789b83bdb3b16a22
#> 39 6128f31e79476bcb789b83bdb3b16a22
#> 40 6128f31e79476bcb789b83bdb3b16a22
#> 41 6128f31e79476bcb789b83bdb3b16a22
#> 42 6128f31e79476bcb789b83bdb3b16a22
#> 43 6128f31e79476bcb789b83bdb3b16a22
#> 44 6128f31e79476bcb789b83bdb3b16a22
#> 45 6128f31e79476bcb789b83bdb3b16a22
#> 46 6128f31e79476bcb789b83bdb3b16a22
#> 47 6128f31e79476bcb789b83bdb3b16a22
#> 48 6128f31e79476bcb789b83bdb3b16a22
#> 49 6128f31e79476bcb789b83bdb3b16a22
#> 50 6128f31e79476bcb789b83bdb3b16a22
#> 51 6128f31e79476bcb789b83bdb3b16a22
#> 52 6128f31e79476bcb789b83bdb3b16a22
#> 53 6128f31e79476bcb789b83bdb3b16a22
#> 54 6128f31e79476bcb789b83bdb3b16a22
#> 55 6128f31e79476bcb789b83bdb3b16a22
#> 56 6128f31e79476bcb789b83bdb3b16a22
#> 57 6128f31e79476bcb789b83bdb3b16a22
#> 58 6128f31e79476bcb789b83bdb3b16a22
#> 59 6128f31e79476bcb789b83bdb3b16a22
#> 60 6128f31e79476bcb789b83bdb3b16a22
#> 61 6128f31e79476bcb789b83bdb3b16a22
#> 62 6128f31e79476bcb789b83bdb3b16a22
#> 63 6128f31e79476bcb789b83bdb3b16a22
#> 64 6128f31e79476bcb789b83bdb3b16a22
#> 65 6128f31e79476bcb789b83bdb3b16a22
#> 69 6128f31e79476bcb789b83bdb3b16a22
#> 66 6128f31e79476bcb789b83bdb3b16a22
#> 67 6128f31e79476bcb789b83bdb3b16a22
#> 70 6128f31e79476bcb789b83bdb3b16a22
#> 
#> [[3]]
#>             event   category parent               start               at
#> 1          create   overhead   <NA> 2026-01-25 16:33:35 0.000000000 secs
#> 4          launch   overhead   <NA> 2026-01-25 16:33:35 0.007320881 secs
#> 2       getWorker   overhead launch 2026-01-25 16:33:35 0.007440329 secs
#> 3     eraseWorker   overhead launch 2026-01-25 16:33:36 1.317763567 secs
#> 5        resolved       <NA>   <NA> 2026-01-25 16:33:36 1.338515759 secs
#> 122      evaluate evaluation   <NA> 2026-01-25 16:33:36 1.365253925 secs
#> 6        resolved       <NA>   <NA> 2026-01-25 16:33:36 1.374849796 secs
#> 7        resolved       <NA>   <NA> 2026-01-25 16:33:37 1.410955429 secs
#> 8        resolved       <NA>   <NA> 2026-01-25 16:33:37 1.447143316 secs
#> 9        resolved       <NA>   <NA> 2026-01-25 16:33:37 1.483145237 secs
#> 10       resolved       <NA>   <NA> 2026-01-25 16:33:37 1.519413233 secs
#> 11       resolved       <NA>   <NA> 2026-01-25 16:33:37 1.555641890 secs
#> 12       resolved       <NA>   <NA> 2026-01-25 16:33:37 1.592360020 secs
#> 13       resolved       <NA>   <NA> 2026-01-25 16:33:37 1.628858089 secs
#> 14       resolved       <NA>   <NA> 2026-01-25 16:33:37 1.665126324 secs
#> 15       resolved       <NA>   <NA> 2026-01-25 16:33:37 1.701095581 secs
#> 16       resolved       <NA>   <NA> 2026-01-25 16:33:37 1.737651825 secs
#> 17       resolved       <NA>   <NA> 2026-01-25 16:33:37 1.774615526 secs
#> 18       resolved       <NA>   <NA> 2026-01-25 16:33:37 1.811347246 secs
#> 19       resolved       <NA>   <NA> 2026-01-25 16:33:37 1.848105907 secs
#> 20       resolved       <NA>   <NA> 2026-01-25 16:33:37 1.884543657 secs
#> 21       resolved       <NA>   <NA> 2026-01-25 16:33:37 1.920820951 secs
#> 22       resolved       <NA>   <NA> 2026-01-25 16:33:37 1.956900835 secs
#> 23       resolved       <NA>   <NA> 2026-01-25 16:33:37 1.993028641 secs
#> 24       resolved       <NA>   <NA> 2026-01-25 16:33:37 2.029608965 secs
#> 25       resolved       <NA>   <NA> 2026-01-25 16:33:37 2.066231251 secs
#> 26       resolved       <NA>   <NA> 2026-01-25 16:33:37 2.102916479 secs
#> 27       resolved       <NA>   <NA> 2026-01-25 16:33:37 2.139271498 secs
#> 28       resolved       <NA>   <NA> 2026-01-25 16:33:37 2.175607204 secs
#> 29       resolved       <NA>   <NA> 2026-01-25 16:33:37 2.211904764 secs
#> 30       resolved       <NA>   <NA> 2026-01-25 16:33:37 2.246837378 secs
#> 31       resolved       <NA>   <NA> 2026-01-25 16:33:37 2.281729698 secs
#> 32       resolved       <NA>   <NA> 2026-01-25 16:33:37 2.316659927 secs
#> 33       resolved       <NA>   <NA> 2026-01-25 16:33:37 2.339271784 secs
#> 34       resolved       <NA>   <NA> 2026-01-25 16:33:37 2.362524271 secs
#> 35       resolved       <NA>   <NA> 2026-01-25 16:33:38 2.386105061 secs
#> 36       resolved       <NA>   <NA> 2026-01-25 16:33:38 2.416207314 secs
#> 37       resolved       <NA>   <NA> 2026-01-25 16:33:38 2.440034628 secs
#> 38       resolved       <NA>   <NA> 2026-01-25 16:33:38 2.461459398 secs
#> 39       resolved       <NA>   <NA> 2026-01-25 16:33:38 2.484273434 secs
#> 40       resolved       <NA>   <NA> 2026-01-25 16:33:38 2.507467270 secs
#> 41       resolved       <NA>   <NA> 2026-01-25 16:33:38 2.530823231 secs
#> 42       resolved       <NA>   <NA> 2026-01-25 16:33:38 2.554276705 secs
#> 43       resolved       <NA>   <NA> 2026-01-25 16:33:38 2.577857971 secs
#> 44       resolved       <NA>   <NA> 2026-01-25 16:33:38 2.601193905 secs
#> 45       resolved       <NA>   <NA> 2026-01-25 16:33:38 2.624724388 secs
#> 46       resolved       <NA>   <NA> 2026-01-25 16:33:38 2.648548126 secs
#> 47       resolved       <NA>   <NA> 2026-01-25 16:33:38 2.672096729 secs
#> 48       resolved       <NA>   <NA> 2026-01-25 16:33:38 2.695498943 secs
#> 49       resolved       <NA>   <NA> 2026-01-25 16:33:38 2.719219446 secs
#> 50       resolved       <NA>   <NA> 2026-01-25 16:33:38 2.742659807 secs
#> 51       resolved       <NA>   <NA> 2026-01-25 16:33:38 2.766181469 secs
#> 52       resolved       <NA>   <NA> 2026-01-25 16:33:38 2.789592743 secs
#> 53       resolved       <NA>   <NA> 2026-01-25 16:33:38 2.813032150 secs
#> 54       resolved       <NA>   <NA> 2026-01-25 16:33:38 2.836514473 secs
#> 55       resolved       <NA>   <NA> 2026-01-25 16:33:38 2.859836102 secs
#> 56       resolved       <NA>   <NA> 2026-01-25 16:33:38 2.883285999 secs
#> 57       resolved       <NA>   <NA> 2026-01-25 16:33:38 2.906767607 secs
#> 58       resolved       <NA>   <NA> 2026-01-25 16:33:38 2.930446863 secs
#> 59       resolved       <NA>   <NA> 2026-01-25 16:33:38 2.953891039 secs
#> 60       resolved       <NA>   <NA> 2026-01-25 16:33:38 2.977423429 secs
#> 61       resolved       <NA>   <NA> 2026-01-25 16:33:38 3.001384974 secs
#> 62       resolved       <NA>   <NA> 2026-01-25 16:33:38 3.024548531 secs
#> 63       resolved       <NA>   <NA> 2026-01-25 16:33:38 3.047954559 secs
#> 64       resolved       <NA>   <NA> 2026-01-25 16:33:38 3.071238756 secs
#> 65       resolved       <NA>   <NA> 2026-01-25 16:33:38 3.094474792 secs
#> 66       resolved       <NA>   <NA> 2026-01-25 16:33:38 3.117755651 secs
#> 67       resolved       <NA>   <NA> 2026-01-25 16:33:38 3.141097069 secs
#> 68       resolved       <NA>   <NA> 2026-01-25 16:33:38 3.164389133 secs
#> 69       resolved       <NA>   <NA> 2026-01-25 16:33:38 3.187465906 secs
#> 70       resolved       <NA>   <NA> 2026-01-25 16:33:38 3.210608482 secs
#> 71       resolved       <NA>   <NA> 2026-01-25 16:33:38 3.233842850 secs
#> 72       resolved       <NA>   <NA> 2026-01-25 16:33:38 3.257544994 secs
#> 73       resolved       <NA>   <NA> 2026-01-25 16:33:38 3.281060696 secs
#> 74       resolved       <NA>   <NA> 2026-01-25 16:33:38 3.304469347 secs
#> 75       resolved       <NA>   <NA> 2026-01-25 16:33:38 3.327663660 secs
#> 76       resolved       <NA>   <NA> 2026-01-25 16:33:38 3.350910187 secs
#> 77       resolved       <NA>   <NA> 2026-01-25 16:33:38 3.374244690 secs
#> 78       resolved       <NA>   <NA> 2026-01-25 16:33:39 3.397573948 secs
#> 79       resolved       <NA>   <NA> 2026-01-25 16:33:39 3.421113491 secs
#> 80       resolved       <NA>   <NA> 2026-01-25 16:33:39 3.444471598 secs
#> 81       resolved       <NA>   <NA> 2026-01-25 16:33:39 3.467830658 secs
#> 82       resolved       <NA>   <NA> 2026-01-25 16:33:39 3.491191864 secs
#> 83       resolved       <NA>   <NA> 2026-01-25 16:33:39 3.514596701 secs
#> 84       resolved       <NA>   <NA> 2026-01-25 16:33:39 3.537999868 secs
#> 85       resolved       <NA>   <NA> 2026-01-25 16:33:39 3.561046839 secs
#> 86       resolved       <NA>   <NA> 2026-01-25 16:33:39 3.584465265 secs
#> 87       resolved       <NA>   <NA> 2026-01-25 16:33:39 3.607805967 secs
#> 88       resolved       <NA>   <NA> 2026-01-25 16:33:39 3.631267309 secs
#> 89       resolved       <NA>   <NA> 2026-01-25 16:33:39 3.655032396 secs
#> 90       resolved       <NA>   <NA> 2026-01-25 16:33:39 3.678470612 secs
#> 91       resolved       <NA>   <NA> 2026-01-25 16:33:39 3.702100039 secs
#> 92       resolved       <NA>   <NA> 2026-01-25 16:33:39 3.725183964 secs
#> 93       resolved       <NA>   <NA> 2026-01-25 16:33:39 3.748593569 secs
#> 94       resolved       <NA>   <NA> 2026-01-25 16:33:39 3.771654367 secs
#> 95       resolved       <NA>   <NA> 2026-01-25 16:33:39 3.795060396 secs
#> 96       resolved       <NA>   <NA> 2026-01-25 16:33:39 3.818395138 secs
#> 97       resolved       <NA>   <NA> 2026-01-25 16:33:39 3.841557980 secs
#> 98       resolved       <NA>   <NA> 2026-01-25 16:33:39 3.864935160 secs
#> 99       resolved       <NA>   <NA> 2026-01-25 16:33:39 3.888220549 secs
#> 100      resolved       <NA>   <NA> 2026-01-25 16:33:39 3.911517859 secs
#> 101      resolved       <NA>   <NA> 2026-01-25 16:33:39 3.935136795 secs
#> 102      resolved       <NA>   <NA> 2026-01-25 16:33:39 3.958524942 secs
#> 103      resolved       <NA>   <NA> 2026-01-25 16:33:39 3.981934547 secs
#> 104      resolved       <NA>   <NA> 2026-01-25 16:33:39 4.005767107 secs
#> 105      resolved       <NA>   <NA> 2026-01-25 16:33:39 4.029177666 secs
#> 106      resolved       <NA>   <NA> 2026-01-25 16:33:39 4.053013802 secs
#> 107      resolved       <NA>   <NA> 2026-01-25 16:33:39 4.076233864 secs
#> 108      resolved       <NA>   <NA> 2026-01-25 16:33:39 4.099394798 secs
#> 109      resolved       <NA>   <NA> 2026-01-25 16:33:39 4.122993469 secs
#> 110      resolved       <NA>   <NA> 2026-01-25 16:33:39 4.146145344 secs
#> 111      resolved       <NA>   <NA> 2026-01-25 16:33:39 4.169326544 secs
#> 112      resolved       <NA>   <NA> 2026-01-25 16:33:39 4.192461014 secs
#> 113      resolved       <NA>   <NA> 2026-01-25 16:33:39 4.215644360 secs
#> 114      resolved       <NA>   <NA> 2026-01-25 16:33:39 4.238884211 secs
#> 115      resolved       <NA>   <NA> 2026-01-25 16:33:39 4.262605190 secs
#> 116      resolved       <NA>   <NA> 2026-01-25 16:33:39 4.285928965 secs
#> 117      resolved       <NA>   <NA> 2026-01-25 16:33:39 4.308999777 secs
#> 118      resolved       <NA>   <NA> 2026-01-25 16:33:39 4.332273960 secs
#> 119      resolved       <NA>   <NA> 2026-01-25 16:33:39 4.355391026 secs
#> 123      resolved       <NA>   <NA> 2026-01-25 16:33:39 4.378534794 secs
#> 120 receiveResult   overhead gather 2026-01-25 16:33:39 4.379876137 secs
#> 121        gather   overhead   <NA> 2026-01-25 16:33:39 4.381886005 secs
#> 124       resolve   overhead   <NA> 2026-01-25 16:33:40 4.388683319 secs
#>              duration future_label                         future_uuid
#> 1   7.028341e-03 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 4   1.316774e+00 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 2   1.309005e+00 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 3   7.355213e-04 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 5   1.145029e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 122 3.009427e+00 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 6   1.132274e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 7   1.129293e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 8   1.134753e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 9   1.139402e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 10  1.140547e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 11  1.142478e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 12  1.143527e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 13  1.124835e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 14  1.118898e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 15  1.140285e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 16  1.142979e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 17  1.168370e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 18  1.150846e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 19  1.152682e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 20  1.140308e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 21  1.123953e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 22  1.120281e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 23  1.159239e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 24  1.140618e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 25  1.137209e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 26  1.136065e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 27  1.130414e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 28  1.142001e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 29  1.140356e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 30  1.061177e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 31  1.117349e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 32  1.105070e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 33  1.146030e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 34  1.169252e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 35  1.133132e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 36  1.137280e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 37  1.071882e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 38  1.109147e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 39  1.144242e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 40  1.147294e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 41  1.152158e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 42  1.161814e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 43  1.152468e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 44  1.161146e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 45  1.172876e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 46  1.160741e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 47  1.150846e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 48  1.180887e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 49  1.156211e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 50  1.163149e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 51  1.157355e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 52  1.157188e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 53  1.163816e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 54  1.149821e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 55  1.158214e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 56  1.160455e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 57  1.168394e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 58  1.149368e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 59  1.161289e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 60  1.161313e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 61  1.142430e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 62  1.160383e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 63  1.143289e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 64  1.150274e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 65  1.145077e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 66  1.155734e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 67  1.145530e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 68  1.139832e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 69  1.146436e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 70  1.154208e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 71  1.174879e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 72  1.168013e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 73  1.143384e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 74  1.147604e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 75  1.141000e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 76  1.167464e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 77  1.150084e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 78  1.148057e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 79  1.151085e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 80  1.146960e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 81  1.159692e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 82  1.159906e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 83  1.156449e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 84  1.138353e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 85  1.146579e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 86  1.156521e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 87  1.162457e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 88  1.155639e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 89  1.153398e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 90  1.151586e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 91  1.135564e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 92  1.158214e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 93  1.138186e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 94  1.163960e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 95  1.144338e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 96  1.142287e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 97  1.157975e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 98  1.146436e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 99  1.146054e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 100 1.154137e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 101 1.149821e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 102 1.156831e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 103 1.192236e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 104 1.157522e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 105 1.148200e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 106 1.140738e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 107 1.135015e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 108 1.192045e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 109 1.151156e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 110 1.137495e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 111 1.135397e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 112 1.146173e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 113 1.150465e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 114 1.197457e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 115 1.145172e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 116 1.139569e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 117 1.142311e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 118 1.135421e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 119 1.147914e-02 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 123 8.184910e-03 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 120 3.449917e-04 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 121 1.597404e-04 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#> 124 2.479553e-05 secs         <NA> 6128f31e79476bcb789b83bdb3b16a22-19
#>                         session_uuid
#> 1   6128f31e79476bcb789b83bdb3b16a22
#> 4   6128f31e79476bcb789b83bdb3b16a22
#> 2   6128f31e79476bcb789b83bdb3b16a22
#> 3   6128f31e79476bcb789b83bdb3b16a22
#> 5   6128f31e79476bcb789b83bdb3b16a22
#> 122 6128f31e79476bcb789b83bdb3b16a22
#> 6   6128f31e79476bcb789b83bdb3b16a22
#> 7   6128f31e79476bcb789b83bdb3b16a22
#> 8   6128f31e79476bcb789b83bdb3b16a22
#> 9   6128f31e79476bcb789b83bdb3b16a22
#> 10  6128f31e79476bcb789b83bdb3b16a22
#> 11  6128f31e79476bcb789b83bdb3b16a22
#> 12  6128f31e79476bcb789b83bdb3b16a22
#> 13  6128f31e79476bcb789b83bdb3b16a22
#> 14  6128f31e79476bcb789b83bdb3b16a22
#> 15  6128f31e79476bcb789b83bdb3b16a22
#> 16  6128f31e79476bcb789b83bdb3b16a22
#> 17  6128f31e79476bcb789b83bdb3b16a22
#> 18  6128f31e79476bcb789b83bdb3b16a22
#> 19  6128f31e79476bcb789b83bdb3b16a22
#> 20  6128f31e79476bcb789b83bdb3b16a22
#> 21  6128f31e79476bcb789b83bdb3b16a22
#> 22  6128f31e79476bcb789b83bdb3b16a22
#> 23  6128f31e79476bcb789b83bdb3b16a22
#> 24  6128f31e79476bcb789b83bdb3b16a22
#> 25  6128f31e79476bcb789b83bdb3b16a22
#> 26  6128f31e79476bcb789b83bdb3b16a22
#> 27  6128f31e79476bcb789b83bdb3b16a22
#> 28  6128f31e79476bcb789b83bdb3b16a22
#> 29  6128f31e79476bcb789b83bdb3b16a22
#> 30  6128f31e79476bcb789b83bdb3b16a22
#> 31  6128f31e79476bcb789b83bdb3b16a22
#> 32  6128f31e79476bcb789b83bdb3b16a22
#> 33  6128f31e79476bcb789b83bdb3b16a22
#> 34  6128f31e79476bcb789b83bdb3b16a22
#> 35  6128f31e79476bcb789b83bdb3b16a22
#> 36  6128f31e79476bcb789b83bdb3b16a22
#> 37  6128f31e79476bcb789b83bdb3b16a22
#> 38  6128f31e79476bcb789b83bdb3b16a22
#> 39  6128f31e79476bcb789b83bdb3b16a22
#> 40  6128f31e79476bcb789b83bdb3b16a22
#> 41  6128f31e79476bcb789b83bdb3b16a22
#> 42  6128f31e79476bcb789b83bdb3b16a22
#> 43  6128f31e79476bcb789b83bdb3b16a22
#> 44  6128f31e79476bcb789b83bdb3b16a22
#> 45  6128f31e79476bcb789b83bdb3b16a22
#> 46  6128f31e79476bcb789b83bdb3b16a22
#> 47  6128f31e79476bcb789b83bdb3b16a22
#> 48  6128f31e79476bcb789b83bdb3b16a22
#> 49  6128f31e79476bcb789b83bdb3b16a22
#> 50  6128f31e79476bcb789b83bdb3b16a22
#> 51  6128f31e79476bcb789b83bdb3b16a22
#> 52  6128f31e79476bcb789b83bdb3b16a22
#> 53  6128f31e79476bcb789b83bdb3b16a22
#> 54  6128f31e79476bcb789b83bdb3b16a22
#> 55  6128f31e79476bcb789b83bdb3b16a22
#> 56  6128f31e79476bcb789b83bdb3b16a22
#> 57  6128f31e79476bcb789b83bdb3b16a22
#> 58  6128f31e79476bcb789b83bdb3b16a22
#> 59  6128f31e79476bcb789b83bdb3b16a22
#> 60  6128f31e79476bcb789b83bdb3b16a22
#> 61  6128f31e79476bcb789b83bdb3b16a22
#> 62  6128f31e79476bcb789b83bdb3b16a22
#> 63  6128f31e79476bcb789b83bdb3b16a22
#> 64  6128f31e79476bcb789b83bdb3b16a22
#> 65  6128f31e79476bcb789b83bdb3b16a22
#> 66  6128f31e79476bcb789b83bdb3b16a22
#> 67  6128f31e79476bcb789b83bdb3b16a22
#> 68  6128f31e79476bcb789b83bdb3b16a22
#> 69  6128f31e79476bcb789b83bdb3b16a22
#> 70  6128f31e79476bcb789b83bdb3b16a22
#> 71  6128f31e79476bcb789b83bdb3b16a22
#> 72  6128f31e79476bcb789b83bdb3b16a22
#> 73  6128f31e79476bcb789b83bdb3b16a22
#> 74  6128f31e79476bcb789b83bdb3b16a22
#> 75  6128f31e79476bcb789b83bdb3b16a22
#> 76  6128f31e79476bcb789b83bdb3b16a22
#> 77  6128f31e79476bcb789b83bdb3b16a22
#> 78  6128f31e79476bcb789b83bdb3b16a22
#> 79  6128f31e79476bcb789b83bdb3b16a22
#> 80  6128f31e79476bcb789b83bdb3b16a22
#> 81  6128f31e79476bcb789b83bdb3b16a22
#> 82  6128f31e79476bcb789b83bdb3b16a22
#> 83  6128f31e79476bcb789b83bdb3b16a22
#> 84  6128f31e79476bcb789b83bdb3b16a22
#> 85  6128f31e79476bcb789b83bdb3b16a22
#> 86  6128f31e79476bcb789b83bdb3b16a22
#> 87  6128f31e79476bcb789b83bdb3b16a22
#> 88  6128f31e79476bcb789b83bdb3b16a22
#> 89  6128f31e79476bcb789b83bdb3b16a22
#> 90  6128f31e79476bcb789b83bdb3b16a22
#> 91  6128f31e79476bcb789b83bdb3b16a22
#> 92  6128f31e79476bcb789b83bdb3b16a22
#> 93  6128f31e79476bcb789b83bdb3b16a22
#> 94  6128f31e79476bcb789b83bdb3b16a22
#> 95  6128f31e79476bcb789b83bdb3b16a22
#> 96  6128f31e79476bcb789b83bdb3b16a22
#> 97  6128f31e79476bcb789b83bdb3b16a22
#> 98  6128f31e79476bcb789b83bdb3b16a22
#> 99  6128f31e79476bcb789b83bdb3b16a22
#> 100 6128f31e79476bcb789b83bdb3b16a22
#> 101 6128f31e79476bcb789b83bdb3b16a22
#> 102 6128f31e79476bcb789b83bdb3b16a22
#> 103 6128f31e79476bcb789b83bdb3b16a22
#> 104 6128f31e79476bcb789b83bdb3b16a22
#> 105 6128f31e79476bcb789b83bdb3b16a22
#> 106 6128f31e79476bcb789b83bdb3b16a22
#> 107 6128f31e79476bcb789b83bdb3b16a22
#> 108 6128f31e79476bcb789b83bdb3b16a22
#> 109 6128f31e79476bcb789b83bdb3b16a22
#> 110 6128f31e79476bcb789b83bdb3b16a22
#> 111 6128f31e79476bcb789b83bdb3b16a22
#> 112 6128f31e79476bcb789b83bdb3b16a22
#> 113 6128f31e79476bcb789b83bdb3b16a22
#> 114 6128f31e79476bcb789b83bdb3b16a22
#> 115 6128f31e79476bcb789b83bdb3b16a22
#> 116 6128f31e79476bcb789b83bdb3b16a22
#> 117 6128f31e79476bcb789b83bdb3b16a22
#> 118 6128f31e79476bcb789b83bdb3b16a22
#> 119 6128f31e79476bcb789b83bdb3b16a22
#> 123 6128f31e79476bcb789b83bdb3b16a22
#> 120 6128f31e79476bcb789b83bdb3b16a22
#> 121 6128f31e79476bcb789b83bdb3b16a22
#> 124 6128f31e79476bcb789b83bdb3b16a22
#> 

## Stop parallel workers and disable journal logging and signaling
plan(sequential)
options(oopts)