Gets the logged journal of events for a future

WARNING: This function is under development. It can change at any time. For now, please, do not depend on this function in a published R package.

Usage

journal(x, ...)

Arguments

x

A future::Future object.

...

Not used.

Value

A data frame of class FutureJournal with columns:

event (character string) - type of event that took place

category (character string) - the category of the event

parent (character string) - (to be describe)

start (POSIXct) - the timestamp when the event started

at (difftime) - the time when the event started relative to first event

duration (difftime) - the duration of the event

future_label (character string) - the label of the future

future_uuid (character string) - the UUID of the future

session_uuid (character string) - the UUID of the R session where the event took place

The common events are:

create - the future was created (an overhead)

launch - the future was launched (an overhead)

evaluate - the future was evaluated (an evaluation)

resolved - the future was queried (may be occur multiple times) (an overhead)

gather - the results was retrieved (an overhead)

but others may be added by other Future classes.

Common event categories are:

evaluation - processing time is spent on evaluation

overhead - processing time is spent on orchestrating the future

waiting - processing time is spent on waiting to set up or querying the future

but others may be added by other Future classes.

The data frame is sorted by the at time. Note that the timestamps for the evaluate event are based on the local time on the worker. The system clocks on the worker and the calling R system may not be in perfect sync.

Enabling and disabling event logging

To enable logging of events, set option future.journal is TRUE. To disable, set it to FALSE (default).

See also

Examples

## Enable journaling of futures
oopts <- options(future.journal = TRUE)

plan(multisession, workers = 2L)

t_start <- Sys.time()
fs <- lapply(1:3, FUN = function(x) future({ Sys.sleep(x); sqrt(x) }))
vs <- value(fs)
js <- lapply(fs, FUN = journal, baseline = t_start)
print(js)
#> [[1]]
#>            event   category parent               start               at
#> 1         create   overhead   <NA> 2026-01-14 19:51:23 0.000000000 secs
#> 4         launch   overhead   <NA> 2026-01-14 19:51:23 0.008593321 secs
#> 2      getWorker   overhead launch 2026-01-14 19:51:23 0.008731842 secs
#> 3    eraseWorker   overhead launch 2026-01-14 19:51:23 0.010023594 secs
#> 5       resolved       <NA>   <NA> 2026-01-14 19:51:23 0.059194803 secs
#> 6       resolved       <NA>   <NA> 2026-01-14 19:51:24 0.095294952 secs
#> 7       resolved       <NA>   <NA> 2026-01-14 19:51:24 0.130805731 secs
#> 8       resolved       <NA>   <NA> 2026-01-14 19:51:24 0.166574478 secs
#> 9       resolved       <NA>   <NA> 2026-01-14 19:51:24 0.202359915 secs
#> 42      evaluate evaluation   <NA> 2026-01-14 19:51:24 0.224875212 secs
#> 10      resolved       <NA>   <NA> 2026-01-14 19:51:24 0.238552332 secs
#> 11      resolved       <NA>   <NA> 2026-01-14 19:51:24 0.274458885 secs
#> 12      resolved       <NA>   <NA> 2026-01-14 19:51:24 0.309192896 secs
#> 13      resolved       <NA>   <NA> 2026-01-14 19:51:24 0.343973637 secs
#> 14      resolved       <NA>   <NA> 2026-01-14 19:51:24 0.383368492 secs
#> 15      resolved       <NA>   <NA> 2026-01-14 19:51:24 0.418266773 secs
#> 16      resolved       <NA>   <NA> 2026-01-14 19:51:24 0.454656363 secs
#> 17      resolved       <NA>   <NA> 2026-01-14 19:51:24 0.494719028 secs
#> 18      resolved       <NA>   <NA> 2026-01-14 19:51:24 0.534206390 secs
#> 19      resolved       <NA>   <NA> 2026-01-14 19:51:24 0.572472095 secs
#> 20      resolved       <NA>   <NA> 2026-01-14 19:51:24 0.610657454 secs
#> 21      resolved       <NA>   <NA> 2026-01-14 19:51:24 0.648465872 secs
#> 22      resolved       <NA>   <NA> 2026-01-14 19:51:24 0.686907530 secs
#> 23      resolved       <NA>   <NA> 2026-01-14 19:51:24 0.724753141 secs
#> 24      resolved       <NA>   <NA> 2026-01-14 19:51:24 0.760356426 secs
#> 25      resolved       <NA>   <NA> 2026-01-14 19:51:24 0.795613289 secs
#> 26      resolved       <NA>   <NA> 2026-01-14 19:51:24 0.830964088 secs
#> 27      resolved       <NA>   <NA> 2026-01-14 19:51:24 0.867417812 secs
#> 28      resolved       <NA>   <NA> 2026-01-14 19:51:24 0.903293133 secs
#> 29      resolved       <NA>   <NA> 2026-01-14 19:51:24 0.939261675 secs
#> 30      resolved       <NA>   <NA> 2026-01-14 19:51:24 0.975107670 secs
#> 31      resolved       <NA>   <NA> 2026-01-14 19:51:24 1.011286497 secs
#> 32      resolved       <NA>   <NA> 2026-01-14 19:51:24 1.048685551 secs
#> 33      resolved       <NA>   <NA> 2026-01-14 19:51:25 1.087046385 secs
#> 34      resolved       <NA>   <NA> 2026-01-14 19:51:25 1.126947880 secs
#> 35      resolved       <NA>   <NA> 2026-01-14 19:51:25 1.167100668 secs
#> 36      resolved       <NA>   <NA> 2026-01-14 19:51:25 1.207737684 secs
#> 37      resolved       <NA>   <NA> 2026-01-14 19:51:25 1.248580694 secs
#> 38      resolved       <NA>   <NA> 2026-01-14 19:51:25 1.289432049 secs
#> 39      resolved       <NA>   <NA> 2026-01-14 19:51:25 1.329952240 secs
#> 43      resolved       <NA>   <NA> 2026-01-14 19:51:25 1.370924711 secs
#> 40 receiveResult   overhead gather 2026-01-14 19:51:25 1.372236729 secs
#> 41        gather   overhead   <NA> 2026-01-14 19:51:25 1.374105930 secs
#> 44      resolved       <NA>   <NA> 2026-01-14 19:51:25 1.395249844 secs
#> 45       resolve   overhead   <NA> 2026-01-14 19:51:25 1.397588015 secs
#>             duration future_label                         future_uuid
#> 1  8.218527e-03 secs         <NA> 9388f6f894218a6b8e5b343286474110-17
#> 4  5.913019e-03 secs         <NA> 9388f6f894218a6b8e5b343286474110-17
#> 2  6.432533e-04 secs         <NA> 9388f6f894218a6b8e5b343286474110-17
#> 3  4.737377e-04 secs         <NA> 9388f6f894218a6b8e5b343286474110-17
#> 5  1.148319e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-17
#> 6  1.123667e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-17
#> 7  1.118755e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-17
#> 8  1.127815e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-17
#> 9  1.122069e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-17
#> 42 1.134307e+00 secs         <NA> 9388f6f894218a6b8e5b343286474110-17
#> 10 1.126456e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-17
#> 11 1.117277e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-17
#> 12 1.078320e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-17
#> 13 1.495528e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-17
#> 14 1.093221e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-17
#> 15 1.123691e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-17
#> 16 1.169920e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-17
#> 17 1.172900e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-17
#> 18 1.141882e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-17
#> 19 1.137137e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-17
#> 20 1.136374e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-17
#> 21 1.195693e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-17
#> 22 1.135492e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-17
#> 23 1.089573e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-17
#> 24 1.074386e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-17
#> 25 1.076078e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-17
#> 26 1.104450e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-17
#> 27 1.083422e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-17
#> 28 1.082492e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-17
#> 29 1.072359e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-17
#> 30 1.081324e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-17
#> 31 1.081967e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-17
#> 32 1.115346e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-17
#> 33 1.159263e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-17
#> 34 1.159024e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-17
#> 35 1.156425e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-17
#> 36 1.166296e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-17
#> 37 1.177788e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-17
#> 38 1.146626e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-17
#> 39 1.158428e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-17
#> 43 8.057833e-03 secs         <NA> 9388f6f894218a6b8e5b343286474110-17
#> 40 2.801418e-04 secs         <NA> 9388f6f894218a6b8e5b343286474110-17
#> 41 1.380444e-04 secs         <NA> 9388f6f894218a6b8e5b343286474110-17
#> 44 1.239777e-04 secs         <NA> 9388f6f894218a6b8e5b343286474110-17
#> 45 3.290176e-05 secs         <NA> 9388f6f894218a6b8e5b343286474110-17
#>                        session_uuid
#> 1  9388f6f894218a6b8e5b343286474110
#> 4  9388f6f894218a6b8e5b343286474110
#> 2  9388f6f894218a6b8e5b343286474110
#> 3  9388f6f894218a6b8e5b343286474110
#> 5  9388f6f894218a6b8e5b343286474110
#> 6  9388f6f894218a6b8e5b343286474110
#> 7  9388f6f894218a6b8e5b343286474110
#> 8  9388f6f894218a6b8e5b343286474110
#> 9  9388f6f894218a6b8e5b343286474110
#> 42 9388f6f894218a6b8e5b343286474110
#> 10 9388f6f894218a6b8e5b343286474110
#> 11 9388f6f894218a6b8e5b343286474110
#> 12 9388f6f894218a6b8e5b343286474110
#> 13 9388f6f894218a6b8e5b343286474110
#> 14 9388f6f894218a6b8e5b343286474110
#> 15 9388f6f894218a6b8e5b343286474110
#> 16 9388f6f894218a6b8e5b343286474110
#> 17 9388f6f894218a6b8e5b343286474110
#> 18 9388f6f894218a6b8e5b343286474110
#> 19 9388f6f894218a6b8e5b343286474110
#> 20 9388f6f894218a6b8e5b343286474110
#> 21 9388f6f894218a6b8e5b343286474110
#> 22 9388f6f894218a6b8e5b343286474110
#> 23 9388f6f894218a6b8e5b343286474110
#> 24 9388f6f894218a6b8e5b343286474110
#> 25 9388f6f894218a6b8e5b343286474110
#> 26 9388f6f894218a6b8e5b343286474110
#> 27 9388f6f894218a6b8e5b343286474110
#> 28 9388f6f894218a6b8e5b343286474110
#> 29 9388f6f894218a6b8e5b343286474110
#> 30 9388f6f894218a6b8e5b343286474110
#> 31 9388f6f894218a6b8e5b343286474110
#> 32 9388f6f894218a6b8e5b343286474110
#> 33 9388f6f894218a6b8e5b343286474110
#> 34 9388f6f894218a6b8e5b343286474110
#> 35 9388f6f894218a6b8e5b343286474110
#> 36 9388f6f894218a6b8e5b343286474110
#> 37 9388f6f894218a6b8e5b343286474110
#> 38 9388f6f894218a6b8e5b343286474110
#> 39 9388f6f894218a6b8e5b343286474110
#> 43 9388f6f894218a6b8e5b343286474110
#> 40 9388f6f894218a6b8e5b343286474110
#> 41 9388f6f894218a6b8e5b343286474110
#> 44 9388f6f894218a6b8e5b343286474110
#> 45 9388f6f894218a6b8e5b343286474110
#> 
#> [[2]]
#>            event   category parent               start               at
#> 1         create   overhead   <NA> 2026-01-14 19:51:23 0.000000000 secs
#> 4         launch   overhead   <NA> 2026-01-14 19:51:23 0.008690834 secs
#> 2      getWorker   overhead launch 2026-01-14 19:51:23 0.008833885 secs
#> 3    eraseWorker   overhead launch 2026-01-14 19:51:23 0.010107756 secs
#> 5       resolved       <NA>   <NA> 2026-01-14 19:51:23 0.056959152 secs
#> 6       resolved       <NA>   <NA> 2026-01-14 19:51:24 0.092629910 secs
#> 7       resolved       <NA>   <NA> 2026-01-14 19:51:24 0.128264904 secs
#> 8       resolved       <NA>   <NA> 2026-01-14 19:51:24 0.164082527 secs
#> 9       resolved       <NA>   <NA> 2026-01-14 19:51:24 0.199860573 secs
#> 10      resolved       <NA>   <NA> 2026-01-14 19:51:24 0.235960245 secs
#> 70      evaluate evaluation   <NA> 2026-01-14 19:51:24 0.245939970 secs
#> 11      resolved       <NA>   <NA> 2026-01-14 19:51:24 0.271466255 secs
#> 12      resolved       <NA>   <NA> 2026-01-14 19:51:24 0.305741072 secs
#> 13      resolved       <NA>   <NA> 2026-01-14 19:51:24 0.345114470 secs
#> 14      resolved       <NA>   <NA> 2026-01-14 19:51:24 0.380093098 secs
#> 15      resolved       <NA>   <NA> 2026-01-14 19:51:24 0.415643215 secs
#> 16      resolved       <NA>   <NA> 2026-01-14 19:51:24 0.453804255 secs
#> 17      resolved       <NA>   <NA> 2026-01-14 19:51:24 0.494336605 secs
#> 18      resolved       <NA>   <NA> 2026-01-14 19:51:24 0.532330751 secs
#> 19      resolved       <NA>   <NA> 2026-01-14 19:51:24 0.570294857 secs
#> 20      resolved       <NA>   <NA> 2026-01-14 19:51:24 0.608433723 secs
#> 21      resolved       <NA>   <NA> 2026-01-14 19:51:24 0.646826029 secs
#> 22      resolved       <NA>   <NA> 2026-01-14 19:51:24 0.684694052 secs
#> 23      resolved       <NA>   <NA> 2026-01-14 19:51:24 0.721357346 secs
#> 24      resolved       <NA>   <NA> 2026-01-14 19:51:24 0.756757736 secs
#> 25      resolved       <NA>   <NA> 2026-01-14 19:51:24 0.791934967 secs
#> 26      resolved       <NA>   <NA> 2026-01-14 19:51:24 0.827788353 secs
#> 27      resolved       <NA>   <NA> 2026-01-14 19:51:24 0.863865852 secs
#> 28      resolved       <NA>   <NA> 2026-01-14 19:51:24 0.899749994 secs
#> 29      resolved       <NA>   <NA> 2026-01-14 19:51:24 0.935565233 secs
#> 30      resolved       <NA>   <NA> 2026-01-14 19:51:24 0.971535921 secs
#> 31      resolved       <NA>   <NA> 2026-01-14 19:51:24 1.008052588 secs
#> 32      resolved       <NA>   <NA> 2026-01-14 19:51:24 1.045762539 secs
#> 33      resolved       <NA>   <NA> 2026-01-14 19:51:25 1.085123539 secs
#> 34      resolved       <NA>   <NA> 2026-01-14 19:51:25 1.125020742 secs
#> 35      resolved       <NA>   <NA> 2026-01-14 19:51:25 1.165237665 secs
#> 36      resolved       <NA>   <NA> 2026-01-14 19:51:25 1.206022024 secs
#> 37      resolved       <NA>   <NA> 2026-01-14 19:51:25 1.246818304 secs
#> 38      resolved       <NA>   <NA> 2026-01-14 19:51:25 1.287380219 secs
#> 39      resolved       <NA>   <NA> 2026-01-14 19:51:25 1.328033447 secs
#> 40      resolved       <NA>   <NA> 2026-01-14 19:51:25 1.383724689 secs
#> 41      resolved       <NA>   <NA> 2026-01-14 19:51:25 1.420138121 secs
#> 42      resolved       <NA>   <NA> 2026-01-14 19:51:25 1.456385136 secs
#> 43      resolved       <NA>   <NA> 2026-01-14 19:51:25 1.492659807 secs
#> 44      resolved       <NA>   <NA> 2026-01-14 19:51:25 1.529336214 secs
#> 45      resolved       <NA>   <NA> 2026-01-14 19:51:25 1.567932367 secs
#> 46      resolved       <NA>   <NA> 2026-01-14 19:51:25 1.605030060 secs
#> 47      resolved       <NA>   <NA> 2026-01-14 19:51:25 1.639468431 secs
#> 48      resolved       <NA>   <NA> 2026-01-14 19:51:25 1.675348759 secs
#> 49      resolved       <NA>   <NA> 2026-01-14 19:51:25 1.712017298 secs
#> 50      resolved       <NA>   <NA> 2026-01-14 19:51:25 1.748287439 secs
#> 51      resolved       <NA>   <NA> 2026-01-14 19:51:25 1.784829378 secs
#> 52      resolved       <NA>   <NA> 2026-01-14 19:51:25 1.821671486 secs
#> 53      resolved       <NA>   <NA> 2026-01-14 19:51:25 1.858786583 secs
#> 54      resolved       <NA>   <NA> 2026-01-14 19:51:25 1.895516634 secs
#> 55      resolved       <NA>   <NA> 2026-01-14 19:51:25 1.932428122 secs
#> 56      resolved       <NA>   <NA> 2026-01-14 19:51:25 1.968952179 secs
#> 57      resolved       <NA>   <NA> 2026-01-14 19:51:25 2.005539179 secs
#> 58      resolved       <NA>   <NA> 2026-01-14 19:51:25 2.042342424 secs
#> 59      resolved       <NA>   <NA> 2026-01-14 19:51:26 2.079275846 secs
#> 60      resolved       <NA>   <NA> 2026-01-14 19:51:26 2.116383076 secs
#> 61      resolved       <NA>   <NA> 2026-01-14 19:51:26 2.153044462 secs
#> 62      resolved       <NA>   <NA> 2026-01-14 19:51:26 2.190050602 secs
#> 63      resolved       <NA>   <NA> 2026-01-14 19:51:26 2.226627350 secs
#> 64      resolved       <NA>   <NA> 2026-01-14 19:51:26 2.263197899 secs
#> 65      resolved       <NA>   <NA> 2026-01-14 19:51:26 2.300268412 secs
#> 66      resolved       <NA>   <NA> 2026-01-14 19:51:26 2.336972475 secs
#> 67      resolved       <NA>   <NA> 2026-01-14 19:51:26 2.373982191 secs
#> 71      resolved       <NA>   <NA> 2026-01-14 19:51:26 2.412017822 secs
#> 68 receiveResult   overhead gather 2026-01-14 19:51:26 2.413352489 secs
#> 69        gather   overhead   <NA> 2026-01-14 19:51:26 2.415272474 secs
#> 72       resolve   overhead   <NA> 2026-01-14 19:51:26 2.422301054 secs
#>             duration future_label                         future_uuid
#> 1  0.0083487034 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 4  0.0072801113 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 2  0.0005986691 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 3  0.0004315376 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 5  0.0114965439 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 6  0.0111734867 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 7  0.0111689568 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 8  0.0111176968 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 9  0.0111572742 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 10 0.0111353397 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 70 2.1423957348 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 11 0.0107886791 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 12 0.0108544827 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 13 0.0110054016 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 14 0.0107867718 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 15 0.0110213757 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 16 0.0122077465 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 17 0.0113203526 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 18 0.0115141869 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 19 0.0115172863 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 20 0.0111925602 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 21 0.0111925602 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 22 0.0115509033 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 23 0.0106775761 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 24 0.0105996132 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 25 0.0106060505 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 26 0.0107378960 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 27 0.0106866360 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 28 0.0107059479 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 29 0.0106689930 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 30 0.0106673241 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 31 0.0109395981 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 32 0.0109376907 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 33 0.0112581253 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 34 0.0113098621 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 35 0.0114750862 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 36 0.0114998817 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 37 0.0114574432 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 38 0.0113873482 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 39 0.0114471912 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 40 0.0112783909 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 41 0.0115525723 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 42 0.0116338730 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 43 0.0116488934 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 44 0.0116424561 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 45 0.0119168758 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 46 0.0110011101 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 47 0.0110068321 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 48 0.0118172169 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 49 0.0115103722 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 50 0.0114767551 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 51 0.0116755962 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 52 0.0117161274 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 53 0.0115003586 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 54 0.0117769241 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 55 0.0115990639 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 56 0.0116174221 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 57 0.0116286278 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 58 0.0117111206 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 59 0.0117437840 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 60 0.0117790699 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 61 0.0118145943 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 62 0.0116651058 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 63 0.0116121769 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 64 0.0115909576 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 65 0.0116145611 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 66 0.0117459297 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 67 0.0116200447 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 71 0.0080277920 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 68 0.0002868176 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 69 0.0001275539 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#> 72 0.0000333786 secs         <NA> 9388f6f894218a6b8e5b343286474110-18
#>                        session_uuid
#> 1  9388f6f894218a6b8e5b343286474110
#> 4  9388f6f894218a6b8e5b343286474110
#> 2  9388f6f894218a6b8e5b343286474110
#> 3  9388f6f894218a6b8e5b343286474110
#> 5  9388f6f894218a6b8e5b343286474110
#> 6  9388f6f894218a6b8e5b343286474110
#> 7  9388f6f894218a6b8e5b343286474110
#> 8  9388f6f894218a6b8e5b343286474110
#> 9  9388f6f894218a6b8e5b343286474110
#> 10 9388f6f894218a6b8e5b343286474110
#> 70 9388f6f894218a6b8e5b343286474110
#> 11 9388f6f894218a6b8e5b343286474110
#> 12 9388f6f894218a6b8e5b343286474110
#> 13 9388f6f894218a6b8e5b343286474110
#> 14 9388f6f894218a6b8e5b343286474110
#> 15 9388f6f894218a6b8e5b343286474110
#> 16 9388f6f894218a6b8e5b343286474110
#> 17 9388f6f894218a6b8e5b343286474110
#> 18 9388f6f894218a6b8e5b343286474110
#> 19 9388f6f894218a6b8e5b343286474110
#> 20 9388f6f894218a6b8e5b343286474110
#> 21 9388f6f894218a6b8e5b343286474110
#> 22 9388f6f894218a6b8e5b343286474110
#> 23 9388f6f894218a6b8e5b343286474110
#> 24 9388f6f894218a6b8e5b343286474110
#> 25 9388f6f894218a6b8e5b343286474110
#> 26 9388f6f894218a6b8e5b343286474110
#> 27 9388f6f894218a6b8e5b343286474110
#> 28 9388f6f894218a6b8e5b343286474110
#> 29 9388f6f894218a6b8e5b343286474110
#> 30 9388f6f894218a6b8e5b343286474110
#> 31 9388f6f894218a6b8e5b343286474110
#> 32 9388f6f894218a6b8e5b343286474110
#> 33 9388f6f894218a6b8e5b343286474110
#> 34 9388f6f894218a6b8e5b343286474110
#> 35 9388f6f894218a6b8e5b343286474110
#> 36 9388f6f894218a6b8e5b343286474110
#> 37 9388f6f894218a6b8e5b343286474110
#> 38 9388f6f894218a6b8e5b343286474110
#> 39 9388f6f894218a6b8e5b343286474110
#> 40 9388f6f894218a6b8e5b343286474110
#> 41 9388f6f894218a6b8e5b343286474110
#> 42 9388f6f894218a6b8e5b343286474110
#> 43 9388f6f894218a6b8e5b343286474110
#> 44 9388f6f894218a6b8e5b343286474110
#> 45 9388f6f894218a6b8e5b343286474110
#> 46 9388f6f894218a6b8e5b343286474110
#> 47 9388f6f894218a6b8e5b343286474110
#> 48 9388f6f894218a6b8e5b343286474110
#> 49 9388f6f894218a6b8e5b343286474110
#> 50 9388f6f894218a6b8e5b343286474110
#> 51 9388f6f894218a6b8e5b343286474110
#> 52 9388f6f894218a6b8e5b343286474110
#> 53 9388f6f894218a6b8e5b343286474110
#> 54 9388f6f894218a6b8e5b343286474110
#> 55 9388f6f894218a6b8e5b343286474110
#> 56 9388f6f894218a6b8e5b343286474110
#> 57 9388f6f894218a6b8e5b343286474110
#> 58 9388f6f894218a6b8e5b343286474110
#> 59 9388f6f894218a6b8e5b343286474110
#> 60 9388f6f894218a6b8e5b343286474110
#> 61 9388f6f894218a6b8e5b343286474110
#> 62 9388f6f894218a6b8e5b343286474110
#> 63 9388f6f894218a6b8e5b343286474110
#> 64 9388f6f894218a6b8e5b343286474110
#> 65 9388f6f894218a6b8e5b343286474110
#> 66 9388f6f894218a6b8e5b343286474110
#> 67 9388f6f894218a6b8e5b343286474110
#> 71 9388f6f894218a6b8e5b343286474110
#> 68 9388f6f894218a6b8e5b343286474110
#> 69 9388f6f894218a6b8e5b343286474110
#> 72 9388f6f894218a6b8e5b343286474110
#> 
#> [[3]]
#>             event   category parent               start              at
#> 1          create   overhead   <NA> 2026-01-14 19:51:23 0.00000000 secs
#> 4          launch   overhead   <NA> 2026-01-14 19:51:23 0.01646185 secs
#> 2       getWorker   overhead launch 2026-01-14 19:51:23 0.01671529 secs
#> 3     eraseWorker   overhead launch 2026-01-14 19:51:25 1.35126853 secs
#> 5        resolved       <NA>   <NA> 2026-01-14 19:51:25 1.37954617 secs
#> 120      evaluate evaluation   <NA> 2026-01-14 19:51:25 1.40209532 secs
#> 6        resolved       <NA>   <NA> 2026-01-14 19:51:25 1.41623664 secs
#> 7        resolved       <NA>   <NA> 2026-01-14 19:51:25 1.45257664 secs
#> 8        resolved       <NA>   <NA> 2026-01-14 19:51:25 1.48895407 secs
#> 9        resolved       <NA>   <NA> 2026-01-14 19:51:25 1.52644610 secs
#> 10       resolved       <NA>   <NA> 2026-01-14 19:51:25 1.56494784 secs
#> 11       resolved       <NA>   <NA> 2026-01-14 19:51:25 1.60010362 secs
#> 12       resolved       <NA>   <NA> 2026-01-14 19:51:25 1.63499188 secs
#> 13       resolved       <NA>   <NA> 2026-01-14 19:51:25 1.67172122 secs
#> 14       resolved       <NA>   <NA> 2026-01-14 19:51:25 1.70810080 secs
#> 15       resolved       <NA>   <NA> 2026-01-14 19:51:25 1.74429917 secs
#> 16       resolved       <NA>   <NA> 2026-01-14 19:51:25 1.78124952 secs
#> 17       resolved       <NA>   <NA> 2026-01-14 19:51:25 1.81818700 secs
#> 18       resolved       <NA>   <NA> 2026-01-14 19:51:25 1.85495353 secs
#> 19       resolved       <NA>   <NA> 2026-01-14 19:51:25 1.89190841 secs
#> 20       resolved       <NA>   <NA> 2026-01-14 19:51:25 1.92858934 secs
#> 21       resolved       <NA>   <NA> 2026-01-14 19:51:25 1.96513486 secs
#> 22       resolved       <NA>   <NA> 2026-01-14 19:51:25 2.00184464 secs
#> 23       resolved       <NA>   <NA> 2026-01-14 19:51:25 2.03872943 secs
#> 24       resolved       <NA>   <NA> 2026-01-14 19:51:26 2.07578254 secs
#> 25       resolved       <NA>   <NA> 2026-01-14 19:51:26 2.11272717 secs
#> 26       resolved       <NA>   <NA> 2026-01-14 19:51:26 2.14950967 secs
#> 27       resolved       <NA>   <NA> 2026-01-14 19:51:26 2.18625116 secs
#> 28       resolved       <NA>   <NA> 2026-01-14 19:51:26 2.22279882 secs
#> 29       resolved       <NA>   <NA> 2026-01-14 19:51:26 2.25977659 secs
#> 30       resolved       <NA>   <NA> 2026-01-14 19:51:26 2.29645896 secs
#> 31       resolved       <NA>   <NA> 2026-01-14 19:51:26 2.33339381 secs
#> 32       resolved       <NA>   <NA> 2026-01-14 19:51:26 2.37108541 secs
#> 33       resolved       <NA>   <NA> 2026-01-14 19:51:26 2.40655184 secs
#> 34       resolved       <NA>   <NA> 2026-01-14 19:51:26 2.42976522 secs
#> 35       resolved       <NA>   <NA> 2026-01-14 19:51:26 2.46064878 secs
#> 36       resolved       <NA>   <NA> 2026-01-14 19:51:26 2.48442626 secs
#> 37       resolved       <NA>   <NA> 2026-01-14 19:51:26 2.50810289 secs
#> 38       resolved       <NA>   <NA> 2026-01-14 19:51:26 2.53181505 secs
#> 39       resolved       <NA>   <NA> 2026-01-14 19:51:26 2.55546403 secs
#> 40       resolved       <NA>   <NA> 2026-01-14 19:51:26 2.57935596 secs
#> 41       resolved       <NA>   <NA> 2026-01-14 19:51:26 2.60325074 secs
#> 42       resolved       <NA>   <NA> 2026-01-14 19:51:26 2.62663746 secs
#> 43       resolved       <NA>   <NA> 2026-01-14 19:51:26 2.65007901 secs
#> 44       resolved       <NA>   <NA> 2026-01-14 19:51:26 2.67365503 secs
#> 45       resolved       <NA>   <NA> 2026-01-14 19:51:26 2.69722557 secs
#> 46       resolved       <NA>   <NA> 2026-01-14 19:51:26 2.72080612 secs
#> 47       resolved       <NA>   <NA> 2026-01-14 19:51:26 2.74456954 secs
#> 48       resolved       <NA>   <NA> 2026-01-14 19:51:26 2.76836514 secs
#> 49       resolved       <NA>   <NA> 2026-01-14 19:51:26 2.79246759 secs
#> 50       resolved       <NA>   <NA> 2026-01-14 19:51:26 2.81587434 secs
#> 51       resolved       <NA>   <NA> 2026-01-14 19:51:26 2.83951855 secs
#> 52       resolved       <NA>   <NA> 2026-01-14 19:51:26 2.86318731 secs
#> 53       resolved       <NA>   <NA> 2026-01-14 19:51:26 2.88679361 secs
#> 54       resolved       <NA>   <NA> 2026-01-14 19:51:26 2.91044855 secs
#> 55       resolved       <NA>   <NA> 2026-01-14 19:51:26 2.93423748 secs
#> 56       resolved       <NA>   <NA> 2026-01-14 19:51:26 2.95792294 secs
#> 57       resolved       <NA>   <NA> 2026-01-14 19:51:26 2.98179770 secs
#> 58       resolved       <NA>   <NA> 2026-01-14 19:51:26 3.00538588 secs
#> 59       resolved       <NA>   <NA> 2026-01-14 19:51:26 3.02919340 secs
#> 60       resolved       <NA>   <NA> 2026-01-14 19:51:27 3.05276179 secs
#> 61       resolved       <NA>   <NA> 2026-01-14 19:51:27 3.07632470 secs
#> 62       resolved       <NA>   <NA> 2026-01-14 19:51:27 3.09978437 secs
#> 63       resolved       <NA>   <NA> 2026-01-14 19:51:27 3.12329507 secs
#> 64       resolved       <NA>   <NA> 2026-01-14 19:51:27 3.14665842 secs
#> 65       resolved       <NA>   <NA> 2026-01-14 19:51:27 3.17024422 secs
#> 66       resolved       <NA>   <NA> 2026-01-14 19:51:27 3.19375062 secs
#> 67       resolved       <NA>   <NA> 2026-01-14 19:51:27 3.21788096 secs
#> 68       resolved       <NA>   <NA> 2026-01-14 19:51:27 3.24142861 secs
#> 69       resolved       <NA>   <NA> 2026-01-14 19:51:27 3.26496863 secs
#> 70       resolved       <NA>   <NA> 2026-01-14 19:51:27 3.28856897 secs
#> 71       resolved       <NA>   <NA> 2026-01-14 19:51:27 3.31209350 secs
#> 72       resolved       <NA>   <NA> 2026-01-14 19:51:27 3.33598399 secs
#> 73       resolved       <NA>   <NA> 2026-01-14 19:51:27 3.35950661 secs
#> 74       resolved       <NA>   <NA> 2026-01-14 19:51:27 3.38313484 secs
#> 75       resolved       <NA>   <NA> 2026-01-14 19:51:27 3.40655160 secs
#> 76       resolved       <NA>   <NA> 2026-01-14 19:51:27 3.43022442 secs
#> 77       resolved       <NA>   <NA> 2026-01-14 19:51:27 3.45376134 secs
#> 78       resolved       <NA>   <NA> 2026-01-14 19:51:27 3.47752690 secs
#> 79       resolved       <NA>   <NA> 2026-01-14 19:51:27 3.50126672 secs
#> 80       resolved       <NA>   <NA> 2026-01-14 19:51:27 3.52507782 secs
#> 81       resolved       <NA>   <NA> 2026-01-14 19:51:27 3.54857564 secs
#> 82       resolved       <NA>   <NA> 2026-01-14 19:51:27 3.57217360 secs
#> 83       resolved       <NA>   <NA> 2026-01-14 19:51:27 3.59587383 secs
#> 84       resolved       <NA>   <NA> 2026-01-14 19:51:27 3.61937690 secs
#> 85       resolved       <NA>   <NA> 2026-01-14 19:51:27 3.64265823 secs
#> 86       resolved       <NA>   <NA> 2026-01-14 19:51:27 3.66597199 secs
#> 87       resolved       <NA>   <NA> 2026-01-14 19:51:27 3.68952203 secs
#> 88       resolved       <NA>   <NA> 2026-01-14 19:51:27 3.71300244 secs
#> 89       resolved       <NA>   <NA> 2026-01-14 19:51:27 3.73645735 secs
#> 90       resolved       <NA>   <NA> 2026-01-14 19:51:27 3.75992084 secs
#> 91       resolved       <NA>   <NA> 2026-01-14 19:51:27 3.78355575 secs
#> 92       resolved       <NA>   <NA> 2026-01-14 19:51:27 3.80702305 secs
#> 93       resolved       <NA>   <NA> 2026-01-14 19:51:27 3.83048391 secs
#> 94       resolved       <NA>   <NA> 2026-01-14 19:51:27 3.85372424 secs
#> 95       resolved       <NA>   <NA> 2026-01-14 19:51:27 3.87691855 secs
#> 96       resolved       <NA>   <NA> 2026-01-14 19:51:27 3.90043330 secs
#> 97       resolved       <NA>   <NA> 2026-01-14 19:51:27 3.92410374 secs
#> 98       resolved       <NA>   <NA> 2026-01-14 19:51:27 3.94784880 secs
#> 99       resolved       <NA>   <NA> 2026-01-14 19:51:27 3.97135472 secs
#> 100      resolved       <NA>   <NA> 2026-01-14 19:51:27 3.99482131 secs
#> 101      resolved       <NA>   <NA> 2026-01-14 19:51:27 4.01835322 secs
#> 102      resolved       <NA>   <NA> 2026-01-14 19:51:27 4.04209208 secs
#> 103      resolved       <NA>   <NA> 2026-01-14 19:51:28 4.06577325 secs
#> 104      resolved       <NA>   <NA> 2026-01-14 19:51:28 4.08929205 secs
#> 105      resolved       <NA>   <NA> 2026-01-14 19:51:28 4.11275959 secs
#> 106      resolved       <NA>   <NA> 2026-01-14 19:51:28 4.13634491 secs
#> 107      resolved       <NA>   <NA> 2026-01-14 19:51:28 4.16013074 secs
#> 108      resolved       <NA>   <NA> 2026-01-14 19:51:28 4.18381763 secs
#> 109      resolved       <NA>   <NA> 2026-01-14 19:51:28 4.20785069 secs
#> 110      resolved       <NA>   <NA> 2026-01-14 19:51:28 4.23163080 secs
#> 111      resolved       <NA>   <NA> 2026-01-14 19:51:28 4.25528288 secs
#> 112      resolved       <NA>   <NA> 2026-01-14 19:51:28 4.27892804 secs
#> 113      resolved       <NA>   <NA> 2026-01-14 19:51:28 4.30268049 secs
#> 114      resolved       <NA>   <NA> 2026-01-14 19:51:28 4.32620764 secs
#> 115      resolved       <NA>   <NA> 2026-01-14 19:51:28 4.34959984 secs
#> 116      resolved       <NA>   <NA> 2026-01-14 19:51:28 4.37312341 secs
#> 117      resolved       <NA>   <NA> 2026-01-14 19:51:28 4.39659977 secs
#> 121      resolved       <NA>   <NA> 2026-01-14 19:51:28 4.42011452 secs
#> 118 receiveResult   overhead gather 2026-01-14 19:51:28 4.42143059 secs
#> 119        gather   overhead   <NA> 2026-01-14 19:51:28 4.42367697 secs
#> 122       resolve   overhead   <NA> 2026-01-14 19:51:28 4.43108368 secs
#>              duration future_label                         future_uuid
#> 1   1.582623e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 4   1.344097e+00 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 2   1.333214e+00 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 3   9.860992e-04 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 5   1.178265e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 120 3.009451e+00 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 6   1.125860e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 7   1.127696e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 8   1.136065e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 9   1.184702e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 10  1.163244e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 11  1.108074e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 12  1.148796e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 13  1.134372e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 14  1.124835e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 15  1.140046e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 16  1.139688e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 17  1.163936e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 18  1.151395e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 19  1.148820e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 20  1.146340e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 21  1.140380e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 22  1.148272e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 23  1.145387e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 24  1.151299e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 25  1.141977e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 26  1.150632e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 27  1.138067e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 28  1.142192e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 29  1.144481e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 30  1.143456e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 31  1.149845e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 32  1.198769e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 33  1.125956e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 34  1.869798e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 35  1.170850e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 36  1.157856e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 37  1.165748e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 38  1.180553e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 39  1.169586e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 40  1.180100e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 41  1.156235e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 42  1.154661e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 43  1.163626e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 44  1.161122e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 45  1.156569e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 46  1.171947e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 47  1.169157e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 48  1.172876e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 49  1.150727e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 50  1.160359e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 51  1.173091e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 52  1.168942e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 53  1.171446e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 54  1.179719e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 55  1.170635e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 56  1.182747e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 57  1.164579e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 58  1.171327e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 59  1.166081e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 60  1.162410e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 61  1.150322e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 62  1.162338e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 63  1.155925e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 64  1.150036e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 65  1.169109e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 66  1.169014e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 67  1.165342e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 68  1.172662e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 69  1.180053e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 70  1.163507e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 71  1.192522e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 72  1.162004e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 73  1.170611e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 74  1.163363e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 75  1.160717e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 76  1.151824e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 77  1.172566e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 78  1.170254e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 79  1.164317e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 80  1.144361e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 81  1.170707e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 82  1.166916e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 83  1.163435e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 84  1.143694e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 85  1.156664e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 86  1.156116e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 87  1.163983e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 88  1.158118e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 89  1.149917e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 90  1.174331e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 91  1.150608e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 92  1.161313e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 93  1.146364e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 94  1.145577e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 95  1.160002e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 96  1.167083e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 97  1.170635e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 98  1.161790e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 99  1.153541e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 100 1.160669e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 101 1.170278e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 102 1.161861e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 103 1.155186e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 104 1.156950e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 105 1.161575e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 106 1.167941e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 107 1.167655e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 108 1.160288e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 109 1.170421e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 110 1.161146e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 111 1.156163e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 112 1.171207e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 113 1.164865e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 114 1.149154e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 115 1.160264e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 116 1.156688e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 117 1.153517e-02 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 121 8.545876e-03 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 118 3.273487e-04 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 119 1.461506e-04 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#> 122 3.767014e-05 secs         <NA> 9388f6f894218a6b8e5b343286474110-19
#>                         session_uuid
#> 1   9388f6f894218a6b8e5b343286474110
#> 4   9388f6f894218a6b8e5b343286474110
#> 2   9388f6f894218a6b8e5b343286474110
#> 3   9388f6f894218a6b8e5b343286474110
#> 5   9388f6f894218a6b8e5b343286474110
#> 120 9388f6f894218a6b8e5b343286474110
#> 6   9388f6f894218a6b8e5b343286474110
#> 7   9388f6f894218a6b8e5b343286474110
#> 8   9388f6f894218a6b8e5b343286474110
#> 9   9388f6f894218a6b8e5b343286474110
#> 10  9388f6f894218a6b8e5b343286474110
#> 11  9388f6f894218a6b8e5b343286474110
#> 12  9388f6f894218a6b8e5b343286474110
#> 13  9388f6f894218a6b8e5b343286474110
#> 14  9388f6f894218a6b8e5b343286474110
#> 15  9388f6f894218a6b8e5b343286474110
#> 16  9388f6f894218a6b8e5b343286474110
#> 17  9388f6f894218a6b8e5b343286474110
#> 18  9388f6f894218a6b8e5b343286474110
#> 19  9388f6f894218a6b8e5b343286474110
#> 20  9388f6f894218a6b8e5b343286474110
#> 21  9388f6f894218a6b8e5b343286474110
#> 22  9388f6f894218a6b8e5b343286474110
#> 23  9388f6f894218a6b8e5b343286474110
#> 24  9388f6f894218a6b8e5b343286474110
#> 25  9388f6f894218a6b8e5b343286474110
#> 26  9388f6f894218a6b8e5b343286474110
#> 27  9388f6f894218a6b8e5b343286474110
#> 28  9388f6f894218a6b8e5b343286474110
#> 29  9388f6f894218a6b8e5b343286474110
#> 30  9388f6f894218a6b8e5b343286474110
#> 31  9388f6f894218a6b8e5b343286474110
#> 32  9388f6f894218a6b8e5b343286474110
#> 33  9388f6f894218a6b8e5b343286474110
#> 34  9388f6f894218a6b8e5b343286474110
#> 35  9388f6f894218a6b8e5b343286474110
#> 36  9388f6f894218a6b8e5b343286474110
#> 37  9388f6f894218a6b8e5b343286474110
#> 38  9388f6f894218a6b8e5b343286474110
#> 39  9388f6f894218a6b8e5b343286474110
#> 40  9388f6f894218a6b8e5b343286474110
#> 41  9388f6f894218a6b8e5b343286474110
#> 42  9388f6f894218a6b8e5b343286474110
#> 43  9388f6f894218a6b8e5b343286474110
#> 44  9388f6f894218a6b8e5b343286474110
#> 45  9388f6f894218a6b8e5b343286474110
#> 46  9388f6f894218a6b8e5b343286474110
#> 47  9388f6f894218a6b8e5b343286474110
#> 48  9388f6f894218a6b8e5b343286474110
#> 49  9388f6f894218a6b8e5b343286474110
#> 50  9388f6f894218a6b8e5b343286474110
#> 51  9388f6f894218a6b8e5b343286474110
#> 52  9388f6f894218a6b8e5b343286474110
#> 53  9388f6f894218a6b8e5b343286474110
#> 54  9388f6f894218a6b8e5b343286474110
#> 55  9388f6f894218a6b8e5b343286474110
#> 56  9388f6f894218a6b8e5b343286474110
#> 57  9388f6f894218a6b8e5b343286474110
#> 58  9388f6f894218a6b8e5b343286474110
#> 59  9388f6f894218a6b8e5b343286474110
#> 60  9388f6f894218a6b8e5b343286474110
#> 61  9388f6f894218a6b8e5b343286474110
#> 62  9388f6f894218a6b8e5b343286474110
#> 63  9388f6f894218a6b8e5b343286474110
#> 64  9388f6f894218a6b8e5b343286474110
#> 65  9388f6f894218a6b8e5b343286474110
#> 66  9388f6f894218a6b8e5b343286474110
#> 67  9388f6f894218a6b8e5b343286474110
#> 68  9388f6f894218a6b8e5b343286474110
#> 69  9388f6f894218a6b8e5b343286474110
#> 70  9388f6f894218a6b8e5b343286474110
#> 71  9388f6f894218a6b8e5b343286474110
#> 72  9388f6f894218a6b8e5b343286474110
#> 73  9388f6f894218a6b8e5b343286474110
#> 74  9388f6f894218a6b8e5b343286474110
#> 75  9388f6f894218a6b8e5b343286474110
#> 76  9388f6f894218a6b8e5b343286474110
#> 77  9388f6f894218a6b8e5b343286474110
#> 78  9388f6f894218a6b8e5b343286474110
#> 79  9388f6f894218a6b8e5b343286474110
#> 80  9388f6f894218a6b8e5b343286474110
#> 81  9388f6f894218a6b8e5b343286474110
#> 82  9388f6f894218a6b8e5b343286474110
#> 83  9388f6f894218a6b8e5b343286474110
#> 84  9388f6f894218a6b8e5b343286474110
#> 85  9388f6f894218a6b8e5b343286474110
#> 86  9388f6f894218a6b8e5b343286474110
#> 87  9388f6f894218a6b8e5b343286474110
#> 88  9388f6f894218a6b8e5b343286474110
#> 89  9388f6f894218a6b8e5b343286474110
#> 90  9388f6f894218a6b8e5b343286474110
#> 91  9388f6f894218a6b8e5b343286474110
#> 92  9388f6f894218a6b8e5b343286474110
#> 93  9388f6f894218a6b8e5b343286474110
#> 94  9388f6f894218a6b8e5b343286474110
#> 95  9388f6f894218a6b8e5b343286474110
#> 96  9388f6f894218a6b8e5b343286474110
#> 97  9388f6f894218a6b8e5b343286474110
#> 98  9388f6f894218a6b8e5b343286474110
#> 99  9388f6f894218a6b8e5b343286474110
#> 100 9388f6f894218a6b8e5b343286474110
#> 101 9388f6f894218a6b8e5b343286474110
#> 102 9388f6f894218a6b8e5b343286474110
#> 103 9388f6f894218a6b8e5b343286474110
#> 104 9388f6f894218a6b8e5b343286474110
#> 105 9388f6f894218a6b8e5b343286474110
#> 106 9388f6f894218a6b8e5b343286474110
#> 107 9388f6f894218a6b8e5b343286474110
#> 108 9388f6f894218a6b8e5b343286474110
#> 109 9388f6f894218a6b8e5b343286474110
#> 110 9388f6f894218a6b8e5b343286474110
#> 111 9388f6f894218a6b8e5b343286474110
#> 112 9388f6f894218a6b8e5b343286474110
#> 113 9388f6f894218a6b8e5b343286474110
#> 114 9388f6f894218a6b8e5b343286474110
#> 115 9388f6f894218a6b8e5b343286474110
#> 116 9388f6f894218a6b8e5b343286474110
#> 117 9388f6f894218a6b8e5b343286474110
#> 121 9388f6f894218a6b8e5b343286474110
#> 118 9388f6f894218a6b8e5b343286474110
#> 119 9388f6f894218a6b8e5b343286474110
#> 122 9388f6f894218a6b8e5b343286474110
#> 

## Stop parallel workers and disable journal logging and signaling
plan(sequential)
options(oopts)