Gets the logged journal of events for a future

WARNING: This function is under development. It can change at any time. For now, please, do not depend on this function in a published R package.

journal(x, ...)

Arguments

x

A Future object.

...

Not used.

Value

A data frame of class FutureJournal with columns:

event (character string) - type of event that took place

category (character string) - the category of the event

parent (character string) - (to be describe)

start (POSIXct) - the timestamp when the event started

at (difftime) - the time when the event started relative to first event

duration (difftime) - the duration of the event

future_label (character string) - the label of the future

future_uuid (character string) - the UUID of the future

session_uuid (character string) - the UUID of the R session where the event took place

The common events are:

create - the future was created (an overhead)

launch - the future was launched (an overhead)

evaluate - the future was evaluated (an evaluation)

resolved - the future was queried (may be occur multiple times) (an overhead)

gather - the results was retrieved (an overhead)

but others may be added by other Future classes.

Common event categories are:

evaluation - processing time is spent on evaluation

overhead - processing time is spent on orchestrating the future

waiting - processing time is spent on waiting to set up or querying the future

but others may be added by other Future classes.

The data frame is sorted by the at time. Note that the timestamps for the evaluate event are based on the local time on the worker. The system clocks on the worker and the calling R system may not be in perfect sync.

Enabling and disabling event logging

To enable logging of events, set option future.journal is TRUE. To disable, set it to FALSE (default).

See also

Examples

## Enable journaling of futures
oopts <- options(future.journal = TRUE)

plan(multisession, workers = 2L)

t_start <- Sys.time()
fs <- lapply(1:3, FUN = function(x) future({ Sys.sleep(x); sqrt(x) }))
vs <- value(fs)
js <- lapply(fs, FUN = journal, baseline = t_start)
print(js)
#> [[1]]
#>            event   category parent               start              at
#> 1         create   overhead   <NA> 2023-07-31 15:02:28 0.00000000 secs
#> 6         launch   overhead   <NA> 2023-07-31 15:02:28 0.01189375 secs
#> 2      getWorker   overhead launch 2023-07-31 15:02:28 0.01233602 secs
#> 3    eraseWorker   overhead launch 2023-07-31 15:02:28 0.01284480 secs
#> 4 attachPackages   overhead launch 2023-07-31 15:02:28 0.01367402 secs
#> 5  exportGlobals   overhead launch 2023-07-31 15:02:28 0.01406813 secs
#> 9       evaluate evaluation   <NA> 2023-07-31 15:02:28 0.05918574 secs
#> 7  receiveResult   overhead gather 2023-07-31 15:02:29 1.06237626 secs
#> 8         gather   overhead   <NA> 2023-07-31 15:02:29 1.06414104 secs
#>            duration future_label                          future_uuid
#> 1 3.189564e-03 secs         <NA> 0dae9bbd-e8aa-ec15-3516-44cebb55af50
#> 6 3.320217e-03 secs         <NA> 0dae9bbd-e8aa-ec15-3516-44cebb55af50
#> 2 9.679794e-05 secs         <NA> 0dae9bbd-e8aa-ec15-3516-44cebb55af50
#> 3 4.096031e-04 secs         <NA> 0dae9bbd-e8aa-ec15-3516-44cebb55af50
#> 4 2.861023e-05 secs         <NA> 0dae9bbd-e8aa-ec15-3516-44cebb55af50
#> 5 3.354549e-04 secs         <NA> 0dae9bbd-e8aa-ec15-3516-44cebb55af50
#> 9 1.002677e+00 secs         <NA> 0dae9bbd-e8aa-ec15-3516-44cebb55af50
#> 7 2.377033e-04 secs         <NA> 0dae9bbd-e8aa-ec15-3516-44cebb55af50
#> 8 6.651878e-05 secs         <NA> 0dae9bbd-e8aa-ec15-3516-44cebb55af50
#>                           session_uuid
#> 1 10e01b80-cd51-6252-e464-346f9de48b4c
#> 6 10e01b80-cd51-6252-e464-346f9de48b4c
#> 2 10e01b80-cd51-6252-e464-346f9de48b4c
#> 3 10e01b80-cd51-6252-e464-346f9de48b4c
#> 4 10e01b80-cd51-6252-e464-346f9de48b4c
#> 5 10e01b80-cd51-6252-e464-346f9de48b4c
#> 9 10e01b80-cd51-6252-e464-346f9de48b4c
#> 7 10e01b80-cd51-6252-e464-346f9de48b4c
#> 8 10e01b80-cd51-6252-e464-346f9de48b4c
#> 
#> [[2]]
#>            event   category parent               start              at
#> 1         create   overhead   <NA> 2023-07-31 15:02:28 0.00000000 secs
#> 6         launch   overhead   <NA> 2023-07-31 15:02:28 0.01007032 secs
#> 2      getWorker   overhead launch 2023-07-31 15:02:28 0.01048231 secs
#> 3    eraseWorker   overhead launch 2023-07-31 15:02:28 0.01096845 secs
#> 4 attachPackages   overhead launch 2023-07-31 15:02:28 0.03865051 secs
#> 5  exportGlobals   overhead launch 2023-07-31 15:02:28 0.03940964 secs
#> 9       evaluate evaluation   <NA> 2023-07-31 15:02:28 0.08401108 secs
#> 7  receiveResult   overhead gather 2023-07-31 15:02:30 2.14306188 secs
#> 8         gather   overhead   <NA> 2023-07-31 15:02:30 2.14472318 secs
#>            duration future_label                          future_uuid
#> 1 1.556158e-03 secs         <NA> baf3c5c8-9594-b86a-dfc0-7466c43578bc
#> 6 3.095579e-02 secs         <NA> baf3c5c8-9594-b86a-dfc0-7466c43578bc
#> 2 8.845329e-05 secs         <NA> baf3c5c8-9594-b86a-dfc0-7466c43578bc
#> 3 2.690768e-02 secs         <NA> baf3c5c8-9594-b86a-dfc0-7466c43578bc
#> 4 6.341934e-05 secs         <NA> baf3c5c8-9594-b86a-dfc0-7466c43578bc
#> 5 3.988743e-04 secs         <NA> baf3c5c8-9594-b86a-dfc0-7466c43578bc
#> 9 2.042913e+00 secs         <NA> baf3c5c8-9594-b86a-dfc0-7466c43578bc
#> 7 2.152920e-04 secs         <NA> baf3c5c8-9594-b86a-dfc0-7466c43578bc
#> 8 6.508827e-05 secs         <NA> baf3c5c8-9594-b86a-dfc0-7466c43578bc
#>                           session_uuid
#> 1 10e01b80-cd51-6252-e464-346f9de48b4c
#> 6 10e01b80-cd51-6252-e464-346f9de48b4c
#> 2 10e01b80-cd51-6252-e464-346f9de48b4c
#> 3 10e01b80-cd51-6252-e464-346f9de48b4c
#> 4 10e01b80-cd51-6252-e464-346f9de48b4c
#> 5 10e01b80-cd51-6252-e464-346f9de48b4c
#> 9 10e01b80-cd51-6252-e464-346f9de48b4c
#> 7 10e01b80-cd51-6252-e464-346f9de48b4c
#> 8 10e01b80-cd51-6252-e464-346f9de48b4c
#> 
#> [[3]]
#>            event   category parent               start              at
#> 1         create   overhead   <NA> 2023-07-31 15:02:28 0.00000000 secs
#> 6         launch   overhead   <NA> 2023-07-31 15:02:28 0.01103950 secs
#> 2      getWorker   overhead launch 2023-07-31 15:02:28 0.01148462 secs
#> 3    eraseWorker   overhead launch 2023-07-31 15:02:29 1.01271534 secs
#> 4 attachPackages   overhead launch 2023-07-31 15:02:29 1.01476693 secs
#> 5  exportGlobals   overhead launch 2023-07-31 15:02:29 1.01605678 secs
#> 9       evaluate evaluation   <NA> 2023-07-31 15:02:29 1.06245065 secs
#> 7  receiveResult   overhead gather 2023-07-31 15:02:32 4.06788611 secs
#> 8         gather   overhead   <NA> 2023-07-31 15:02:32 4.06967878 secs
#>            duration future_label                          future_uuid
#> 1 1.952410e-03 secs         <NA> f8d802ee-bd69-4d7e-e617-d61c6d85149b
#> 6 1.007492e+00 secs         <NA> f8d802ee-bd69-4d7e-e617-d61c6d85149b
#> 2 1.000095e+00 secs         <NA> f8d802ee-bd69-4d7e-e617-d61c6d85149b
#> 3 8.327961e-04 secs         <NA> f8d802ee-bd69-4d7e-e617-d61c6d85149b
#> 4 7.867813e-05 secs         <NA> f8d802ee-bd69-4d7e-e617-d61c6d85149b
#> 5 6.575584e-04 secs         <NA> f8d802ee-bd69-4d7e-e617-d61c6d85149b
#> 9 3.004953e+00 secs         <NA> f8d802ee-bd69-4d7e-e617-d61c6d85149b
#> 7 2.441406e-04 secs         <NA> f8d802ee-bd69-4d7e-e617-d61c6d85149b
#> 8 7.224083e-05 secs         <NA> f8d802ee-bd69-4d7e-e617-d61c6d85149b
#>                           session_uuid
#> 1 10e01b80-cd51-6252-e464-346f9de48b4c
#> 6 10e01b80-cd51-6252-e464-346f9de48b4c
#> 2 10e01b80-cd51-6252-e464-346f9de48b4c
#> 3 10e01b80-cd51-6252-e464-346f9de48b4c
#> 4 10e01b80-cd51-6252-e464-346f9de48b4c
#> 5 10e01b80-cd51-6252-e464-346f9de48b4c
#> 9 10e01b80-cd51-6252-e464-346f9de48b4c
#> 7 10e01b80-cd51-6252-e464-346f9de48b4c
#> 8 10e01b80-cd51-6252-e464-346f9de48b4c
#> 

## Stop parallel workers and disable journal logging and signaling
plan(sequential)
options(oopts)