Gets the logged journal of events for a future

WARNING: This function is under development. It can change at any time. For now, please, do not depend on this function in a published R package.

Usage

journal(x, ...)

Arguments

x

A future::Future object.

...

Not used.

Value

A data frame of class FutureJournal with columns:

event (character string) - type of event that took place

category (character string) - the category of the event

parent (character string) - (to be describe)

start (POSIXct) - the timestamp when the event started

at (difftime) - the time when the event started relative to first event

duration (difftime) - the duration of the event

future_label (character string) - the label of the future

future_uuid (character string) - the UUID of the future

session_uuid (character string) - the UUID of the R session where the event took place

The common events are:

create - the future was created (an overhead)

launch - the future was launched (an overhead)

evaluate - the future was evaluated (an evaluation)

resolved - the future was queried (may be occur multiple times) (an overhead)

gather - the results was retrieved (an overhead)

but others may be added by other Future classes.

Common event categories are:

evaluation - processing time is spent on evaluation

overhead - processing time is spent on orchestrating the future

waiting - processing time is spent on waiting to set up or querying the future

but others may be added by other Future classes.

The data frame is sorted by the at time. Note that the timestamps for the evaluate event are based on the local time on the worker. The system clocks on the worker and the calling R system may not be in perfect sync.

Enabling and disabling event logging

To enable logging of events, set option future.journal is TRUE. To disable, set it to FALSE (default).

See also

Examples

## Enable journaling of futures
oopts <- options(future.journal = TRUE)

plan(multisession, workers = 2L)

t_start <- Sys.time()
fs <- lapply(1:3, FUN = function(x) future({ Sys.sleep(x); sqrt(x) }))
vs <- value(fs)
js <- lapply(fs, FUN = journal, baseline = t_start)
print(js)
#> [[1]]
#>            event   category parent               start              at
#> 1         create   overhead   <NA> 2025-10-31 14:05:57 0.00000000 secs
#> 4         launch   overhead   <NA> 2025-10-31 14:05:57 0.01029587 secs
#> 2      getWorker   overhead launch 2025-10-31 14:05:57 0.01055384 secs
#> 3    eraseWorker   overhead launch 2025-10-31 14:05:57 0.01208162 secs
#> 5       resolved       <NA>   <NA> 2025-10-31 14:05:57 0.06345868 secs
#> 6       resolved       <NA>   <NA> 2025-10-31 14:05:57 0.09775162 secs
#> 7       resolved       <NA>   <NA> 2025-10-31 14:05:57 0.13193083 secs
#> 8       resolved       <NA>   <NA> 2025-10-31 14:05:57 0.16630578 secs
#> 9       resolved       <NA>   <NA> 2025-10-31 14:05:57 0.20131850 secs
#> 10      resolved       <NA>   <NA> 2025-10-31 14:05:57 0.23664379 secs
#> 11      resolved       <NA>   <NA> 2025-10-31 14:05:57 0.27100778 secs
#> 44      evaluate evaluation   <NA> 2025-10-31 14:05:57 0.28682089 secs
#> 12      resolved       <NA>   <NA> 2025-10-31 14:05:57 0.30590129 secs
#> 13      resolved       <NA>   <NA> 2025-10-31 14:05:57 0.34939241 secs
#> 14      resolved       <NA>   <NA> 2025-10-31 14:05:57 0.38502765 secs
#> 15      resolved       <NA>   <NA> 2025-10-31 14:05:57 0.42125511 secs
#> 16      resolved       <NA>   <NA> 2025-10-31 14:05:57 0.45713091 secs
#> 17      resolved       <NA>   <NA> 2025-10-31 14:05:57 0.49272871 secs
#> 18      resolved       <NA>   <NA> 2025-10-31 14:05:58 0.52868462 secs
#> 19      resolved       <NA>   <NA> 2025-10-31 14:05:58 0.56499553 secs
#> 20      resolved       <NA>   <NA> 2025-10-31 14:05:58 0.60076165 secs
#> 21      resolved       <NA>   <NA> 2025-10-31 14:05:58 0.63705444 secs
#> 22      resolved       <NA>   <NA> 2025-10-31 14:05:58 0.67517328 secs
#> 23      resolved       <NA>   <NA> 2025-10-31 14:05:58 0.71332216 secs
#> 24      resolved       <NA>   <NA> 2025-10-31 14:05:58 0.75109243 secs
#> 25      resolved       <NA>   <NA> 2025-10-31 14:05:58 0.78751898 secs
#> 26      resolved       <NA>   <NA> 2025-10-31 14:05:58 0.82420182 secs
#> 27      resolved       <NA>   <NA> 2025-10-31 14:05:58 0.86039376 secs
#> 28      resolved       <NA>   <NA> 2025-10-31 14:05:58 0.89719081 secs
#> 29      resolved       <NA>   <NA> 2025-10-31 14:05:58 0.93469238 secs
#> 30      resolved       <NA>   <NA> 2025-10-31 14:05:58 0.97455072 secs
#> 31      resolved       <NA>   <NA> 2025-10-31 14:05:58 1.01219320 secs
#> 32      resolved       <NA>   <NA> 2025-10-31 14:05:58 1.05035639 secs
#> 33      resolved       <NA>   <NA> 2025-10-31 14:05:58 1.08909082 secs
#> 34      resolved       <NA>   <NA> 2025-10-31 14:05:58 1.12751603 secs
#> 35      resolved       <NA>   <NA> 2025-10-31 14:05:58 1.16583586 secs
#> 36      resolved       <NA>   <NA> 2025-10-31 14:05:58 1.20567775 secs
#> 37      resolved       <NA>   <NA> 2025-10-31 14:05:58 1.24484348 secs
#> 38      resolved       <NA>   <NA> 2025-10-31 14:05:58 1.28503227 secs
#> 39      resolved       <NA>   <NA> 2025-10-31 14:05:58 1.32468390 secs
#> 40      resolved       <NA>   <NA> 2025-10-31 14:05:58 1.36518335 secs
#> 41      resolved       <NA>   <NA> 2025-10-31 14:05:58 1.41064358 secs
#> 45      resolved       <NA>   <NA> 2025-10-31 14:05:58 1.45771933 secs
#> 42 receiveResult   overhead gather 2025-10-31 14:05:58 1.46739411 secs
#> 43        gather   overhead   <NA> 2025-10-31 14:05:58 1.47129297 secs
#> 46      resolved       <NA>   <NA> 2025-10-31 14:05:58 1.49809527 secs
#> 47       resolve   overhead   <NA> 2025-10-31 14:05:58 1.50048470 secs
#>             duration future_label                         future_uuid
#> 1  9.684563e-03 secs         <NA> 2107504b07336839b4018835cafa74f4-17
#> 4  7.871628e-03 secs         <NA> 2107504b07336839b4018835cafa74f4-17
#> 2  7.557869e-04 secs         <NA> 2107504b07336839b4018835cafa74f4-17
#> 3  5.278587e-04 secs         <NA> 2107504b07336839b4018835cafa74f4-17
#> 5  1.103926e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-17
#> 6  1.103449e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-17
#> 7  1.087379e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-17
#> 8  1.085496e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-17
#> 9  1.115727e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-17
#> 10 1.083755e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-17
#> 11 1.095247e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-17
#> 44 1.176986e+00 secs         <NA> 2107504b07336839b4018835cafa74f4-17
#> 12 1.151752e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-17
#> 13 1.103306e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-17
#> 14 1.106882e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-17
#> 15 1.113319e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-17
#> 16 1.108265e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-17
#> 17 1.115465e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-17
#> 18 1.100397e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-17
#> 19 1.098752e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-17
#> 20 1.086164e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-17
#> 21 1.136732e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-17
#> 22 1.139164e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-17
#> 23 1.139402e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-17
#> 24 1.100016e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-17
#> 25 1.102567e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-17
#> 26 1.095247e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-17
#> 27 1.085424e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-17
#> 28 1.119876e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-17
#> 29 1.115727e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-17
#> 30 1.119494e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-17
#> 31 1.101017e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-17
#> 32 1.123691e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-17
#> 33 1.153374e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-17
#> 34 1.119328e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-17
#> 35 1.109886e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-17
#> 36 1.175046e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-17
#> 37 1.120305e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-17
#> 38 1.102471e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-17
#> 39 1.185393e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-17
#> 40 1.124692e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-17
#> 41 1.260996e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-17
#> 45 2.153134e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-17
#> 42 5.803108e-04 secs         <NA> 2107504b07336839b4018835cafa74f4-17
#> 43 3.058910e-04 secs         <NA> 2107504b07336839b4018835cafa74f4-17
#> 46 9.655952e-05 secs         <NA> 2107504b07336839b4018835cafa74f4-17
#> 47 3.170967e-05 secs         <NA> 2107504b07336839b4018835cafa74f4-17
#>                        session_uuid
#> 1  2107504b07336839b4018835cafa74f4
#> 4  2107504b07336839b4018835cafa74f4
#> 2  2107504b07336839b4018835cafa74f4
#> 3  2107504b07336839b4018835cafa74f4
#> 5  2107504b07336839b4018835cafa74f4
#> 6  2107504b07336839b4018835cafa74f4
#> 7  2107504b07336839b4018835cafa74f4
#> 8  2107504b07336839b4018835cafa74f4
#> 9  2107504b07336839b4018835cafa74f4
#> 10 2107504b07336839b4018835cafa74f4
#> 11 2107504b07336839b4018835cafa74f4
#> 44 2107504b07336839b4018835cafa74f4
#> 12 2107504b07336839b4018835cafa74f4
#> 13 2107504b07336839b4018835cafa74f4
#> 14 2107504b07336839b4018835cafa74f4
#> 15 2107504b07336839b4018835cafa74f4
#> 16 2107504b07336839b4018835cafa74f4
#> 17 2107504b07336839b4018835cafa74f4
#> 18 2107504b07336839b4018835cafa74f4
#> 19 2107504b07336839b4018835cafa74f4
#> 20 2107504b07336839b4018835cafa74f4
#> 21 2107504b07336839b4018835cafa74f4
#> 22 2107504b07336839b4018835cafa74f4
#> 23 2107504b07336839b4018835cafa74f4
#> 24 2107504b07336839b4018835cafa74f4
#> 25 2107504b07336839b4018835cafa74f4
#> 26 2107504b07336839b4018835cafa74f4
#> 27 2107504b07336839b4018835cafa74f4
#> 28 2107504b07336839b4018835cafa74f4
#> 29 2107504b07336839b4018835cafa74f4
#> 30 2107504b07336839b4018835cafa74f4
#> 31 2107504b07336839b4018835cafa74f4
#> 32 2107504b07336839b4018835cafa74f4
#> 33 2107504b07336839b4018835cafa74f4
#> 34 2107504b07336839b4018835cafa74f4
#> 35 2107504b07336839b4018835cafa74f4
#> 36 2107504b07336839b4018835cafa74f4
#> 37 2107504b07336839b4018835cafa74f4
#> 38 2107504b07336839b4018835cafa74f4
#> 39 2107504b07336839b4018835cafa74f4
#> 40 2107504b07336839b4018835cafa74f4
#> 41 2107504b07336839b4018835cafa74f4
#> 45 2107504b07336839b4018835cafa74f4
#> 42 2107504b07336839b4018835cafa74f4
#> 43 2107504b07336839b4018835cafa74f4
#> 46 2107504b07336839b4018835cafa74f4
#> 47 2107504b07336839b4018835cafa74f4
#> 
#> [[2]]
#>            event   category parent               start              at
#> 1         create   overhead   <NA> 2025-10-31 14:05:57 0.00000000 secs
#> 4         launch   overhead   <NA> 2025-10-31 14:05:57 0.01272726 secs
#> 2      getWorker   overhead launch 2025-10-31 14:05:57 0.01295114 secs
#> 3    eraseWorker   overhead launch 2025-10-31 14:05:57 0.01457095 secs
#> 5       resolved       <NA>   <NA> 2025-10-31 14:05:57 0.05645037 secs
#> 6       resolved       <NA>   <NA> 2025-10-31 14:05:57 0.09076929 secs
#> 7       resolved       <NA>   <NA> 2025-10-31 14:05:57 0.12473631 secs
#> 8       resolved       <NA>   <NA> 2025-10-31 14:05:57 0.15937424 secs
#> 9       resolved       <NA>   <NA> 2025-10-31 14:05:57 0.19454336 secs
#> 10      resolved       <NA>   <NA> 2025-10-31 14:05:57 0.22945261 secs
#> 11      resolved       <NA>   <NA> 2025-10-31 14:05:57 0.26395488 secs
#> 69      evaluate evaluation   <NA> 2025-10-31 14:05:57 0.29649329 secs
#> 12      resolved       <NA>   <NA> 2025-10-31 14:05:57 0.30666113 secs
#> 13      resolved       <NA>   <NA> 2025-10-31 14:05:57 0.34254265 secs
#> 14      resolved       <NA>   <NA> 2025-10-31 14:05:57 0.37824774 secs
#> 15      resolved       <NA>   <NA> 2025-10-31 14:05:57 0.41454983 secs
#> 16      resolved       <NA>   <NA> 2025-10-31 14:05:57 0.45029116 secs
#> 17      resolved       <NA>   <NA> 2025-10-31 14:05:57 0.48595071 secs
#> 18      resolved       <NA>   <NA> 2025-10-31 14:05:58 0.52215266 secs
#> 19      resolved       <NA>   <NA> 2025-10-31 14:05:58 0.55812907 secs
#> 20      resolved       <NA>   <NA> 2025-10-31 14:05:58 0.59367204 secs
#> 21      resolved       <NA>   <NA> 2025-10-31 14:05:58 0.63108468 secs
#> 22      resolved       <NA>   <NA> 2025-10-31 14:05:58 0.66918731 secs
#> 23      resolved       <NA>   <NA> 2025-10-31 14:05:58 0.70735669 secs
#> 24      resolved       <NA>   <NA> 2025-10-31 14:05:58 0.74415946 secs
#> 25      resolved       <NA>   <NA> 2025-10-31 14:05:58 0.78065658 secs
#> 26      resolved       <NA>   <NA> 2025-10-31 14:05:58 0.81713915 secs
#> 27      resolved       <NA>   <NA> 2025-10-31 14:05:58 0.85328031 secs
#> 28      resolved       <NA>   <NA> 2025-10-31 14:05:58 0.89048147 secs
#> 29      resolved       <NA>   <NA> 2025-10-31 14:05:58 0.92812920 secs
#> 30      resolved       <NA>   <NA> 2025-10-31 14:05:58 0.96787763 secs
#> 31      resolved       <NA>   <NA> 2025-10-31 14:05:58 1.00531650 secs
#> 32      resolved       <NA>   <NA> 2025-10-31 14:05:58 1.04421020 secs
#> 33      resolved       <NA>   <NA> 2025-10-31 14:05:58 1.08280301 secs
#> 34      resolved       <NA>   <NA> 2025-10-31 14:05:58 1.12084913 secs
#> 35      resolved       <NA>   <NA> 2025-10-31 14:05:58 1.15916038 secs
#> 36      resolved       <NA>   <NA> 2025-10-31 14:05:58 1.19963551 secs
#> 37      resolved       <NA>   <NA> 2025-10-31 14:05:58 1.23908639 secs
#> 38      resolved       <NA>   <NA> 2025-10-31 14:05:58 1.27819538 secs
#> 39      resolved       <NA>   <NA> 2025-10-31 14:05:58 1.31874418 secs
#> 40      resolved       <NA>   <NA> 2025-10-31 14:05:58 1.36073923 secs
#> 41      resolved       <NA>   <NA> 2025-10-31 14:05:58 1.40764332 secs
#> 42      resolved       <NA>   <NA> 2025-10-31 14:05:58 1.48250508 secs
#> 43      resolved       <NA>   <NA> 2025-10-31 14:05:59 1.51604295 secs
#> 44      resolved       <NA>   <NA> 2025-10-31 14:05:59 1.55049562 secs
#> 45      resolved       <NA>   <NA> 2025-10-31 14:05:59 1.59101343 secs
#> 46      resolved       <NA>   <NA> 2025-10-31 14:05:59 1.63433456 secs
#> 47      resolved       <NA>   <NA> 2025-10-31 14:05:59 1.67701745 secs
#> 48      resolved       <NA>   <NA> 2025-10-31 14:05:59 1.71897578 secs
#> 49      resolved       <NA>   <NA> 2025-10-31 14:05:59 1.76286936 secs
#> 50      resolved       <NA>   <NA> 2025-10-31 14:05:59 1.80492902 secs
#> 51      resolved       <NA>   <NA> 2025-10-31 14:05:59 1.84963179 secs
#> 52      resolved       <NA>   <NA> 2025-10-31 14:05:59 1.89211941 secs
#> 53      resolved       <NA>   <NA> 2025-10-31 14:05:59 1.93459940 secs
#> 54      resolved       <NA>   <NA> 2025-10-31 14:05:59 1.97819734 secs
#> 55      resolved       <NA>   <NA> 2025-10-31 14:05:59 2.02024603 secs
#> 56      resolved       <NA>   <NA> 2025-10-31 14:05:59 2.06273508 secs
#> 57      resolved       <NA>   <NA> 2025-10-31 14:05:59 2.10477638 secs
#> 58      resolved       <NA>   <NA> 2025-10-31 14:05:59 2.14152980 secs
#> 59      resolved       <NA>   <NA> 2025-10-31 14:05:59 2.18065810 secs
#> 60      resolved       <NA>   <NA> 2025-10-31 14:05:59 2.21584177 secs
#> 61      resolved       <NA>   <NA> 2025-10-31 14:05:59 2.25175118 secs
#> 62      resolved       <NA>   <NA> 2025-10-31 14:05:59 2.29276490 secs
#> 63      resolved       <NA>   <NA> 2025-10-31 14:05:59 2.32972860 secs
#> 64      resolved       <NA>   <NA> 2025-10-31 14:05:59 2.36422944 secs
#> 65      resolved       <NA>   <NA> 2025-10-31 14:05:59 2.39974165 secs
#> 66      resolved       <NA>   <NA> 2025-10-31 14:05:59 2.43499732 secs
#> 70      resolved       <NA>   <NA> 2025-10-31 14:05:59 2.48515368 secs
#> 67 receiveResult   overhead gather 2025-10-31 14:05:59 2.48963785 secs
#> 68        gather   overhead   <NA> 2025-10-31 14:05:59 2.49670482 secs
#> 71       resolve   overhead   <NA> 2025-10-31 14:06:00 2.51849389 secs
#>             duration future_label                         future_uuid
#> 1  0.0122005939 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 4  0.0079245567 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 2  0.0008194447 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 3  0.0004818439 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 5  0.0110654831 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 6  0.0107886791 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 7  0.0110144615 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 8  0.0108623505 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 9  0.0110263824 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 10 0.0108132362 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 11 0.0108551979 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 69 2.1848299503 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 12 0.0115029812 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 13 0.0110146999 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 14 0.0109879971 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 15 0.0109660625 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 16 0.0108940601 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 17 0.0110027790 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 18 0.0110433102 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 19 0.0108807087 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 20 0.0108425617 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 21 0.0113320351 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 22 0.0113024712 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 23 0.0112907887 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 24 0.0109014511 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 25 0.0109207630 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 26 0.0107507706 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 27 0.0108680725 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 28 0.0110971928 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 29 0.0110909939 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 30 0.0109899044 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 31 0.0109601021 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 32 0.0110349655 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 33 0.0110163689 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 34 0.0111846924 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 35 0.0111811161 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 36 0.0110087395 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 37 0.0116021633 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 38 0.0111017227 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 39 0.0115537643 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 40 0.0125374794 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 41 0.0125777721 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 42 0.0108931065 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 43 0.0111167431 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 44 0.0119707584 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 45 0.0128045082 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 46 0.0132336617 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 47 0.0128345490 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 48 0.0135736465 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 49 0.0127692223 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 50 0.0133521557 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 51 0.0130529404 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 52 0.0127599239 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 53 0.0135855675 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 54 0.0128390789 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 55 0.0128674507 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 56 0.0127818584 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 57 0.0116987228 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 58 0.0119769573 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 59 0.0109701157 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 60 0.0110602379 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 61 0.0140614510 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 62 0.0117242336 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 63 0.0110096931 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 64 0.0118205547 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 65 0.0113356113 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 66 0.0145692825 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 70 0.0254354477 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 67 0.0010631084 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 68 0.0003190041 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#> 71 0.0001337528 secs         <NA> 2107504b07336839b4018835cafa74f4-18
#>                        session_uuid
#> 1  2107504b07336839b4018835cafa74f4
#> 4  2107504b07336839b4018835cafa74f4
#> 2  2107504b07336839b4018835cafa74f4
#> 3  2107504b07336839b4018835cafa74f4
#> 5  2107504b07336839b4018835cafa74f4
#> 6  2107504b07336839b4018835cafa74f4
#> 7  2107504b07336839b4018835cafa74f4
#> 8  2107504b07336839b4018835cafa74f4
#> 9  2107504b07336839b4018835cafa74f4
#> 10 2107504b07336839b4018835cafa74f4
#> 11 2107504b07336839b4018835cafa74f4
#> 69 2107504b07336839b4018835cafa74f4
#> 12 2107504b07336839b4018835cafa74f4
#> 13 2107504b07336839b4018835cafa74f4
#> 14 2107504b07336839b4018835cafa74f4
#> 15 2107504b07336839b4018835cafa74f4
#> 16 2107504b07336839b4018835cafa74f4
#> 17 2107504b07336839b4018835cafa74f4
#> 18 2107504b07336839b4018835cafa74f4
#> 19 2107504b07336839b4018835cafa74f4
#> 20 2107504b07336839b4018835cafa74f4
#> 21 2107504b07336839b4018835cafa74f4
#> 22 2107504b07336839b4018835cafa74f4
#> 23 2107504b07336839b4018835cafa74f4
#> 24 2107504b07336839b4018835cafa74f4
#> 25 2107504b07336839b4018835cafa74f4
#> 26 2107504b07336839b4018835cafa74f4
#> 27 2107504b07336839b4018835cafa74f4
#> 28 2107504b07336839b4018835cafa74f4
#> 29 2107504b07336839b4018835cafa74f4
#> 30 2107504b07336839b4018835cafa74f4
#> 31 2107504b07336839b4018835cafa74f4
#> 32 2107504b07336839b4018835cafa74f4
#> 33 2107504b07336839b4018835cafa74f4
#> 34 2107504b07336839b4018835cafa74f4
#> 35 2107504b07336839b4018835cafa74f4
#> 36 2107504b07336839b4018835cafa74f4
#> 37 2107504b07336839b4018835cafa74f4
#> 38 2107504b07336839b4018835cafa74f4
#> 39 2107504b07336839b4018835cafa74f4
#> 40 2107504b07336839b4018835cafa74f4
#> 41 2107504b07336839b4018835cafa74f4
#> 42 2107504b07336839b4018835cafa74f4
#> 43 2107504b07336839b4018835cafa74f4
#> 44 2107504b07336839b4018835cafa74f4
#> 45 2107504b07336839b4018835cafa74f4
#> 46 2107504b07336839b4018835cafa74f4
#> 47 2107504b07336839b4018835cafa74f4
#> 48 2107504b07336839b4018835cafa74f4
#> 49 2107504b07336839b4018835cafa74f4
#> 50 2107504b07336839b4018835cafa74f4
#> 51 2107504b07336839b4018835cafa74f4
#> 52 2107504b07336839b4018835cafa74f4
#> 53 2107504b07336839b4018835cafa74f4
#> 54 2107504b07336839b4018835cafa74f4
#> 55 2107504b07336839b4018835cafa74f4
#> 56 2107504b07336839b4018835cafa74f4
#> 57 2107504b07336839b4018835cafa74f4
#> 58 2107504b07336839b4018835cafa74f4
#> 59 2107504b07336839b4018835cafa74f4
#> 60 2107504b07336839b4018835cafa74f4
#> 61 2107504b07336839b4018835cafa74f4
#> 62 2107504b07336839b4018835cafa74f4
#> 63 2107504b07336839b4018835cafa74f4
#> 64 2107504b07336839b4018835cafa74f4
#> 65 2107504b07336839b4018835cafa74f4
#> 66 2107504b07336839b4018835cafa74f4
#> 70 2107504b07336839b4018835cafa74f4
#> 67 2107504b07336839b4018835cafa74f4
#> 68 2107504b07336839b4018835cafa74f4
#> 71 2107504b07336839b4018835cafa74f4
#> 
#> [[3]]
#>             event   category parent               start              at
#> 1          create   overhead   <NA> 2025-10-31 14:05:57 0.00000000 secs
#> 4          launch   overhead   <NA> 2025-10-31 14:05:57 0.01230550 secs
#> 2       getWorker   overhead launch 2025-10-31 14:05:57 0.01245904 secs
#> 3     eraseWorker   overhead launch 2025-10-31 14:05:58 1.44300628 secs
#> 5        resolved       <NA>   <NA> 2025-10-31 14:05:58 1.47231412 secs
#> 114      evaluate evaluation   <NA> 2025-10-31 14:05:59 1.49540949 secs
#> 6        resolved       <NA>   <NA> 2025-10-31 14:05:59 1.50620937 secs
#> 7        resolved       <NA>   <NA> 2025-10-31 14:05:59 1.54320025 secs
#> 8        resolved       <NA>   <NA> 2025-10-31 14:05:59 1.58575273 secs
#> 9        resolved       <NA>   <NA> 2025-10-31 14:05:59 1.62889361 secs
#> 10       resolved       <NA>   <NA> 2025-10-31 14:05:59 1.67118144 secs
#> 11       resolved       <NA>   <NA> 2025-10-31 14:05:59 1.71393061 secs
#> 12       resolved       <NA>   <NA> 2025-10-31 14:05:59 1.75702238 secs
#> 13       resolved       <NA>   <NA> 2025-10-31 14:05:59 1.80053473 secs
#> 14       resolved       <NA>   <NA> 2025-10-31 14:05:59 1.84408379 secs
#> 15       resolved       <NA>   <NA> 2025-10-31 14:05:59 1.88624001 secs
#> 16       resolved       <NA>   <NA> 2025-10-31 14:05:59 1.92951798 secs
#> 17       resolved       <NA>   <NA> 2025-10-31 14:05:59 1.97238207 secs
#> 18       resolved       <NA>   <NA> 2025-10-31 14:05:59 2.01444602 secs
#> 19       resolved       <NA>   <NA> 2025-10-31 14:05:59 2.05714941 secs
#> 20       resolved       <NA>   <NA> 2025-10-31 14:05:59 2.09677124 secs
#> 21       resolved       <NA>   <NA> 2025-10-31 14:05:59 2.13490057 secs
#> 22       resolved       <NA>   <NA> 2025-10-31 14:05:59 2.17100835 secs
#> 23       resolved       <NA>   <NA> 2025-10-31 14:05:59 2.20624065 secs
#> 24       resolved       <NA>   <NA> 2025-10-31 14:05:59 2.24623919 secs
#> 25       resolved       <NA>   <NA> 2025-10-31 14:05:59 2.28449750 secs
#> 26       resolved       <NA>   <NA> 2025-10-31 14:05:59 2.32007098 secs
#> 27       resolved       <NA>   <NA> 2025-10-31 14:05:59 2.35554099 secs
#> 28       resolved       <NA>   <NA> 2025-10-31 14:05:59 2.39060020 secs
#> 29       resolved       <NA>   <NA> 2025-10-31 14:05:59 2.43331790 secs
#> 30       resolved       <NA>   <NA> 2025-10-31 14:06:00 2.50257301 secs
#> 31       resolved       <NA>   <NA> 2025-10-31 14:06:00 2.53054833 secs
#> 32       resolved       <NA>   <NA> 2025-10-31 14:06:00 2.55955577 secs
#> 33       resolved       <NA>   <NA> 2025-10-31 14:06:00 2.58242297 secs
#> 34       resolved       <NA>   <NA> 2025-10-31 14:06:00 2.60527802 secs
#> 35       resolved       <NA>   <NA> 2025-10-31 14:06:00 2.62955713 secs
#> 36       resolved       <NA>   <NA> 2025-10-31 14:06:00 2.65655994 secs
#> 37       resolved       <NA>   <NA> 2025-10-31 14:06:00 2.68333673 secs
#> 38       resolved       <NA>   <NA> 2025-10-31 14:06:00 2.70994210 secs
#> 39       resolved       <NA>   <NA> 2025-10-31 14:06:00 2.73611927 secs
#> 40       resolved       <NA>   <NA> 2025-10-31 14:06:00 2.76244521 secs
#> 41       resolved       <NA>   <NA> 2025-10-31 14:06:00 2.78885841 secs
#> 42       resolved       <NA>   <NA> 2025-10-31 14:06:00 2.81544566 secs
#> 43       resolved       <NA>   <NA> 2025-10-31 14:06:00 2.84285498 secs
#> 44       resolved       <NA>   <NA> 2025-10-31 14:06:00 2.86927056 secs
#> 45       resolved       <NA>   <NA> 2025-10-31 14:06:00 2.89570999 secs
#> 46       resolved       <NA>   <NA> 2025-10-31 14:06:00 2.92253470 secs
#> 47       resolved       <NA>   <NA> 2025-10-31 14:06:00 2.94889450 secs
#> 48       resolved       <NA>   <NA> 2025-10-31 14:06:00 2.97606468 secs
#> 49       resolved       <NA>   <NA> 2025-10-31 14:06:00 3.00258207 secs
#> 50       resolved       <NA>   <NA> 2025-10-31 14:06:00 3.02901554 secs
#> 51       resolved       <NA>   <NA> 2025-10-31 14:06:00 3.05819702 secs
#> 52       resolved       <NA>   <NA> 2025-10-31 14:06:00 3.08467865 secs
#> 53       resolved       <NA>   <NA> 2025-10-31 14:06:00 3.11198330 secs
#> 54       resolved       <NA>   <NA> 2025-10-31 14:06:00 3.13799453 secs
#> 55       resolved       <NA>   <NA> 2025-10-31 14:06:00 3.16394067 secs
#> 56       resolved       <NA>   <NA> 2025-10-31 14:06:00 3.19068670 secs
#> 57       resolved       <NA>   <NA> 2025-10-31 14:06:00 3.21655178 secs
#> 58       resolved       <NA>   <NA> 2025-10-31 14:06:00 3.24334431 secs
#> 59       resolved       <NA>   <NA> 2025-10-31 14:06:00 3.26851773 secs
#> 60       resolved       <NA>   <NA> 2025-10-31 14:06:00 3.29216552 secs
#> 61       resolved       <NA>   <NA> 2025-10-31 14:06:00 3.31470490 secs
#> 62       resolved       <NA>   <NA> 2025-10-31 14:06:00 3.33726263 secs
#> 63       resolved       <NA>   <NA> 2025-10-31 14:06:00 3.35992217 secs
#> 64       resolved       <NA>   <NA> 2025-10-31 14:06:00 3.38531613 secs
#> 65       resolved       <NA>   <NA> 2025-10-31 14:06:00 3.41153598 secs
#> 66       resolved       <NA>   <NA> 2025-10-31 14:06:00 3.43608141 secs
#> 67       resolved       <NA>   <NA> 2025-10-31 14:06:00 3.45921636 secs
#> 68       resolved       <NA>   <NA> 2025-10-31 14:06:01 3.48183680 secs
#> 69       resolved       <NA>   <NA> 2025-10-31 14:06:01 3.50462914 secs
#> 70       resolved       <NA>   <NA> 2025-10-31 14:06:01 3.52711821 secs
#> 71       resolved       <NA>   <NA> 2025-10-31 14:06:01 3.55065799 secs
#> 72       resolved       <NA>   <NA> 2025-10-31 14:06:01 3.57374454 secs
#> 73       resolved       <NA>   <NA> 2025-10-31 14:06:01 3.59607410 secs
#> 74       resolved       <NA>   <NA> 2025-10-31 14:06:01 3.61841822 secs
#> 75       resolved       <NA>   <NA> 2025-10-31 14:06:01 3.64149022 secs
#> 76       resolved       <NA>   <NA> 2025-10-31 14:06:01 3.66631246 secs
#> 77       resolved       <NA>   <NA> 2025-10-31 14:06:01 3.68956065 secs
#> 78       resolved       <NA>   <NA> 2025-10-31 14:06:01 3.71194339 secs
#> 79       resolved       <NA>   <NA> 2025-10-31 14:06:01 3.73480797 secs
#> 80       resolved       <NA>   <NA> 2025-10-31 14:06:01 3.75801206 secs
#> 81       resolved       <NA>   <NA> 2025-10-31 14:06:01 3.78121829 secs
#> 82       resolved       <NA>   <NA> 2025-10-31 14:06:01 3.80615735 secs
#> 83       resolved       <NA>   <NA> 2025-10-31 14:06:01 3.82966805 secs
#> 84       resolved       <NA>   <NA> 2025-10-31 14:06:01 3.85311937 secs
#> 85       resolved       <NA>   <NA> 2025-10-31 14:06:01 3.87633848 secs
#> 86       resolved       <NA>   <NA> 2025-10-31 14:06:01 3.89871120 secs
#> 87       resolved       <NA>   <NA> 2025-10-31 14:06:01 3.92170715 secs
#> 88       resolved       <NA>   <NA> 2025-10-31 14:06:01 3.94475555 secs
#> 89       resolved       <NA>   <NA> 2025-10-31 14:06:01 3.96699429 secs
#> 90       resolved       <NA>   <NA> 2025-10-31 14:06:01 3.98936486 secs
#> 91       resolved       <NA>   <NA> 2025-10-31 14:06:01 4.01177382 secs
#> 92       resolved       <NA>   <NA> 2025-10-31 14:06:01 4.03424549 secs
#> 93       resolved       <NA>   <NA> 2025-10-31 14:06:01 4.05706525 secs
#> 94       resolved       <NA>   <NA> 2025-10-31 14:06:01 4.08193040 secs
#> 95       resolved       <NA>   <NA> 2025-10-31 14:06:01 4.10749483 secs
#> 96       resolved       <NA>   <NA> 2025-10-31 14:06:01 4.13283730 secs
#> 97       resolved       <NA>   <NA> 2025-10-31 14:06:01 4.15644121 secs
#> 98       resolved       <NA>   <NA> 2025-10-31 14:06:01 4.17918611 secs
#> 99       resolved       <NA>   <NA> 2025-10-31 14:06:01 4.20353937 secs
#> 100      resolved       <NA>   <NA> 2025-10-31 14:06:01 4.22708201 secs
#> 101      resolved       <NA>   <NA> 2025-10-31 14:06:01 4.25034523 secs
#> 102      resolved       <NA>   <NA> 2025-10-31 14:06:01 4.27277231 secs
#> 103      resolved       <NA>   <NA> 2025-10-31 14:06:01 4.29499722 secs
#> 104      resolved       <NA>   <NA> 2025-10-31 14:06:01 4.31738114 secs
#> 105      resolved       <NA>   <NA> 2025-10-31 14:06:01 4.34020925 secs
#> 106      resolved       <NA>   <NA> 2025-10-31 14:06:01 4.36288214 secs
#> 107      resolved       <NA>   <NA> 2025-10-31 14:06:01 4.38668847 secs
#> 108      resolved       <NA>   <NA> 2025-10-31 14:06:01 4.41010332 secs
#> 109      resolved       <NA>   <NA> 2025-10-31 14:06:01 4.43265891 secs
#> 110      resolved       <NA>   <NA> 2025-10-31 14:06:01 4.45530200 secs
#> 111      resolved       <NA>   <NA> 2025-10-31 14:06:02 4.47869182 secs
#> 115      resolved       <NA>   <NA> 2025-10-31 14:06:02 4.50149727 secs
#> 112 receiveResult   overhead gather 2025-10-31 14:06:02 4.50290227 secs
#> 113        gather   overhead   <NA> 2025-10-31 14:06:02 4.50424385 secs
#> 116       resolve   overhead   <NA> 2025-10-31 14:06:02 4.50906515 secs
#>              duration future_label                         future_uuid
#> 1   1.188040e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 4   1.442502e+00 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 2   1.429242e+00 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 3   7.386208e-04 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 5   1.083589e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 114 3.006380e+00 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 6   1.080179e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 7   1.265454e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 8   1.261759e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 9   1.285625e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 10  1.271200e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 11  1.347542e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 12  1.278925e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 13  1.285219e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 14  1.260519e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 15  1.258707e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 16  1.341629e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 17  1.263261e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 18  1.278448e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 19  1.260662e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 20  1.161766e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 21  1.219678e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 22  1.095533e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 23  1.129723e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 24  1.263642e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 25  1.184893e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 26  1.102686e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 27  1.114058e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 28  1.112986e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 29  1.396132e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 30  1.442361e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 31  1.180053e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 32  1.133728e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 33  1.121521e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 34  1.160407e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 35  1.345301e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 36  1.336288e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 37  1.270199e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 38  1.277566e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 39  1.274443e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 40  1.297712e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 41  1.286674e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 42  1.305842e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 43  1.300430e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 44  1.294112e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 45  1.329827e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 46  1.271200e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 47  1.298046e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 48  1.277113e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 49  1.311874e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 50  1.355743e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 51  1.265883e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 52  1.280141e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 53  1.275420e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 54  1.257992e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 55  1.339054e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 56  1.263237e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 57  1.272202e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 58  1.263762e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 59  1.193500e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 60  1.122332e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 61  1.115608e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 62  1.106095e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 63  1.196885e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 64  1.292467e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 65  1.197958e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 66  1.101327e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 67  1.103592e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 68  1.103997e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 69  1.107097e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 70  1.221490e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 71  1.126146e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 72  1.103163e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 73  1.111007e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 74  1.138544e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 75  1.218867e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 76  1.124406e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 77  1.108027e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 78  1.103210e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 79  1.129580e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 80  1.112151e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 81  1.218224e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 82  1.152110e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 83  1.188493e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 84  1.186538e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 85  1.102638e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 86  1.175117e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 87  1.113701e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 88  1.101232e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 89  1.106644e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 90  1.124191e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 91  1.113415e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 92  1.105714e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 93  1.184130e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 94  1.245880e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 95  1.240516e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 96  1.159430e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 97  1.145673e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 98  1.273990e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 99  1.195025e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 100 1.194525e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 101 1.111412e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 102 1.099539e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 103 1.112342e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 104 1.105571e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 105 1.109672e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 106 1.245642e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 107 1.118326e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 108 1.113820e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 109 1.136279e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 110 1.114416e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 111 1.147389e-02 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 115 6.048679e-03 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 112 2.119541e-04 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 113 9.560585e-05 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#> 116 2.336502e-05 secs         <NA> 2107504b07336839b4018835cafa74f4-19
#>                         session_uuid
#> 1   2107504b07336839b4018835cafa74f4
#> 4   2107504b07336839b4018835cafa74f4
#> 2   2107504b07336839b4018835cafa74f4
#> 3   2107504b07336839b4018835cafa74f4
#> 5   2107504b07336839b4018835cafa74f4
#> 114 2107504b07336839b4018835cafa74f4
#> 6   2107504b07336839b4018835cafa74f4
#> 7   2107504b07336839b4018835cafa74f4
#> 8   2107504b07336839b4018835cafa74f4
#> 9   2107504b07336839b4018835cafa74f4
#> 10  2107504b07336839b4018835cafa74f4
#> 11  2107504b07336839b4018835cafa74f4
#> 12  2107504b07336839b4018835cafa74f4
#> 13  2107504b07336839b4018835cafa74f4
#> 14  2107504b07336839b4018835cafa74f4
#> 15  2107504b07336839b4018835cafa74f4
#> 16  2107504b07336839b4018835cafa74f4
#> 17  2107504b07336839b4018835cafa74f4
#> 18  2107504b07336839b4018835cafa74f4
#> 19  2107504b07336839b4018835cafa74f4
#> 20  2107504b07336839b4018835cafa74f4
#> 21  2107504b07336839b4018835cafa74f4
#> 22  2107504b07336839b4018835cafa74f4
#> 23  2107504b07336839b4018835cafa74f4
#> 24  2107504b07336839b4018835cafa74f4
#> 25  2107504b07336839b4018835cafa74f4
#> 26  2107504b07336839b4018835cafa74f4
#> 27  2107504b07336839b4018835cafa74f4
#> 28  2107504b07336839b4018835cafa74f4
#> 29  2107504b07336839b4018835cafa74f4
#> 30  2107504b07336839b4018835cafa74f4
#> 31  2107504b07336839b4018835cafa74f4
#> 32  2107504b07336839b4018835cafa74f4
#> 33  2107504b07336839b4018835cafa74f4
#> 34  2107504b07336839b4018835cafa74f4
#> 35  2107504b07336839b4018835cafa74f4
#> 36  2107504b07336839b4018835cafa74f4
#> 37  2107504b07336839b4018835cafa74f4
#> 38  2107504b07336839b4018835cafa74f4
#> 39  2107504b07336839b4018835cafa74f4
#> 40  2107504b07336839b4018835cafa74f4
#> 41  2107504b07336839b4018835cafa74f4
#> 42  2107504b07336839b4018835cafa74f4
#> 43  2107504b07336839b4018835cafa74f4
#> 44  2107504b07336839b4018835cafa74f4
#> 45  2107504b07336839b4018835cafa74f4
#> 46  2107504b07336839b4018835cafa74f4
#> 47  2107504b07336839b4018835cafa74f4
#> 48  2107504b07336839b4018835cafa74f4
#> 49  2107504b07336839b4018835cafa74f4
#> 50  2107504b07336839b4018835cafa74f4
#> 51  2107504b07336839b4018835cafa74f4
#> 52  2107504b07336839b4018835cafa74f4
#> 53  2107504b07336839b4018835cafa74f4
#> 54  2107504b07336839b4018835cafa74f4
#> 55  2107504b07336839b4018835cafa74f4
#> 56  2107504b07336839b4018835cafa74f4
#> 57  2107504b07336839b4018835cafa74f4
#> 58  2107504b07336839b4018835cafa74f4
#> 59  2107504b07336839b4018835cafa74f4
#> 60  2107504b07336839b4018835cafa74f4
#> 61  2107504b07336839b4018835cafa74f4
#> 62  2107504b07336839b4018835cafa74f4
#> 63  2107504b07336839b4018835cafa74f4
#> 64  2107504b07336839b4018835cafa74f4
#> 65  2107504b07336839b4018835cafa74f4
#> 66  2107504b07336839b4018835cafa74f4
#> 67  2107504b07336839b4018835cafa74f4
#> 68  2107504b07336839b4018835cafa74f4
#> 69  2107504b07336839b4018835cafa74f4
#> 70  2107504b07336839b4018835cafa74f4
#> 71  2107504b07336839b4018835cafa74f4
#> 72  2107504b07336839b4018835cafa74f4
#> 73  2107504b07336839b4018835cafa74f4
#> 74  2107504b07336839b4018835cafa74f4
#> 75  2107504b07336839b4018835cafa74f4
#> 76  2107504b07336839b4018835cafa74f4
#> 77  2107504b07336839b4018835cafa74f4
#> 78  2107504b07336839b4018835cafa74f4
#> 79  2107504b07336839b4018835cafa74f4
#> 80  2107504b07336839b4018835cafa74f4
#> 81  2107504b07336839b4018835cafa74f4
#> 82  2107504b07336839b4018835cafa74f4
#> 83  2107504b07336839b4018835cafa74f4
#> 84  2107504b07336839b4018835cafa74f4
#> 85  2107504b07336839b4018835cafa74f4
#> 86  2107504b07336839b4018835cafa74f4
#> 87  2107504b07336839b4018835cafa74f4
#> 88  2107504b07336839b4018835cafa74f4
#> 89  2107504b07336839b4018835cafa74f4
#> 90  2107504b07336839b4018835cafa74f4
#> 91  2107504b07336839b4018835cafa74f4
#> 92  2107504b07336839b4018835cafa74f4
#> 93  2107504b07336839b4018835cafa74f4
#> 94  2107504b07336839b4018835cafa74f4
#> 95  2107504b07336839b4018835cafa74f4
#> 96  2107504b07336839b4018835cafa74f4
#> 97  2107504b07336839b4018835cafa74f4
#> 98  2107504b07336839b4018835cafa74f4
#> 99  2107504b07336839b4018835cafa74f4
#> 100 2107504b07336839b4018835cafa74f4
#> 101 2107504b07336839b4018835cafa74f4
#> 102 2107504b07336839b4018835cafa74f4
#> 103 2107504b07336839b4018835cafa74f4
#> 104 2107504b07336839b4018835cafa74f4
#> 105 2107504b07336839b4018835cafa74f4
#> 106 2107504b07336839b4018835cafa74f4
#> 107 2107504b07336839b4018835cafa74f4
#> 108 2107504b07336839b4018835cafa74f4
#> 109 2107504b07336839b4018835cafa74f4
#> 110 2107504b07336839b4018835cafa74f4
#> 111 2107504b07336839b4018835cafa74f4
#> 115 2107504b07336839b4018835cafa74f4
#> 112 2107504b07336839b4018835cafa74f4
#> 113 2107504b07336839b4018835cafa74f4
#> 116 2107504b07336839b4018835cafa74f4
#> 

## Stop parallel workers and disable journal logging and signaling
plan(sequential)
options(oopts)