Gets the logged journal of events for a future

WARNING: This function is under development. It can change at any time. For now, please, do not depend on this function in a published R package.

Usage

journal(x, ...)

Arguments

x

A future::Future object.

...

Not used.

Value

A data frame of class FutureJournal with columns:

event (character string) - type of event that took place

category (character string) - the category of the event

parent (character string) - (to be describe)

start (POSIXct) - the timestamp when the event started

at (difftime) - the time when the event started relative to first event

duration (difftime) - the duration of the event

future_label (character string) - the label of the future

future_uuid (character string) - the UUID of the future

session_uuid (character string) - the UUID of the R session where the event took place

The common events are:

create - the future was created (an overhead)

launch - the future was launched (an overhead)

evaluate - the future was evaluated (an evaluation)

resolved - the future was queried (may be occur multiple times) (an overhead)

gather - the results was retrieved (an overhead)

but others may be added by other Future classes.

Common event categories are:

evaluation - processing time is spent on evaluation

overhead - processing time is spent on orchestrating the future

waiting - processing time is spent on waiting to set up or querying the future

but others may be added by other Future classes.

The data frame is sorted by the at time. Note that the timestamps for the evaluate event are based on the local time on the worker. The system clocks on the worker and the calling R system may not be in perfect sync.

Enabling and disabling event logging

To enable logging of events, set option future.journal is TRUE. To disable, set it to FALSE (default).

See also

Examples

## Enable journaling of futures
oopts <- options(future.journal = TRUE)

plan(multisession, workers = 2L)

t_start <- Sys.time()
fs <- lapply(1:3, FUN = function(x) future({ Sys.sleep(x); sqrt(x) }))
vs <- value(fs)
js <- lapply(fs, FUN = journal, baseline = t_start)
print(js)
#> [[1]]
#>            event   category parent               start              at
#> 1         create   overhead   <NA> 2025-10-31 14:11:23 0.00000000 secs
#> 4         launch   overhead   <NA> 2025-10-31 14:11:23 0.01063299 secs
#> 2      getWorker   overhead launch 2025-10-31 14:11:23 0.01077867 secs
#> 3    eraseWorker   overhead launch 2025-10-31 14:11:23 0.01217175 secs
#> 5       resolved       <NA>   <NA> 2025-10-31 14:11:23 0.07049584 secs
#> 6       resolved       <NA>   <NA> 2025-10-31 14:11:23 0.10472775 secs
#> 7       resolved       <NA>   <NA> 2025-10-31 14:11:23 0.13864779 secs
#> 8       resolved       <NA>   <NA> 2025-10-31 14:11:23 0.17262840 secs
#> 9       resolved       <NA>   <NA> 2025-10-31 14:11:23 0.20675659 secs
#> 10      resolved       <NA>   <NA> 2025-10-31 14:11:24 0.24162412 secs
#> 44      evaluate evaluation   <NA> 2025-10-31 14:11:24 0.25561714 secs
#> 11      resolved       <NA>   <NA> 2025-10-31 14:11:24 0.27599454 secs
#> 12      resolved       <NA>   <NA> 2025-10-31 14:11:24 0.31234598 secs
#> 13      resolved       <NA>   <NA> 2025-10-31 14:11:24 0.35297894 secs
#> 14      resolved       <NA>   <NA> 2025-10-31 14:11:24 0.38843107 secs
#> 15      resolved       <NA>   <NA> 2025-10-31 14:11:24 0.42521429 secs
#> 16      resolved       <NA>   <NA> 2025-10-31 14:11:24 0.46078920 secs
#> 17      resolved       <NA>   <NA> 2025-10-31 14:11:24 0.49561834 secs
#> 18      resolved       <NA>   <NA> 2025-10-31 14:11:24 0.53164053 secs
#> 19      resolved       <NA>   <NA> 2025-10-31 14:11:24 0.56924772 secs
#> 20      resolved       <NA>   <NA> 2025-10-31 14:11:24 0.60731673 secs
#> 21      resolved       <NA>   <NA> 2025-10-31 14:11:24 0.64526463 secs
#> 22      resolved       <NA>   <NA> 2025-10-31 14:11:24 0.68208313 secs
#> 23      resolved       <NA>   <NA> 2025-10-31 14:11:24 0.71823907 secs
#> 24      resolved       <NA>   <NA> 2025-10-31 14:11:24 0.75507760 secs
#> 25      resolved       <NA>   <NA> 2025-10-31 14:11:24 0.79187799 secs
#> 26      resolved       <NA>   <NA> 2025-10-31 14:11:24 0.82759428 secs
#> 27      resolved       <NA>   <NA> 2025-10-31 14:11:24 0.86326051 secs
#> 28      resolved       <NA>   <NA> 2025-10-31 14:11:24 0.89946723 secs
#> 29      resolved       <NA>   <NA> 2025-10-31 14:11:24 0.93592381 secs
#> 30      resolved       <NA>   <NA> 2025-10-31 14:11:24 0.97247076 secs
#> 31      resolved       <NA>   <NA> 2025-10-31 14:11:24 1.00922656 secs
#> 32      resolved       <NA>   <NA> 2025-10-31 14:11:24 1.04601693 secs
#> 33      resolved       <NA>   <NA> 2025-10-31 14:11:24 1.08285046 secs
#> 34      resolved       <NA>   <NA> 2025-10-31 14:11:24 1.12003875 secs
#> 35      resolved       <NA>   <NA> 2025-10-31 14:11:24 1.15699029 secs
#> 36      resolved       <NA>   <NA> 2025-10-31 14:11:24 1.19431424 secs
#> 37      resolved       <NA>   <NA> 2025-10-31 14:11:24 1.23332310 secs
#> 38      resolved       <NA>   <NA> 2025-10-31 14:11:25 1.27117205 secs
#> 39      resolved       <NA>   <NA> 2025-10-31 14:11:25 1.30905986 secs
#> 40      resolved       <NA>   <NA> 2025-10-31 14:11:25 1.34730458 secs
#> 41      resolved       <NA>   <NA> 2025-10-31 14:11:25 1.39151669 secs
#> 45      resolved       <NA>   <NA> 2025-10-31 14:11:25 1.43684626 secs
#> 42 receiveResult   overhead gather 2025-10-31 14:11:25 1.43910027 secs
#> 43        gather   overhead   <NA> 2025-10-31 14:11:25 1.44249678 secs
#> 46      resolved       <NA>   <NA> 2025-10-31 14:11:25 1.48625898 secs
#> 47       resolve   overhead   <NA> 2025-10-31 14:11:25 1.48989153 secs
#>             duration future_label                         future_uuid
#> 1  1.022005e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-17
#> 4  7.390499e-03 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-17
#> 2  6.618500e-04 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-17
#> 3  4.708767e-04 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-17
#> 5  1.107454e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-17
#> 6  1.085234e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-17
#> 7  1.086187e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-17
#> 8  1.087332e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-17
#> 9  1.083899e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-17
#> 10 1.086903e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-17
#> 44 1.170696e+00 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-17
#> 11 1.121974e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-17
#> 12 1.129532e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-17
#> 13 1.109958e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-17
#> 14 1.136112e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-17
#> 15 1.137686e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-17
#> 16 1.081181e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-17
#> 17 1.126003e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-17
#> 18 1.123500e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-17
#> 19 1.148558e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-17
#> 20 1.153636e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-17
#> 21 1.098776e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-17
#> 22 1.106477e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-17
#> 23 1.108551e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-17
#> 24 1.104212e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-17
#> 25 1.083851e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-17
#> 26 1.077318e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-17
#> 27 1.080370e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-17
#> 28 1.086974e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-17
#> 29 1.089883e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-17
#> 30 1.091504e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-17
#> 31 1.088071e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-17
#> 32 1.083994e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-17
#> 33 1.087999e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-17
#> 34 1.079202e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-17
#> 35 1.090169e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-17
#> 36 1.108575e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-17
#> 37 1.084518e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-17
#> 38 1.079798e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-17
#> 39 1.081085e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-17
#> 40 1.141810e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-17
#> 41 1.240897e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-17
#> 45 1.532817e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-17
#> 42 4.832745e-04 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-17
#> 43 2.422333e-04 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-17
#> 46 1.814365e-04 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-17
#> 47 4.315376e-05 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-17
#>                        session_uuid
#> 1  8313298c511e3f14075b4f3bc7be71fb
#> 4  8313298c511e3f14075b4f3bc7be71fb
#> 2  8313298c511e3f14075b4f3bc7be71fb
#> 3  8313298c511e3f14075b4f3bc7be71fb
#> 5  8313298c511e3f14075b4f3bc7be71fb
#> 6  8313298c511e3f14075b4f3bc7be71fb
#> 7  8313298c511e3f14075b4f3bc7be71fb
#> 8  8313298c511e3f14075b4f3bc7be71fb
#> 9  8313298c511e3f14075b4f3bc7be71fb
#> 10 8313298c511e3f14075b4f3bc7be71fb
#> 44 8313298c511e3f14075b4f3bc7be71fb
#> 11 8313298c511e3f14075b4f3bc7be71fb
#> 12 8313298c511e3f14075b4f3bc7be71fb
#> 13 8313298c511e3f14075b4f3bc7be71fb
#> 14 8313298c511e3f14075b4f3bc7be71fb
#> 15 8313298c511e3f14075b4f3bc7be71fb
#> 16 8313298c511e3f14075b4f3bc7be71fb
#> 17 8313298c511e3f14075b4f3bc7be71fb
#> 18 8313298c511e3f14075b4f3bc7be71fb
#> 19 8313298c511e3f14075b4f3bc7be71fb
#> 20 8313298c511e3f14075b4f3bc7be71fb
#> 21 8313298c511e3f14075b4f3bc7be71fb
#> 22 8313298c511e3f14075b4f3bc7be71fb
#> 23 8313298c511e3f14075b4f3bc7be71fb
#> 24 8313298c511e3f14075b4f3bc7be71fb
#> 25 8313298c511e3f14075b4f3bc7be71fb
#> 26 8313298c511e3f14075b4f3bc7be71fb
#> 27 8313298c511e3f14075b4f3bc7be71fb
#> 28 8313298c511e3f14075b4f3bc7be71fb
#> 29 8313298c511e3f14075b4f3bc7be71fb
#> 30 8313298c511e3f14075b4f3bc7be71fb
#> 31 8313298c511e3f14075b4f3bc7be71fb
#> 32 8313298c511e3f14075b4f3bc7be71fb
#> 33 8313298c511e3f14075b4f3bc7be71fb
#> 34 8313298c511e3f14075b4f3bc7be71fb
#> 35 8313298c511e3f14075b4f3bc7be71fb
#> 36 8313298c511e3f14075b4f3bc7be71fb
#> 37 8313298c511e3f14075b4f3bc7be71fb
#> 38 8313298c511e3f14075b4f3bc7be71fb
#> 39 8313298c511e3f14075b4f3bc7be71fb
#> 40 8313298c511e3f14075b4f3bc7be71fb
#> 41 8313298c511e3f14075b4f3bc7be71fb
#> 45 8313298c511e3f14075b4f3bc7be71fb
#> 42 8313298c511e3f14075b4f3bc7be71fb
#> 43 8313298c511e3f14075b4f3bc7be71fb
#> 46 8313298c511e3f14075b4f3bc7be71fb
#> 47 8313298c511e3f14075b4f3bc7be71fb
#> 
#> [[2]]
#>            event   category parent               start              at
#> 1         create   overhead   <NA> 2025-10-31 14:11:23 0.00000000 secs
#> 4         launch   overhead   <NA> 2025-10-31 14:11:23 0.01164722 secs
#> 2      getWorker   overhead launch 2025-10-31 14:11:23 0.01183510 secs
#> 3    eraseWorker   overhead launch 2025-10-31 14:11:23 0.01353693 secs
#> 5       resolved       <NA>   <NA> 2025-10-31 14:11:23 0.06363463 secs
#> 6       resolved       <NA>   <NA> 2025-10-31 14:11:23 0.09763193 secs
#> 7       resolved       <NA>   <NA> 2025-10-31 14:11:23 0.13156891 secs
#> 8       resolved       <NA>   <NA> 2025-10-31 14:11:23 0.16558027 secs
#> 9       resolved       <NA>   <NA> 2025-10-31 14:11:23 0.20012069 secs
#> 10      resolved       <NA>   <NA> 2025-10-31 14:11:24 0.23455739 secs
#> 11      resolved       <NA>   <NA> 2025-10-31 14:11:24 0.26979375 secs
#> 69      evaluate evaluation   <NA> 2025-10-31 14:11:24 0.27137375 secs
#> 12      resolved       <NA>   <NA> 2025-10-31 14:11:24 0.30577755 secs
#> 13      resolved       <NA>   <NA> 2025-10-31 14:11:24 0.34634757 secs
#> 14      resolved       <NA>   <NA> 2025-10-31 14:11:24 0.38263726 secs
#> 15      resolved       <NA>   <NA> 2025-10-31 14:11:24 0.41876507 secs
#> 16      resolved       <NA>   <NA> 2025-10-31 14:11:24 0.45364499 secs
#> 17      resolved       <NA>   <NA> 2025-10-31 14:11:24 0.48895097 secs
#> 18      resolved       <NA>   <NA> 2025-10-31 14:11:24 0.52552938 secs
#> 19      resolved       <NA>   <NA> 2025-10-31 14:11:24 0.56390262 secs
#> 20      resolved       <NA>   <NA> 2025-10-31 14:11:24 0.60163450 secs
#> 21      resolved       <NA>   <NA> 2025-10-31 14:11:24 0.63897491 secs
#> 22      resolved       <NA>   <NA> 2025-10-31 14:11:24 0.67534614 secs
#> 23      resolved       <NA>   <NA> 2025-10-31 14:11:24 0.71167421 secs
#> 24      resolved       <NA>   <NA> 2025-10-31 14:11:24 0.74842811 secs
#> 25      resolved       <NA>   <NA> 2025-10-31 14:11:24 0.78467393 secs
#> 26      resolved       <NA>   <NA> 2025-10-31 14:11:24 0.82039905 secs
#> 27      resolved       <NA>   <NA> 2025-10-31 14:11:24 0.85611558 secs
#> 28      resolved       <NA>   <NA> 2025-10-31 14:11:24 0.89240098 secs
#> 29      resolved       <NA>   <NA> 2025-10-31 14:11:24 0.92887473 secs
#> 30      resolved       <NA>   <NA> 2025-10-31 14:11:24 0.96544147 secs
#> 31      resolved       <NA>   <NA> 2025-10-31 14:11:24 1.00212026 secs
#> 32      resolved       <NA>   <NA> 2025-10-31 14:11:24 1.03886843 secs
#> 33      resolved       <NA>   <NA> 2025-10-31 14:11:24 1.07576919 secs
#> 34      resolved       <NA>   <NA> 2025-10-31 14:11:24 1.11278081 secs
#> 35      resolved       <NA>   <NA> 2025-10-31 14:11:24 1.14992785 secs
#> 36      resolved       <NA>   <NA> 2025-10-31 14:11:24 1.18816543 secs
#> 37      resolved       <NA>   <NA> 2025-10-31 14:11:25 1.22640204 secs
#> 38      resolved       <NA>   <NA> 2025-10-31 14:11:25 1.26425028 secs
#> 39      resolved       <NA>   <NA> 2025-10-31 14:11:25 1.30189610 secs
#> 40      resolved       <NA>   <NA> 2025-10-31 14:11:25 1.34274721 secs
#> 41      resolved       <NA>   <NA> 2025-10-31 14:11:25 1.38775349 secs
#> 42      resolved       <NA>   <NA> 2025-10-31 14:11:25 1.47311401 secs
#> 43      resolved       <NA>   <NA> 2025-10-31 14:11:25 1.51039553 secs
#> 44      resolved       <NA>   <NA> 2025-10-31 14:11:25 1.54356933 secs
#> 45      resolved       <NA>   <NA> 2025-10-31 14:11:25 1.57711029 secs
#> 46      resolved       <NA>   <NA> 2025-10-31 14:11:25 1.61121798 secs
#> 47      resolved       <NA>   <NA> 2025-10-31 14:11:25 1.64918280 secs
#> 48      resolved       <NA>   <NA> 2025-10-31 14:11:25 1.69173646 secs
#> 49      resolved       <NA>   <NA> 2025-10-31 14:11:25 1.73440051 secs
#> 50      resolved       <NA>   <NA> 2025-10-31 14:11:25 1.77335715 secs
#> 51      resolved       <NA>   <NA> 2025-10-31 14:11:25 1.80732512 secs
#> 52      resolved       <NA>   <NA> 2025-10-31 14:11:25 1.84152889 secs
#> 53      resolved       <NA>   <NA> 2025-10-31 14:11:25 1.88052440 secs
#> 54      resolved       <NA>   <NA> 2025-10-31 14:11:25 1.92196608 secs
#> 55      resolved       <NA>   <NA> 2025-10-31 14:11:25 1.96485806 secs
#> 56      resolved       <NA>   <NA> 2025-10-31 14:11:25 2.00580072 secs
#> 57      resolved       <NA>   <NA> 2025-10-31 14:11:25 2.04786730 secs
#> 58      resolved       <NA>   <NA> 2025-10-31 14:11:25 2.09059310 secs
#> 59      resolved       <NA>   <NA> 2025-10-31 14:11:25 2.13247514 secs
#> 60      resolved       <NA>   <NA> 2025-10-31 14:11:25 2.17525935 secs
#> 61      resolved       <NA>   <NA> 2025-10-31 14:11:26 2.21720481 secs
#> 62      resolved       <NA>   <NA> 2025-10-31 14:11:26 2.25867653 secs
#> 63      resolved       <NA>   <NA> 2025-10-31 14:11:26 2.30119991 secs
#> 64      resolved       <NA>   <NA> 2025-10-31 14:11:26 2.34291196 secs
#> 65      resolved       <NA>   <NA> 2025-10-31 14:11:26 2.38521934 secs
#> 66      resolved       <NA>   <NA> 2025-10-31 14:11:26 2.42716026 secs
#> 70      resolved       <NA>   <NA> 2025-10-31 14:11:26 2.46916294 secs
#> 67 receiveResult   overhead gather 2025-10-31 14:11:26 2.47166276 secs
#> 68        gather   overhead   <NA> 2025-10-31 14:11:26 2.47557378 secs
#> 71       resolve   overhead   <NA> 2025-10-31 14:11:26 2.49093533 secs
#>             duration future_label                         future_uuid
#> 1  1.112008e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 4  1.116538e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 2  8.378029e-04 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 3  6.728172e-04 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 5  1.098895e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 6  1.082039e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 7  1.079011e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 8  1.082516e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 9  1.106906e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 10 1.081133e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 11 1.111555e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 69 2.171439e+00 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 12 1.599741e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 13 1.092672e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 14 1.126766e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 15 1.088095e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 16 1.076388e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 17 1.080298e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 18 1.108932e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 19 1.144433e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 20 1.133609e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 21 1.126075e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 22 1.078844e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 23 1.115513e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 24 1.105928e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 25 1.067924e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 26 1.064396e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 27 1.079154e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 28 1.080775e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 29 1.082063e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 30 1.081443e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 31 1.075864e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 32 1.076126e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 33 1.078033e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 34 1.073790e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 35 1.080823e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 36 1.104879e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 37 1.077890e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 38 1.088166e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 39 1.096725e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 40 1.232409e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 41 1.222801e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 42 1.130009e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 43 1.081729e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 44 1.079440e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 45 1.097488e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 46 1.102710e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 47 1.306248e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 48 1.272011e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 49 1.307964e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 50 1.144838e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 51 1.100302e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 52 1.156163e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 53 1.172113e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 54 1.282334e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 55 1.255322e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 56 1.296401e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 57 1.351380e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 58 1.271296e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 59 1.321959e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 60 1.285410e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 61 1.266527e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 62 1.356316e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 63 1.276851e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 64 1.274514e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 65 1.268220e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 66 1.266432e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 70 1.671815e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 67 5.738735e-04 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 68 2.717972e-04 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#> 71 9.441376e-05 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-18
#>                        session_uuid
#> 1  8313298c511e3f14075b4f3bc7be71fb
#> 4  8313298c511e3f14075b4f3bc7be71fb
#> 2  8313298c511e3f14075b4f3bc7be71fb
#> 3  8313298c511e3f14075b4f3bc7be71fb
#> 5  8313298c511e3f14075b4f3bc7be71fb
#> 6  8313298c511e3f14075b4f3bc7be71fb
#> 7  8313298c511e3f14075b4f3bc7be71fb
#> 8  8313298c511e3f14075b4f3bc7be71fb
#> 9  8313298c511e3f14075b4f3bc7be71fb
#> 10 8313298c511e3f14075b4f3bc7be71fb
#> 11 8313298c511e3f14075b4f3bc7be71fb
#> 69 8313298c511e3f14075b4f3bc7be71fb
#> 12 8313298c511e3f14075b4f3bc7be71fb
#> 13 8313298c511e3f14075b4f3bc7be71fb
#> 14 8313298c511e3f14075b4f3bc7be71fb
#> 15 8313298c511e3f14075b4f3bc7be71fb
#> 16 8313298c511e3f14075b4f3bc7be71fb
#> 17 8313298c511e3f14075b4f3bc7be71fb
#> 18 8313298c511e3f14075b4f3bc7be71fb
#> 19 8313298c511e3f14075b4f3bc7be71fb
#> 20 8313298c511e3f14075b4f3bc7be71fb
#> 21 8313298c511e3f14075b4f3bc7be71fb
#> 22 8313298c511e3f14075b4f3bc7be71fb
#> 23 8313298c511e3f14075b4f3bc7be71fb
#> 24 8313298c511e3f14075b4f3bc7be71fb
#> 25 8313298c511e3f14075b4f3bc7be71fb
#> 26 8313298c511e3f14075b4f3bc7be71fb
#> 27 8313298c511e3f14075b4f3bc7be71fb
#> 28 8313298c511e3f14075b4f3bc7be71fb
#> 29 8313298c511e3f14075b4f3bc7be71fb
#> 30 8313298c511e3f14075b4f3bc7be71fb
#> 31 8313298c511e3f14075b4f3bc7be71fb
#> 32 8313298c511e3f14075b4f3bc7be71fb
#> 33 8313298c511e3f14075b4f3bc7be71fb
#> 34 8313298c511e3f14075b4f3bc7be71fb
#> 35 8313298c511e3f14075b4f3bc7be71fb
#> 36 8313298c511e3f14075b4f3bc7be71fb
#> 37 8313298c511e3f14075b4f3bc7be71fb
#> 38 8313298c511e3f14075b4f3bc7be71fb
#> 39 8313298c511e3f14075b4f3bc7be71fb
#> 40 8313298c511e3f14075b4f3bc7be71fb
#> 41 8313298c511e3f14075b4f3bc7be71fb
#> 42 8313298c511e3f14075b4f3bc7be71fb
#> 43 8313298c511e3f14075b4f3bc7be71fb
#> 44 8313298c511e3f14075b4f3bc7be71fb
#> 45 8313298c511e3f14075b4f3bc7be71fb
#> 46 8313298c511e3f14075b4f3bc7be71fb
#> 47 8313298c511e3f14075b4f3bc7be71fb
#> 48 8313298c511e3f14075b4f3bc7be71fb
#> 49 8313298c511e3f14075b4f3bc7be71fb
#> 50 8313298c511e3f14075b4f3bc7be71fb
#> 51 8313298c511e3f14075b4f3bc7be71fb
#> 52 8313298c511e3f14075b4f3bc7be71fb
#> 53 8313298c511e3f14075b4f3bc7be71fb
#> 54 8313298c511e3f14075b4f3bc7be71fb
#> 55 8313298c511e3f14075b4f3bc7be71fb
#> 56 8313298c511e3f14075b4f3bc7be71fb
#> 57 8313298c511e3f14075b4f3bc7be71fb
#> 58 8313298c511e3f14075b4f3bc7be71fb
#> 59 8313298c511e3f14075b4f3bc7be71fb
#> 60 8313298c511e3f14075b4f3bc7be71fb
#> 61 8313298c511e3f14075b4f3bc7be71fb
#> 62 8313298c511e3f14075b4f3bc7be71fb
#> 63 8313298c511e3f14075b4f3bc7be71fb
#> 64 8313298c511e3f14075b4f3bc7be71fb
#> 65 8313298c511e3f14075b4f3bc7be71fb
#> 66 8313298c511e3f14075b4f3bc7be71fb
#> 70 8313298c511e3f14075b4f3bc7be71fb
#> 67 8313298c511e3f14075b4f3bc7be71fb
#> 68 8313298c511e3f14075b4f3bc7be71fb
#> 71 8313298c511e3f14075b4f3bc7be71fb
#> 
#> [[3]]
#>             event   category parent               start              at
#> 1          create   overhead   <NA> 2025-10-31 14:11:23 0.00000000 secs
#> 4          launch   overhead   <NA> 2025-10-31 14:11:23 0.01709008 secs
#> 2       getWorker   overhead launch 2025-10-31 14:11:23 0.01724482 secs
#> 3     eraseWorker   overhead launch 2025-10-31 14:11:25 1.41706443 secs
#> 5        resolved       <NA>   <NA> 2025-10-31 14:11:25 1.46292090 secs
#> 112      evaluate evaluation   <NA> 2025-10-31 14:11:25 1.47782397 secs
#> 6        resolved       <NA>   <NA> 2025-10-31 14:11:25 1.49786973 secs
#> 7        resolved       <NA>   <NA> 2025-10-31 14:11:25 1.53101015 secs
#> 8        resolved       <NA>   <NA> 2025-10-31 14:11:25 1.56473327 secs
#> 9        resolved       <NA>   <NA> 2025-10-31 14:11:25 1.60045362 secs
#> 10       resolved       <NA>   <NA> 2025-10-31 14:11:25 1.64118528 secs
#> 11       resolved       <NA>   <NA> 2025-10-31 14:11:25 1.68302107 secs
#> 12       resolved       <NA>   <NA> 2025-10-31 14:11:25 1.72563624 secs
#> 13       resolved       <NA>   <NA> 2025-10-31 14:11:25 1.76150894 secs
#> 14       resolved       <NA>   <NA> 2025-10-31 14:11:25 1.79523492 secs
#> 15       resolved       <NA>   <NA> 2025-10-31 14:11:25 1.83146834 secs
#> 16       resolved       <NA>   <NA> 2025-10-31 14:11:25 1.87136149 secs
#> 17       resolved       <NA>   <NA> 2025-10-31 14:11:25 1.91420865 secs
#> 18       resolved       <NA>   <NA> 2025-10-31 14:11:25 1.95597792 secs
#> 19       resolved       <NA>   <NA> 2025-10-31 14:11:25 1.99777818 secs
#> 20       resolved       <NA>   <NA> 2025-10-31 14:11:25 2.04033089 secs
#> 21       resolved       <NA>   <NA> 2025-10-31 14:11:25 2.08226204 secs
#> 22       resolved       <NA>   <NA> 2025-10-31 14:11:25 2.12447047 secs
#> 23       resolved       <NA>   <NA> 2025-10-31 14:11:25 2.16697812 secs
#> 24       resolved       <NA>   <NA> 2025-10-31 14:11:26 2.20866275 secs
#> 25       resolved       <NA>   <NA> 2025-10-31 14:11:26 2.25113988 secs
#> 26       resolved       <NA>   <NA> 2025-10-31 14:11:26 2.29290175 secs
#> 27       resolved       <NA>   <NA> 2025-10-31 14:11:26 2.33463764 secs
#> 28       resolved       <NA>   <NA> 2025-10-31 14:11:26 2.37685943 secs
#> 29       resolved       <NA>   <NA> 2025-10-31 14:11:26 2.41866589 secs
#> 30       resolved       <NA>   <NA> 2025-10-31 14:11:26 2.47013569 secs
#> 31       resolved       <NA>   <NA> 2025-10-31 14:11:26 2.49478793 secs
#> 32       resolved       <NA>   <NA> 2025-10-31 14:11:26 2.51921344 secs
#> 33       resolved       <NA>   <NA> 2025-10-31 14:11:26 2.55246401 secs
#> 34       resolved       <NA>   <NA> 2025-10-31 14:11:26 2.57492423 secs
#> 35       resolved       <NA>   <NA> 2025-10-31 14:11:26 2.59741592 secs
#> 36       resolved       <NA>   <NA> 2025-10-31 14:11:26 2.62494469 secs
#> 37       resolved       <NA>   <NA> 2025-10-31 14:11:26 2.65214539 secs
#> 38       resolved       <NA>   <NA> 2025-10-31 14:11:26 2.67739582 secs
#> 39       resolved       <NA>   <NA> 2025-10-31 14:11:26 2.70262027 secs
#> 40       resolved       <NA>   <NA> 2025-10-31 14:11:26 2.72814965 secs
#> 41       resolved       <NA>   <NA> 2025-10-31 14:11:26 2.75405169 secs
#> 42       resolved       <NA>   <NA> 2025-10-31 14:11:26 2.78024006 secs
#> 43       resolved       <NA>   <NA> 2025-10-31 14:11:26 2.80722523 secs
#> 44       resolved       <NA>   <NA> 2025-10-31 14:11:26 2.83360267 secs
#> 45       resolved       <NA>   <NA> 2025-10-31 14:11:26 2.85974407 secs
#> 46       resolved       <NA>   <NA> 2025-10-31 14:11:26 2.88582015 secs
#> 47       resolved       <NA>   <NA> 2025-10-31 14:11:26 2.91208720 secs
#> 48       resolved       <NA>   <NA> 2025-10-31 14:11:26 2.93823981 secs
#> 49       resolved       <NA>   <NA> 2025-10-31 14:11:26 2.96428776 secs
#> 50       resolved       <NA>   <NA> 2025-10-31 14:11:26 2.99038768 secs
#> 51       resolved       <NA>   <NA> 2025-10-31 14:11:26 3.01681066 secs
#> 52       resolved       <NA>   <NA> 2025-10-31 14:11:26 3.04357290 secs
#> 53       resolved       <NA>   <NA> 2025-10-31 14:11:26 3.07056808 secs
#> 54       resolved       <NA>   <NA> 2025-10-31 14:11:26 3.09708357 secs
#> 55       resolved       <NA>   <NA> 2025-10-31 14:11:26 3.12395120 secs
#> 56       resolved       <NA>   <NA> 2025-10-31 14:11:26 3.15017796 secs
#> 57       resolved       <NA>   <NA> 2025-10-31 14:11:26 3.17631555 secs
#> 58       resolved       <NA>   <NA> 2025-10-31 14:11:27 3.20260334 secs
#> 59       resolved       <NA>   <NA> 2025-10-31 14:11:27 3.22921348 secs
#> 60       resolved       <NA>   <NA> 2025-10-31 14:11:27 3.25338006 secs
#> 61       resolved       <NA>   <NA> 2025-10-31 14:11:27 3.27645278 secs
#> 62       resolved       <NA>   <NA> 2025-10-31 14:11:27 3.29928279 secs
#> 63       resolved       <NA>   <NA> 2025-10-31 14:11:27 3.32518888 secs
#> 64       resolved       <NA>   <NA> 2025-10-31 14:11:27 3.35117674 secs
#> 65       resolved       <NA>   <NA> 2025-10-31 14:11:27 3.37705398 secs
#> 66       resolved       <NA>   <NA> 2025-10-31 14:11:27 3.40322518 secs
#> 67       resolved       <NA>   <NA> 2025-10-31 14:11:27 3.42930388 secs
#> 68       resolved       <NA>   <NA> 2025-10-31 14:11:27 3.45569396 secs
#> 69       resolved       <NA>   <NA> 2025-10-31 14:11:27 3.48300219 secs
#> 70       resolved       <NA>   <NA> 2025-10-31 14:11:27 3.50954509 secs
#> 71       resolved       <NA>   <NA> 2025-10-31 14:11:27 3.53609705 secs
#> 72       resolved       <NA>   <NA> 2025-10-31 14:11:27 3.56231213 secs
#> 73       resolved       <NA>   <NA> 2025-10-31 14:11:27 3.58900380 secs
#> 74       resolved       <NA>   <NA> 2025-10-31 14:11:27 3.61548543 secs
#> 75       resolved       <NA>   <NA> 2025-10-31 14:11:27 3.63875055 secs
#> 76       resolved       <NA>   <NA> 2025-10-31 14:11:27 3.66577435 secs
#> 77       resolved       <NA>   <NA> 2025-10-31 14:11:27 3.69199872 secs
#> 78       resolved       <NA>   <NA> 2025-10-31 14:11:27 3.71827197 secs
#> 79       resolved       <NA>   <NA> 2025-10-31 14:11:27 3.74489641 secs
#> 80       resolved       <NA>   <NA> 2025-10-31 14:11:27 3.77123308 secs
#> 81       resolved       <NA>   <NA> 2025-10-31 14:11:27 3.79878116 secs
#> 82       resolved       <NA>   <NA> 2025-10-31 14:11:27 3.82553005 secs
#> 83       resolved       <NA>   <NA> 2025-10-31 14:11:27 3.85216856 secs
#> 84       resolved       <NA>   <NA> 2025-10-31 14:11:27 3.87925839 secs
#> 85       resolved       <NA>   <NA> 2025-10-31 14:11:27 3.90841222 secs
#> 86       resolved       <NA>   <NA> 2025-10-31 14:11:27 3.93179178 secs
#> 87       resolved       <NA>   <NA> 2025-10-31 14:11:27 3.95569372 secs
#> 88       resolved       <NA>   <NA> 2025-10-31 14:11:27 3.97793269 secs
#> 89       resolved       <NA>   <NA> 2025-10-31 14:11:27 3.99981666 secs
#> 90       resolved       <NA>   <NA> 2025-10-31 14:11:27 4.02248669 secs
#> 91       resolved       <NA>   <NA> 2025-10-31 14:11:27 4.04519844 secs
#> 92       resolved       <NA>   <NA> 2025-10-31 14:11:27 4.06767631 secs
#> 93       resolved       <NA>   <NA> 2025-10-31 14:11:27 4.09054017 secs
#> 94       resolved       <NA>   <NA> 2025-10-31 14:11:27 4.11335111 secs
#> 95       resolved       <NA>   <NA> 2025-10-31 14:11:27 4.13607144 secs
#> 96       resolved       <NA>   <NA> 2025-10-31 14:11:27 4.15888238 secs
#> 97       resolved       <NA>   <NA> 2025-10-31 14:11:27 4.18163753 secs
#> 98       resolved       <NA>   <NA> 2025-10-31 14:11:28 4.20465255 secs
#> 99       resolved       <NA>   <NA> 2025-10-31 14:11:28 4.22689700 secs
#> 100      resolved       <NA>   <NA> 2025-10-31 14:11:28 4.25019383 secs
#> 101      resolved       <NA>   <NA> 2025-10-31 14:11:28 4.27353120 secs
#> 102      resolved       <NA>   <NA> 2025-10-31 14:11:28 4.29655266 secs
#> 103      resolved       <NA>   <NA> 2025-10-31 14:11:28 4.31954074 secs
#> 104      resolved       <NA>   <NA> 2025-10-31 14:11:28 4.34227824 secs
#> 105      resolved       <NA>   <NA> 2025-10-31 14:11:28 4.36496091 secs
#> 106      resolved       <NA>   <NA> 2025-10-31 14:11:28 4.38774228 secs
#> 107      resolved       <NA>   <NA> 2025-10-31 14:11:28 4.41041732 secs
#> 108      resolved       <NA>   <NA> 2025-10-31 14:11:28 4.43300152 secs
#> 109      resolved       <NA>   <NA> 2025-10-31 14:11:28 4.45541668 secs
#> 113      resolved       <NA>   <NA> 2025-10-31 14:11:28 4.47767878 secs
#> 110 receiveResult   overhead gather 2025-10-31 14:11:28 4.48507881 secs
#> 111        gather   overhead   <NA> 2025-10-31 14:11:28 4.48644924 secs
#> 114       resolve   overhead   <NA> 2025-10-31 14:11:28 4.49206257 secs
#>              duration future_label                         future_uuid
#> 1   1.664996e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 4   1.420596e+00 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 2   1.397347e+00 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 3   1.362801e-03 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 5   1.227856e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 112 3.006149e+00 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 6   1.066327e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 7   1.087856e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 8   1.087594e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 9   1.158142e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 10  1.289725e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 11  1.213455e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 12  1.187062e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 13  1.071882e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 14  1.090765e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 15  1.213408e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 16  1.226711e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 17  1.295400e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 18  1.239944e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 19  1.250529e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 20  1.267529e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 21  1.268697e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 22  1.243353e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 23  1.265359e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 24  1.249671e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 25  1.252413e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 26  1.252437e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 27  1.302576e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 28  1.278830e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 29  1.279330e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 30  1.174116e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 31  1.210284e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 32  2.061725e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 33  1.112580e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 34  1.107478e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 35  1.150584e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 36  1.418757e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 37  1.233172e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 38  1.236558e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 39  1.246500e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 40  1.249194e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 41  1.272583e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 42  1.339984e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 43  1.293635e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 44  1.277208e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 45  1.269412e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 46  1.285839e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 47  1.272821e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 48  1.274514e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 49  1.276255e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 50  1.301193e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 51  1.318169e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 52  1.333761e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 53  1.298356e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 54  1.347280e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 55  1.293254e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 56  1.271367e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 57  1.302791e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 58  1.247931e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 59  1.279140e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 60  1.129889e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 61  1.108146e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 62  1.222181e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 63  1.272917e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 64  1.261616e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 65  1.287627e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 66  1.270652e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 67  1.303649e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 68  1.390100e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 69  1.272798e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 70  1.317048e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 71  1.271677e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 72  1.270461e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 73  1.389217e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 74  1.148748e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 75  1.172400e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 76  1.288533e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 77  1.308489e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 78  1.288486e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 79  1.303267e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 80  1.406193e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 81  1.292443e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 82  1.287842e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 83  1.262665e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 84  1.273966e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 85  1.185155e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 86  1.183343e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 87  1.110101e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 88  1.087594e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 89  1.114225e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 90  1.118326e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 91  1.117921e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 92  1.127243e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 93  1.122522e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 94  1.123214e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 95  1.118517e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 96  1.150632e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 97  1.103520e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 98  1.105285e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 99  1.127219e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 100 1.154327e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 101 1.135397e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 102 1.132536e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 103 1.116920e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 104 1.119995e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 105 1.121259e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 106 1.119089e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 107 1.112437e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 108 1.090479e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 109 1.086521e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 113 1.255584e-02 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 110 2.117157e-04 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 111 1.058578e-04 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#> 114 2.717972e-05 secs         <NA> 8313298c511e3f14075b4f3bc7be71fb-19
#>                         session_uuid
#> 1   8313298c511e3f14075b4f3bc7be71fb
#> 4   8313298c511e3f14075b4f3bc7be71fb
#> 2   8313298c511e3f14075b4f3bc7be71fb
#> 3   8313298c511e3f14075b4f3bc7be71fb
#> 5   8313298c511e3f14075b4f3bc7be71fb
#> 112 8313298c511e3f14075b4f3bc7be71fb
#> 6   8313298c511e3f14075b4f3bc7be71fb
#> 7   8313298c511e3f14075b4f3bc7be71fb
#> 8   8313298c511e3f14075b4f3bc7be71fb
#> 9   8313298c511e3f14075b4f3bc7be71fb
#> 10  8313298c511e3f14075b4f3bc7be71fb
#> 11  8313298c511e3f14075b4f3bc7be71fb
#> 12  8313298c511e3f14075b4f3bc7be71fb
#> 13  8313298c511e3f14075b4f3bc7be71fb
#> 14  8313298c511e3f14075b4f3bc7be71fb
#> 15  8313298c511e3f14075b4f3bc7be71fb
#> 16  8313298c511e3f14075b4f3bc7be71fb
#> 17  8313298c511e3f14075b4f3bc7be71fb
#> 18  8313298c511e3f14075b4f3bc7be71fb
#> 19  8313298c511e3f14075b4f3bc7be71fb
#> 20  8313298c511e3f14075b4f3bc7be71fb
#> 21  8313298c511e3f14075b4f3bc7be71fb
#> 22  8313298c511e3f14075b4f3bc7be71fb
#> 23  8313298c511e3f14075b4f3bc7be71fb
#> 24  8313298c511e3f14075b4f3bc7be71fb
#> 25  8313298c511e3f14075b4f3bc7be71fb
#> 26  8313298c511e3f14075b4f3bc7be71fb
#> 27  8313298c511e3f14075b4f3bc7be71fb
#> 28  8313298c511e3f14075b4f3bc7be71fb
#> 29  8313298c511e3f14075b4f3bc7be71fb
#> 30  8313298c511e3f14075b4f3bc7be71fb
#> 31  8313298c511e3f14075b4f3bc7be71fb
#> 32  8313298c511e3f14075b4f3bc7be71fb
#> 33  8313298c511e3f14075b4f3bc7be71fb
#> 34  8313298c511e3f14075b4f3bc7be71fb
#> 35  8313298c511e3f14075b4f3bc7be71fb
#> 36  8313298c511e3f14075b4f3bc7be71fb
#> 37  8313298c511e3f14075b4f3bc7be71fb
#> 38  8313298c511e3f14075b4f3bc7be71fb
#> 39  8313298c511e3f14075b4f3bc7be71fb
#> 40  8313298c511e3f14075b4f3bc7be71fb
#> 41  8313298c511e3f14075b4f3bc7be71fb
#> 42  8313298c511e3f14075b4f3bc7be71fb
#> 43  8313298c511e3f14075b4f3bc7be71fb
#> 44  8313298c511e3f14075b4f3bc7be71fb
#> 45  8313298c511e3f14075b4f3bc7be71fb
#> 46  8313298c511e3f14075b4f3bc7be71fb
#> 47  8313298c511e3f14075b4f3bc7be71fb
#> 48  8313298c511e3f14075b4f3bc7be71fb
#> 49  8313298c511e3f14075b4f3bc7be71fb
#> 50  8313298c511e3f14075b4f3bc7be71fb
#> 51  8313298c511e3f14075b4f3bc7be71fb
#> 52  8313298c511e3f14075b4f3bc7be71fb
#> 53  8313298c511e3f14075b4f3bc7be71fb
#> 54  8313298c511e3f14075b4f3bc7be71fb
#> 55  8313298c511e3f14075b4f3bc7be71fb
#> 56  8313298c511e3f14075b4f3bc7be71fb
#> 57  8313298c511e3f14075b4f3bc7be71fb
#> 58  8313298c511e3f14075b4f3bc7be71fb
#> 59  8313298c511e3f14075b4f3bc7be71fb
#> 60  8313298c511e3f14075b4f3bc7be71fb
#> 61  8313298c511e3f14075b4f3bc7be71fb
#> 62  8313298c511e3f14075b4f3bc7be71fb
#> 63  8313298c511e3f14075b4f3bc7be71fb
#> 64  8313298c511e3f14075b4f3bc7be71fb
#> 65  8313298c511e3f14075b4f3bc7be71fb
#> 66  8313298c511e3f14075b4f3bc7be71fb
#> 67  8313298c511e3f14075b4f3bc7be71fb
#> 68  8313298c511e3f14075b4f3bc7be71fb
#> 69  8313298c511e3f14075b4f3bc7be71fb
#> 70  8313298c511e3f14075b4f3bc7be71fb
#> 71  8313298c511e3f14075b4f3bc7be71fb
#> 72  8313298c511e3f14075b4f3bc7be71fb
#> 73  8313298c511e3f14075b4f3bc7be71fb
#> 74  8313298c511e3f14075b4f3bc7be71fb
#> 75  8313298c511e3f14075b4f3bc7be71fb
#> 76  8313298c511e3f14075b4f3bc7be71fb
#> 77  8313298c511e3f14075b4f3bc7be71fb
#> 78  8313298c511e3f14075b4f3bc7be71fb
#> 79  8313298c511e3f14075b4f3bc7be71fb
#> 80  8313298c511e3f14075b4f3bc7be71fb
#> 81  8313298c511e3f14075b4f3bc7be71fb
#> 82  8313298c511e3f14075b4f3bc7be71fb
#> 83  8313298c511e3f14075b4f3bc7be71fb
#> 84  8313298c511e3f14075b4f3bc7be71fb
#> 85  8313298c511e3f14075b4f3bc7be71fb
#> 86  8313298c511e3f14075b4f3bc7be71fb
#> 87  8313298c511e3f14075b4f3bc7be71fb
#> 88  8313298c511e3f14075b4f3bc7be71fb
#> 89  8313298c511e3f14075b4f3bc7be71fb
#> 90  8313298c511e3f14075b4f3bc7be71fb
#> 91  8313298c511e3f14075b4f3bc7be71fb
#> 92  8313298c511e3f14075b4f3bc7be71fb
#> 93  8313298c511e3f14075b4f3bc7be71fb
#> 94  8313298c511e3f14075b4f3bc7be71fb
#> 95  8313298c511e3f14075b4f3bc7be71fb
#> 96  8313298c511e3f14075b4f3bc7be71fb
#> 97  8313298c511e3f14075b4f3bc7be71fb
#> 98  8313298c511e3f14075b4f3bc7be71fb
#> 99  8313298c511e3f14075b4f3bc7be71fb
#> 100 8313298c511e3f14075b4f3bc7be71fb
#> 101 8313298c511e3f14075b4f3bc7be71fb
#> 102 8313298c511e3f14075b4f3bc7be71fb
#> 103 8313298c511e3f14075b4f3bc7be71fb
#> 104 8313298c511e3f14075b4f3bc7be71fb
#> 105 8313298c511e3f14075b4f3bc7be71fb
#> 106 8313298c511e3f14075b4f3bc7be71fb
#> 107 8313298c511e3f14075b4f3bc7be71fb
#> 108 8313298c511e3f14075b4f3bc7be71fb
#> 109 8313298c511e3f14075b4f3bc7be71fb
#> 113 8313298c511e3f14075b4f3bc7be71fb
#> 110 8313298c511e3f14075b4f3bc7be71fb
#> 111 8313298c511e3f14075b4f3bc7be71fb
#> 114 8313298c511e3f14075b4f3bc7be71fb
#> 

## Stop parallel workers and disable journal logging and signaling
plan(sequential)
options(oopts)