Gets the logged journal of events for a future

WARNING: This function is under development. It can change at any time. For now, please, do not depend on this function in a published R package.

journal(x, ...)

Arguments

x

A Future object.

...

Not used.

Value

A data frame of class FutureJournal with columns:

event (character string) - type of event that took place

category (character string) - the category of the event

parent (character string) - (to be describe)

start (POSIXct) - the timestamp when the event started

at (difftime) - the time when the event started relative to first event

duration (difftime) - the duration of the event

future_label (character string) - the label of the future

future_uuid (character string) - the UUID of the future

session_uuid (character string) - the UUID of the R session where the event took place

The common events are:

create - the future was created (an overhead)

launch - the future was launched (an overhead)

evaluate - the future was evaluated (an evaluation)

resolved - the future was queried (may be occur multiple times) (an overhead)

gather - the results was retrieved (an overhead)

but others may be added by other Future classes.

Common event categorys are:

evaluation - processing time is spent on evaluation

overhead - processing time is spent on orchestrating the future

waiting - processing time is spent on waiting to set up or querying the future

but others may be added by other Future classes.

The data frame is sorted by the at time. Note that the timestamps for the evaluate event are based on the local time on the worker. The system clocks on the worker and the calling R system may not be in perfect sync.

Enabling and disabling event logging

To enable logging of events, set option future.journal is TRUE. To disable, set it to FALSE (default).

See also

Examples

## Enable journaling of futures
oopts <- options(future.journal = TRUE)

plan(multisession, workers = 2L)

t_start <- Sys.time()
fs <- lapply(1:3, FUN = function(x) future({ Sys.sleep(x); sqrt(x) }))
vs <- value(fs)
js <- lapply(fs, FUN = journal, baseline = t_start)
print(js)
#> [[1]]
#>            event   category parent               start               at
#> 1         create   overhead   <NA> 2023-05-19 18:50:36 0.000000000 secs
#> 6         launch   overhead   <NA> 2023-05-19 18:50:36 0.003949404 secs
#> 2      getWorker   overhead launch 2023-05-19 18:50:36 0.004347324 secs
#> 3    eraseWorker   overhead launch 2023-05-19 18:50:36 0.004864693 secs
#> 4 attachPackages   overhead launch 2023-05-19 18:50:36 0.005685806 secs
#> 5  exportGlobals   overhead launch 2023-05-19 18:50:36 0.006074190 secs
#> 9       evaluate evaluation   <NA> 2023-05-19 18:50:36 0.049399137 secs
#> 7  receiveResult   overhead gather 2023-05-19 18:50:37 1.059825897 secs
#> 8         gather   overhead   <NA> 2023-05-19 18:50:37 1.060439348 secs
#>            duration future_label                          future_uuid
#> 1 3.261566e-03 secs         <NA> e8a7e69d-992a-400e-b00d-8983d138fa57
#> 6 3.166676e-03 secs         <NA> e8a7e69d-992a-400e-b00d-8983d138fa57
#> 2 8.678436e-05 secs         <NA> e8a7e69d-992a-400e-b00d-8983d138fa57
#> 3 3.883839e-04 secs         <NA> e8a7e69d-992a-400e-b00d-8983d138fa57
#> 4 2.717972e-05 secs         <NA> e8a7e69d-992a-400e-b00d-8983d138fa57
#> 5 3.187656e-04 secs         <NA> e8a7e69d-992a-400e-b00d-8983d138fa57
#> 9 1.002380e+00 secs         <NA> e8a7e69d-992a-400e-b00d-8983d138fa57
#> 7 9.131432e-05 secs         <NA> e8a7e69d-992a-400e-b00d-8983d138fa57
#> 8 2.455711e-05 secs         <NA> e8a7e69d-992a-400e-b00d-8983d138fa57
#>                           session_uuid
#> 1 563f701f-4530-3881-b2f5-2a78d9d3bc51
#> 6 563f701f-4530-3881-b2f5-2a78d9d3bc51
#> 2 563f701f-4530-3881-b2f5-2a78d9d3bc51
#> 3 563f701f-4530-3881-b2f5-2a78d9d3bc51
#> 4 563f701f-4530-3881-b2f5-2a78d9d3bc51
#> 5 563f701f-4530-3881-b2f5-2a78d9d3bc51
#> 9 563f701f-4530-3881-b2f5-2a78d9d3bc51
#> 7 563f701f-4530-3881-b2f5-2a78d9d3bc51
#> 8 563f701f-4530-3881-b2f5-2a78d9d3bc51
#> 
#> [[2]]
#>            event   category parent               start               at
#> 1         create   overhead   <NA> 2023-05-19 18:50:36 0.000000000 secs
#> 6         launch   overhead   <NA> 2023-05-19 18:50:36 0.002240181 secs
#> 2      getWorker   overhead launch 2023-05-19 18:50:36 0.002621889 secs
#> 3    eraseWorker   overhead launch 2023-05-19 18:50:36 0.003193855 secs
#> 4 attachPackages   overhead launch 2023-05-19 18:50:36 0.028426170 secs
#> 5  exportGlobals   overhead launch 2023-05-19 18:50:36 0.028845072 secs
#> 9       evaluate evaluation   <NA> 2023-05-19 18:50:36 0.074164629 secs
#> 7  receiveResult   overhead gather 2023-05-19 18:50:38 2.121537685 secs
#> 8         gather   overhead   <NA> 2023-05-19 18:50:38 2.122930765 secs
#>            duration future_label                          future_uuid
#> 1 1.616240e-03 secs         <NA> 78e58527-768b-1177-5d85-32f265aa704c
#> 6 2.775288e-02 secs         <NA> 78e58527-768b-1177-5d85-32f265aa704c
#> 2 1.037121e-04 secs         <NA> 78e58527-768b-1177-5d85-32f265aa704c
#> 3 2.471590e-02 secs         <NA> 78e58527-768b-1177-5d85-32f265aa704c
#> 4 3.767014e-05 secs         <NA> 78e58527-768b-1177-5d85-32f265aa704c
#> 5 2.641678e-04 secs         <NA> 78e58527-768b-1177-5d85-32f265aa704c
#> 9 2.040046e+00 secs         <NA> 78e58527-768b-1177-5d85-32f265aa704c
#> 7 1.840591e-04 secs         <NA> 78e58527-768b-1177-5d85-32f265aa704c
#> 8 5.149841e-05 secs         <NA> 78e58527-768b-1177-5d85-32f265aa704c
#>                           session_uuid
#> 1 563f701f-4530-3881-b2f5-2a78d9d3bc51
#> 6 563f701f-4530-3881-b2f5-2a78d9d3bc51
#> 2 563f701f-4530-3881-b2f5-2a78d9d3bc51
#> 3 563f701f-4530-3881-b2f5-2a78d9d3bc51
#> 4 563f701f-4530-3881-b2f5-2a78d9d3bc51
#> 5 563f701f-4530-3881-b2f5-2a78d9d3bc51
#> 9 563f701f-4530-3881-b2f5-2a78d9d3bc51
#> 7 563f701f-4530-3881-b2f5-2a78d9d3bc51
#> 8 563f701f-4530-3881-b2f5-2a78d9d3bc51
#> 
#> [[3]]
#>            event   category parent               start               at
#> 1         create   overhead   <NA> 2023-05-19 18:50:36 0.000000000 secs
#> 6         launch   overhead   <NA> 2023-05-19 18:50:36 0.002396584 secs
#> 2      getWorker   overhead launch 2023-05-19 18:50:36 0.002786398 secs
#> 3    eraseWorker   overhead launch 2023-05-19 18:50:37 1.024413347 secs
#> 4 attachPackages   overhead launch 2023-05-19 18:50:37 1.025252819 secs
#> 5  exportGlobals   overhead launch 2023-05-19 18:50:37 1.025630236 secs
#> 9       evaluate evaluation   <NA> 2023-05-19 18:50:37 1.067344427 secs
#> 7  receiveResult   overhead gather 2023-05-19 18:50:40 4.081427574 secs
#> 8         gather   overhead   <NA> 2023-05-19 18:50:40 4.082114935 secs
#>            duration future_label                          future_uuid
#> 1 1.735926e-03 secs         <NA> 64dbf578-f9d8-f8ee-a34e-e6a6db6c61ab
#> 6 1.024341e+00 secs         <NA> 64dbf578-f9d8-f8ee-a34e-e6a6db6c61ab
#> 2 1.021260e+00 secs         <NA> 64dbf578-f9d8-f8ee-a34e-e6a6db6c61ab
#> 3 4.339218e-04 secs         <NA> 64dbf578-f9d8-f8ee-a34e-e6a6db6c61ab
#> 4 2.908707e-05 secs         <NA> 64dbf578-f9d8-f8ee-a34e-e6a6db6c61ab
#> 5 2.918243e-04 secs         <NA> 64dbf578-f9d8-f8ee-a34e-e6a6db6c61ab
#> 9 3.004534e+00 secs         <NA> 64dbf578-f9d8-f8ee-a34e-e6a6db6c61ab
#> 7 1.034737e-04 secs         <NA> 64dbf578-f9d8-f8ee-a34e-e6a6db6c61ab
#> 8 2.908707e-05 secs         <NA> 64dbf578-f9d8-f8ee-a34e-e6a6db6c61ab
#>                           session_uuid
#> 1 563f701f-4530-3881-b2f5-2a78d9d3bc51
#> 6 563f701f-4530-3881-b2f5-2a78d9d3bc51
#> 2 563f701f-4530-3881-b2f5-2a78d9d3bc51
#> 3 563f701f-4530-3881-b2f5-2a78d9d3bc51
#> 4 563f701f-4530-3881-b2f5-2a78d9d3bc51
#> 5 563f701f-4530-3881-b2f5-2a78d9d3bc51
#> 9 563f701f-4530-3881-b2f5-2a78d9d3bc51
#> 7 563f701f-4530-3881-b2f5-2a78d9d3bc51
#> 8 563f701f-4530-3881-b2f5-2a78d9d3bc51
#> 

## Stop parallel workers and disable journal logging and signaling
plan(sequential)
options(oopts)