Gets the logged journal of events for a future

WARNING: This function is under development. It can change at any time. For now, please, do not depend on this function in a published R package.

journal(x, ...)

Arguments

x

A Future object.

...

Not used.

Value

A data frame of class FutureJournal with columns:

event (character string) - type of event that took place

category (character string) - the category of the event

parent (character string) - (to be describe)

start (POSIXct) - the timestamp when the event started

at (difftime) - the time when the event started relative to first event

duration (difftime) - the duration of the event

future_label (character string) - the label of the future

future_uuid (character string) - the UUID of the future

session_uuid (character string) - the UUID of the R session where the event took place

The common events are:

create - the future was created (an overhead)

launch - the future was launched (an overhead)

evaluate - the future was evaluated (an evaluation)

resolved - the future was queried (may be occur multiple times) (an overhead)

gather - the results was retrieved (an overhead)

but others may be added by other Future classes.

Common event categories are:

evaluation - processing time is spent on evaluation

overhead - processing time is spent on orchestrating the future

waiting - processing time is spent on waiting to set up or querying the future

but others may be added by other Future classes.

The data frame is sorted by the at time. Note that the timestamps for the evaluate event are based on the local time on the worker. The system clocks on the worker and the calling R system may not be in perfect sync.

Enabling and disabling event logging

To enable logging of events, set option future.journal is TRUE. To disable, set it to FALSE (default).

See also

Examples

## Enable journaling of futures
oopts <- options(future.journal = TRUE)

plan(multisession, workers = 2L)

t_start <- Sys.time()
fs <- lapply(1:3, FUN = function(x) future({ Sys.sleep(x); sqrt(x) }))
vs <- value(fs)
js <- lapply(fs, FUN = journal, baseline = t_start)
print(js)
#> [[1]]
#>            event   category parent               start              at
#> 1         create   overhead   <NA> 2023-12-13 11:51:17 0.00000000 secs
#> 6         launch   overhead   <NA> 2023-12-13 11:51:17 0.01207590 secs
#> 2      getWorker   overhead launch 2023-12-13 11:51:17 0.01252818 secs
#> 3    eraseWorker   overhead launch 2023-12-13 11:51:17 0.01306963 secs
#> 4 attachPackages   overhead launch 2023-12-13 11:51:17 0.01390934 secs
#> 5  exportGlobals   overhead launch 2023-12-13 11:51:17 0.01435876 secs
#> 9       evaluate evaluation   <NA> 2023-12-13 11:51:17 0.06083989 secs
#> 7  receiveResult   overhead gather 2023-12-13 11:51:18 1.06472182 secs
#> 8         gather   overhead   <NA> 2023-12-13 11:51:18 1.06659436 secs
#>            duration future_label                          future_uuid
#> 1 3.228188e-03 secs         <NA> 610113ef-0f60-2085-4499-6adb4631f387
#> 6 3.619671e-03 secs         <NA> 610113ef-0f60-2085-4499-6adb4631f387
#> 2 1.065731e-04 secs         <NA> 610113ef-0f60-2085-4499-6adb4631f387
#> 3 4.160404e-04 secs         <NA> 610113ef-0f60-2085-4499-6adb4631f387
#> 4 2.861023e-05 secs         <NA> 610113ef-0f60-2085-4499-6adb4631f387
#> 5 3.466606e-04 secs         <NA> 610113ef-0f60-2085-4499-6adb4631f387
#> 9 1.003229e+00 secs         <NA> 610113ef-0f60-2085-4499-6adb4631f387
#> 7 2.593994e-04 secs         <NA> 610113ef-0f60-2085-4499-6adb4631f387
#> 8 7.891655e-05 secs         <NA> 610113ef-0f60-2085-4499-6adb4631f387
#>                           session_uuid
#> 1 bfae1d0f-bdd4-577a-883e-47180f7f5e8e
#> 6 bfae1d0f-bdd4-577a-883e-47180f7f5e8e
#> 2 bfae1d0f-bdd4-577a-883e-47180f7f5e8e
#> 3 bfae1d0f-bdd4-577a-883e-47180f7f5e8e
#> 4 bfae1d0f-bdd4-577a-883e-47180f7f5e8e
#> 5 bfae1d0f-bdd4-577a-883e-47180f7f5e8e
#> 9 bfae1d0f-bdd4-577a-883e-47180f7f5e8e
#> 7 bfae1d0f-bdd4-577a-883e-47180f7f5e8e
#> 8 bfae1d0f-bdd4-577a-883e-47180f7f5e8e
#> 
#> [[2]]
#>            event   category parent               start              at
#> 1         create   overhead   <NA> 2023-12-13 11:51:17 0.00000000 secs
#> 6         launch   overhead   <NA> 2023-12-13 11:51:17 0.01005864 secs
#> 2      getWorker   overhead launch 2023-12-13 11:51:17 0.01046801 secs
#> 3    eraseWorker   overhead launch 2023-12-13 11:51:17 0.01094627 secs
#> 4 attachPackages   overhead launch 2023-12-13 11:51:17 0.03675938 secs
#> 5  exportGlobals   overhead launch 2023-12-13 11:51:17 0.03728580 secs
#> 9       evaluate evaluation   <NA> 2023-12-13 11:51:17 0.08497429 secs
#> 7  receiveResult   overhead gather 2023-12-13 11:51:19 2.12714529 secs
#> 8         gather   overhead   <NA> 2023-12-13 11:51:19 2.12893105 secs
#>            duration future_label                          future_uuid
#> 1 1.660824e-03 secs         <NA> cd58e055-69fe-91c9-5e09-bde9b88a05de
#> 6 2.859259e-02 secs         <NA> cd58e055-69fe-91c9-5e09-bde9b88a05de
#> 2 9.059906e-05 secs         <NA> cd58e055-69fe-91c9-5e09-bde9b88a05de
#> 3 2.510262e-02 secs         <NA> cd58e055-69fe-91c9-5e09-bde9b88a05de
#> 4 4.792213e-05 secs         <NA> cd58e055-69fe-91c9-5e09-bde9b88a05de
#> 5 4.293919e-04 secs         <NA> cd58e055-69fe-91c9-5e09-bde9b88a05de
#> 9 2.041693e+00 secs         <NA> cd58e055-69fe-91c9-5e09-bde9b88a05de
#> 7 2.408028e-04 secs         <NA> cd58e055-69fe-91c9-5e09-bde9b88a05de
#> 8 6.937981e-05 secs         <NA> cd58e055-69fe-91c9-5e09-bde9b88a05de
#>                           session_uuid
#> 1 bfae1d0f-bdd4-577a-883e-47180f7f5e8e
#> 6 bfae1d0f-bdd4-577a-883e-47180f7f5e8e
#> 2 bfae1d0f-bdd4-577a-883e-47180f7f5e8e
#> 3 bfae1d0f-bdd4-577a-883e-47180f7f5e8e
#> 4 bfae1d0f-bdd4-577a-883e-47180f7f5e8e
#> 5 bfae1d0f-bdd4-577a-883e-47180f7f5e8e
#> 9 bfae1d0f-bdd4-577a-883e-47180f7f5e8e
#> 7 bfae1d0f-bdd4-577a-883e-47180f7f5e8e
#> 8 bfae1d0f-bdd4-577a-883e-47180f7f5e8e
#> 
#> [[3]]
#>            event   category parent               start              at
#> 1         create   overhead   <NA> 2023-12-13 11:51:17 0.00000000 secs
#> 6         launch   overhead   <NA> 2023-12-13 11:51:17 0.01056528 secs
#> 2      getWorker   overhead launch 2023-12-13 11:51:17 0.01095867 secs
#> 3    eraseWorker   overhead launch 2023-12-13 11:51:18 1.01698256 secs
#> 4 attachPackages   overhead launch 2023-12-13 11:51:18 1.01928616 secs
#> 5  exportGlobals   overhead launch 2023-12-13 11:51:18 1.02047253 secs
#> 9       evaluate evaluation   <NA> 2023-12-13 11:51:18 1.06593990 secs
#> 7  receiveResult   overhead gather 2023-12-13 11:51:21 4.07098126 secs
#> 8         gather   overhead   <NA> 2023-12-13 11:51:21 4.07270217 secs
#>            duration future_label                          future_uuid
#> 1 1.781702e-03 secs         <NA> 3a030673-f0b5-9d97-54a5-a8b808e68c07
#> 6 1.013139e+00 secs         <NA> 3a030673-f0b5-9d97-54a5-a8b808e68c07
#> 2 1.004880e+00 secs         <NA> 3a030673-f0b5-9d97-54a5-a8b808e68c07
#> 3 8.513927e-04 secs         <NA> 3a030673-f0b5-9d97-54a5-a8b808e68c07
#> 4 8.392334e-05 secs         <NA> 3a030673-f0b5-9d97-54a5-a8b808e68c07
#> 5 8.242130e-04 secs         <NA> 3a030673-f0b5-9d97-54a5-a8b808e68c07
#> 9 3.004596e+00 secs         <NA> 3a030673-f0b5-9d97-54a5-a8b808e68c07
#> 7 2.155304e-04 secs         <NA> 3a030673-f0b5-9d97-54a5-a8b808e68c07
#> 8 7.271767e-05 secs         <NA> 3a030673-f0b5-9d97-54a5-a8b808e68c07
#>                           session_uuid
#> 1 bfae1d0f-bdd4-577a-883e-47180f7f5e8e
#> 6 bfae1d0f-bdd4-577a-883e-47180f7f5e8e
#> 2 bfae1d0f-bdd4-577a-883e-47180f7f5e8e
#> 3 bfae1d0f-bdd4-577a-883e-47180f7f5e8e
#> 4 bfae1d0f-bdd4-577a-883e-47180f7f5e8e
#> 5 bfae1d0f-bdd4-577a-883e-47180f7f5e8e
#> 9 bfae1d0f-bdd4-577a-883e-47180f7f5e8e
#> 7 bfae1d0f-bdd4-577a-883e-47180f7f5e8e
#> 8 bfae1d0f-bdd4-577a-883e-47180f7f5e8e
#> 

## Stop parallel workers and disable journal logging and signaling
plan(sequential)
options(oopts)