Gets the logged journal of events for a future

WARNING: This function is under development. It can change at any time. For now, please, do not depend on this function in a published R package.

journal(x, ...)

Arguments

x

A Future object.

...

Not used.

Value

A data frame of class FutureJournal with columns:

event (character string) - type of event that took place

category (character string) - the category of the event

parent (character string) - (to be describe)

start (POSIXct) - the timestamp when the event started

at (difftime) - the time when the event started relative to first event

duration (difftime) - the duration of the event

future_label (character string) - the label of the future

future_uuid (character string) - the UUID of the future

session_uuid (character string) - the UUID of the R session where the event took place

The common events are:

create - the future was created (an overhead)

launch - the future was launched (an overhead)

evaluate - the future was evaluated (an evaluation)

resolved - the future was queried (may be occur multiple times) (an overhead)

gather - the results was retrieved (an overhead)

but others may be added by other Future classes.

Common event categories are:

evaluation - processing time is spent on evaluation

overhead - processing time is spent on orchestrating the future

waiting - processing time is spent on waiting to set up or querying the future

but others may be added by other Future classes.

The data frame is sorted by the at time. Note that the timestamps for the evaluate event are based on the local time on the worker. The system clocks on the worker and the calling R system may not be in perfect sync.

Enabling and disabling event logging

To enable logging of events, set option future.journal is TRUE. To disable, set it to FALSE (default).

See also

Examples

## Enable journaling of futures
oopts <- options(future.journal = TRUE)

plan(multisession, workers = 2L)

t_start <- Sys.time()
fs <- lapply(1:3, FUN = function(x) future({ Sys.sleep(x); sqrt(x) }))
vs <- value(fs)
js <- lapply(fs, FUN = journal, baseline = t_start)
print(js)
#> [[1]]
#>            event   category parent               start              at
#> 1         create   overhead   <NA> 2023-05-23 20:17:56 0.00000000 secs
#> 6         launch   overhead   <NA> 2023-05-23 20:17:56 0.01192403 secs
#> 2      getWorker   overhead launch 2023-05-23 20:17:56 0.01235819 secs
#> 3    eraseWorker   overhead launch 2023-05-23 20:17:56 0.01294327 secs
#> 4 attachPackages   overhead launch 2023-05-23 20:17:56 0.01367855 secs
#> 5  exportGlobals   overhead launch 2023-05-23 20:17:56 0.01405811 secs
#> 9       evaluate evaluation   <NA> 2023-05-23 20:17:56 0.01534986 secs
#> 7  receiveResult   overhead gather 2023-05-23 20:17:57 1.01885271 secs
#> 8         gather   overhead   <NA> 2023-05-23 20:17:57 1.02068949 secs
#>            duration future_label                          future_uuid
#> 1 3.234386e-03 secs         <NA> e7d7097c-413c-307c-f9d5-3fe4368cca87
#> 6 3.080845e-03 secs         <NA> e7d7097c-413c-307c-f9d5-3fe4368cca87
#> 2 1.125336e-04 secs         <NA> e7d7097c-413c-307c-f9d5-3fe4368cca87
#> 3 3.409386e-04 secs         <NA> e7d7097c-413c-307c-f9d5-3fe4368cca87
#> 4 2.789497e-05 secs         <NA> e7d7097c-413c-307c-f9d5-3fe4368cca87
#> 5 1.990795e-04 secs         <NA> e7d7097c-413c-307c-f9d5-3fe4368cca87
#> 9 1.002998e+00 secs         <NA> e7d7097c-413c-307c-f9d5-3fe4368cca87
#> 7 2.322197e-04 secs         <NA> e7d7097c-413c-307c-f9d5-3fe4368cca87
#> 8 7.414818e-05 secs         <NA> e7d7097c-413c-307c-f9d5-3fe4368cca87
#>                           session_uuid
#> 1 78d1f4cb-e9e3-6067-c223-2d9e7246ebd1
#> 6 78d1f4cb-e9e3-6067-c223-2d9e7246ebd1
#> 2 78d1f4cb-e9e3-6067-c223-2d9e7246ebd1
#> 3 78d1f4cb-e9e3-6067-c223-2d9e7246ebd1
#> 4 78d1f4cb-e9e3-6067-c223-2d9e7246ebd1
#> 5 78d1f4cb-e9e3-6067-c223-2d9e7246ebd1
#> 9 78d1f4cb-e9e3-6067-c223-2d9e7246ebd1
#> 7 78d1f4cb-e9e3-6067-c223-2d9e7246ebd1
#> 8 78d1f4cb-e9e3-6067-c223-2d9e7246ebd1
#> 
#> [[2]]
#>            event   category parent               start              at
#> 1         create   overhead   <NA> 2023-05-23 20:17:56 0.00000000 secs
#> 6         launch   overhead   <NA> 2023-05-23 20:17:56 0.01020408 secs
#> 2      getWorker   overhead launch 2023-05-23 20:17:56 0.01421905 secs
#> 3    eraseWorker   overhead launch 2023-05-23 20:17:56 0.01486111 secs
#> 4 attachPackages   overhead launch 2023-05-23 20:17:56 0.04053187 secs
#> 5  exportGlobals   overhead launch 2023-05-23 20:17:56 0.04101896 secs
#> 9       evaluate evaluation   <NA> 2023-05-23 20:17:56 0.08341956 secs
#> 7  receiveResult   overhead gather 2023-05-23 20:17:58 2.12786531 secs
#> 8         gather   overhead   <NA> 2023-05-23 20:17:58 2.12949061 secs
#>            duration future_label                          future_uuid
#> 1 1.755476e-03 secs         <NA> 5a522dd6-8c12-7672-83ab-5773d097fdaf
#> 6 3.195596e-02 secs         <NA> 5a522dd6-8c12-7672-83ab-5773d097fdaf
#> 2 1.258850e-04 secs         <NA> 5a522dd6-8c12-7672-83ab-5773d097fdaf
#> 3 2.507401e-02 secs         <NA> 5a522dd6-8c12-7672-83ab-5773d097fdaf
#> 4 4.029274e-05 secs         <NA> 5a522dd6-8c12-7672-83ab-5773d097fdaf
#> 5 3.778934e-04 secs         <NA> 5a522dd6-8c12-7672-83ab-5773d097fdaf
#> 9 2.042032e+00 secs         <NA> 5a522dd6-8c12-7672-83ab-5773d097fdaf
#> 7 2.124310e-04 secs         <NA> 5a522dd6-8c12-7672-83ab-5773d097fdaf
#> 8 8.845329e-05 secs         <NA> 5a522dd6-8c12-7672-83ab-5773d097fdaf
#>                           session_uuid
#> 1 78d1f4cb-e9e3-6067-c223-2d9e7246ebd1
#> 6 78d1f4cb-e9e3-6067-c223-2d9e7246ebd1
#> 2 78d1f4cb-e9e3-6067-c223-2d9e7246ebd1
#> 3 78d1f4cb-e9e3-6067-c223-2d9e7246ebd1
#> 4 78d1f4cb-e9e3-6067-c223-2d9e7246ebd1
#> 5 78d1f4cb-e9e3-6067-c223-2d9e7246ebd1
#> 9 78d1f4cb-e9e3-6067-c223-2d9e7246ebd1
#> 7 78d1f4cb-e9e3-6067-c223-2d9e7246ebd1
#> 8 78d1f4cb-e9e3-6067-c223-2d9e7246ebd1
#> 
#> [[3]]
#>            event   category parent               start              at
#> 1         create   overhead   <NA> 2023-05-23 20:17:56 0.00000000 secs
#> 6         launch   overhead   <NA> 2023-05-23 20:17:56 0.01013350 secs
#> 2      getWorker   overhead launch 2023-05-23 20:17:56 0.01051569 secs
#> 3    eraseWorker   overhead launch 2023-05-23 20:17:57 0.96900535 secs
#> 4 attachPackages   overhead launch 2023-05-23 20:17:57 0.97089052 secs
#> 5  exportGlobals   overhead launch 2023-05-23 20:17:57 0.97205400 secs
#> 9       evaluate evaluation   <NA> 2023-05-23 20:17:57 1.01712584 secs
#> 7  receiveResult   overhead gather 2023-05-23 20:18:00 4.02363586 secs
#> 8         gather   overhead   <NA> 2023-05-23 20:18:00 4.02539897 secs
#>            duration future_label                          future_uuid
#> 1 1.670837e-03 secs         <NA> dc7ea6d7-0b14-9c0c-4101-ae8538d3e2db
#> 6 9.646230e-01 secs         <NA> dc7ea6d7-0b14-9c0c-4101-ae8538d3e2db
#> 2 9.572368e-01 secs         <NA> dc7ea6d7-0b14-9c0c-4101-ae8538d3e2db
#> 3 6.775856e-04 secs         <NA> dc7ea6d7-0b14-9c0c-4101-ae8538d3e2db
#> 4 7.748604e-05 secs         <NA> dc7ea6d7-0b14-9c0c-4101-ae8538d3e2db
#> 5 6.113052e-04 secs         <NA> dc7ea6d7-0b14-9c0c-4101-ae8538d3e2db
#> 9 3.004676e+00 secs         <NA> dc7ea6d7-0b14-9c0c-4101-ae8538d3e2db
#> 7 2.036095e-04 secs         <NA> dc7ea6d7-0b14-9c0c-4101-ae8538d3e2db
#> 8 6.675720e-05 secs         <NA> dc7ea6d7-0b14-9c0c-4101-ae8538d3e2db
#>                           session_uuid
#> 1 78d1f4cb-e9e3-6067-c223-2d9e7246ebd1
#> 6 78d1f4cb-e9e3-6067-c223-2d9e7246ebd1
#> 2 78d1f4cb-e9e3-6067-c223-2d9e7246ebd1
#> 3 78d1f4cb-e9e3-6067-c223-2d9e7246ebd1
#> 4 78d1f4cb-e9e3-6067-c223-2d9e7246ebd1
#> 5 78d1f4cb-e9e3-6067-c223-2d9e7246ebd1
#> 9 78d1f4cb-e9e3-6067-c223-2d9e7246ebd1
#> 7 78d1f4cb-e9e3-6067-c223-2d9e7246ebd1
#> 8 78d1f4cb-e9e3-6067-c223-2d9e7246ebd1
#> 

## Stop parallel workers and disable journal logging and signaling
plan(sequential)
options(oopts)